All platforms

List all platforms

allPlatforms(platforms = NULL, filter = NULL, offset = 0, limit = 20,
  sort = "+id", file = NULL, return = TRUE, overwrite = FALSE,
  memoised = FALSE)

Arguments

platforms
A list of identifiers (e.g: GPL96,GPL1355,GPL1261). Identifiers can either be the Platform ID or its short name.

Retrieval by ID is more efficient.

Only platforms that user has access to will be available.

Do not combine different identifiers in one query.

filter
Filtering can be done on any* property of a supported type or nested property that the ExpressionExperiment class has ( and is mapped by hibernate ). E.g: 'curationDetails' or 'curationDetails.lastTroubledEvent.date' Currently supported types are:

String - property of String type, required value can be any String.

Number - any Number implementation. Required value must be a string parseable to the specific Number type.

Boolean - required value will be parsed to true only if the string matches 'true', ignoring case.

Accepted operator keywords are:

'=' - equality

'!=' - non-equality

'<' - smaller than

'=>' - larger or equal

'like' similar string, effectively means 'contains', translates to the sql 'LIKE' operator (given value will be surrounded by

Multiple filters can be chained using 'AND' or 'OR' keywords.

Leave space between the keywords and the previous/next word!

E.g: ?filter=property1 < value1 AND property2 like value2

If chained filters are mixed conjunctions and disjunctions, the query must be in conjunctive normal form (CNF). Parentheses are not necessary - every AND keyword separates blocks of disjunctions. Example:

?filter=p1 = v1 OR p1 != v2 AND p2 <= v2 AND p3 > v3 OR p3 < v4

Above query will translate to:

(p1 = v1 OR p1 != v2) AND (p2 <= v2) AND (p3 > v3 OR p3 < v4;)

Breaking the CNF results in an error.

Filter "curationDetails.troubled" will be ignored if user is not an administrator.

offset
Integer. Optional parameter (defaults to 0) skips the specified amount of datasets when retrieving them from the database.

limit
Integer. Optional parameter (defaults to 20) limits the result to specified amount of datasets. Use 0 for no limit.

sort
Character. Optional parameter (defaults to +id) sets the ordering property and direction. Format is [+,-][property name]. E.g. "-accession" will translate to descending ordering by the Accession property. Note that this does not guarantee the order of the returned entities. Nested properties are also supported (recursively).

E.g: +curationDetails.lastTroubledEvent.date

file
Character. File path. If provided, response will be saved to file

return
Logical. If the response should be returned. Set to false when you only want to save a file

overwrite
Logical. If TRUE, existing files will be overwritten. If FALSE a warning will be thrown and no action is taken.

memoised
Logical. If TRUE a memoised version of the function will be used which is faster for repeated requests. Use forgetGemmaMemoised to clear memory.

Value

List of lists containing platform object.

Examples

allPlatforms()

#> $GPL96
#> $GPL96$shortName
#> [1] "GPL96"
#> 
#> $GPL96$taxon
#> [1] "human"
#> 
#> $GPL96$technologyType
#> [1] "ONECOLOR"
#> 
#> $GPL96$lastGeneMapping
#> NULL
#> 
#> $GPL96$lastSequenceAnalysis
#> NULL
#> 
#> $GPL96$lastSequenceUpdate
#> NULL
#> 
#> $GPL96$numProbeSequences
#> NULL
#> 
#> $GPL96$numProbeAlignments
#> NULL
#> 
#> $GPL96$numProbesToGenes
#> NULL
#> 
#> $GPL96$numGenes
#> NULL
#> 
#> $GPL96$designElementCount
#> NULL
#> 
#> $GPL96$dateCached
#> NULL
#> 
#> $GPL96$color
#> [1] "ONECOLOR"
#> 
#> $GPL96$expressionExperimentCount
#> [1] 411
#> 
#> $GPL96$hasBlatAssociations
#> NULL
#> 
#> $GPL96$hasGeneAssociations
#> NULL
#> 
#> $GPL96$hasSequenceAssociations
#> NULL
#> 
#> $GPL96$isMerged
#> NULL
#> 
#> $GPL96$isMergee
#> [1] TRUE
#> 
#> $GPL96$isSubsumed
#> NULL
#> 
#> $GPL96$isSubsumer
#> NULL
#> 
#> $GPL96$lastRepeatMask
#> NULL
#> 
#> $GPL96$description
#> [1] " The U133 set includes 2 arrays with a total of 44928 entries and was indexed 29-Jan-2002. The set includes over 1,000,000 unique oligonucleotide features covering more than 39,000 transcript variants, which in turn represent greater than 33,000 of the best characterized human genes. Sequences were selected from GenBank, dbEST, and RefSeq. Sequence clusters were created from Build 133 of UniGene (April 20, 2001) and refined by analysis and comparison with a number of other publicly available databases including the Washington University EST trace repository and the University of California, Santa Cruz golden-path human genome database (April 2001 release). In addition, ESTs were analyzed for untrimmed low-quality sequence information, correct orientation, false priming, false clustering, alternative splicing and alternative polyadenylation.  Keywords = high density oligonucleotide array\nFrom GPL96\nLast Updated: Mar 09 2006"
#> 
#> $GPL96$name
#> [1] "Affymetrix GeneChip Human Genome U133 Array Set HG-U133A"
#> 
#> $GPL96$troubled
#> [1] FALSE
#> 
#> $GPL96$lastUpdated
#> [1] 1.487812e+12
#> 
#> $GPL96$lastTroubledEvent
#> NULL
#> 
#> $GPL96$troubleDetails
#> [1] "No trouble details provided."
#> 
#> $GPL96$lastNeedsAttentionEvent
#> NULL
#> 
#> $GPL96$curationNote
#> NULL
#> 
#> $GPL96$needsAttention
#> [1] FALSE
#> 
#> $GPL96$lastNoteUpdateEvent
#> NULL
#> 
#> $GPL96$id
#> [1] 1
#> 
#> 
#> $GPL1355
#> $GPL1355$shortName
#> [1] "GPL1355"
#> 
#> $GPL1355$taxon
#> [1] "rat"
#> 
#> $GPL1355$technologyType
#> [1] "ONECOLOR"
#> 
#> $GPL1355$lastGeneMapping
#> NULL
#> 
#> $GPL1355$lastSequenceAnalysis
#> NULL
#> 
#> $GPL1355$lastSequenceUpdate
#> NULL
#> 
#> $GPL1355$numProbeSequences
#> NULL
#> 
#> $GPL1355$numProbeAlignments
#> NULL
#> 
#> $GPL1355$numProbesToGenes
#> NULL
#> 
#> $GPL1355$numGenes
#> NULL
#> 
#> $GPL1355$designElementCount
#> NULL
#> 
#> $GPL1355$dateCached
#> NULL
#> 
#> $GPL1355$color
#> [1] "ONECOLOR"
#> 
#> $GPL1355$expressionExperimentCount
#> [1] 197
#> 
#> $GPL1355$hasBlatAssociations
#> NULL
#> 
#> $GPL1355$hasGeneAssociations
#> NULL
#> 
#> $GPL1355$hasSequenceAssociations
#> NULL
#> 
#> $GPL1355$isMerged
#> NULL
#> 
#> $GPL1355$isMergee
#> [1] FALSE
#> 
#> $GPL1355$isSubsumed
#> NULL
#> 
#> $GPL1355$isSubsumer
#> NULL
#> 
#> $GPL1355$lastRepeatMask
#> NULL
#> 
#> $GPL1355$description
#> [1] " The GeneChip Rat Genome 230 2.0 Array is a powerful tool for toxicology, neurobiology, and other applications using rat as a model organism.  - Provides comprehensive coverage of the transcribed rat genome on a single array - Comprised of more than 31,000 probe sets, analyzing over 30,000 transcripts and variants from over 28,000 well-substantiated rat genes - The publicly available draft of the rat genome and leading public rat databases were used to refine sequences and provide a higher quality of data output  All probe sets represented on the GeneChip Rat Expression Set 230 are included on the GeneChip Rat Genome 230 2.0 Array. Sequences used in the design of the GeneChip Rat Genome 230 2.0 Array were selected from GenBank, dbEST, and RefSeq. The sequence clusters were created from the UniGene database (Build 99, June 2002) and then refined by analysis and comparison with the publicly available draft assembly of the rat genome from the Baylor College of Medicine Human Genome Sequencing Center (June 2002).  The GeneChip Rat Genome 230 2.0 Array includes representation of the RefSeq database sequences and probe sets related to sequences and refined EST clusters previously represented on the GeneChip Rat Genome U34 Set.  Oligonucleotide probes complementary to each corresponding sequence are synthesized in situ on the arrays. Eleven pairs of oligonucleotide probes are used to measure the level of transcription of each sequence represented on the GeneChip Rat Genome 230 2.0 Array.  Annotations derived from Affymetrix CSV file dated 6/23/2004\nFrom GPL1355\nLast Updated: May 31 2005"
#> 
#> $GPL1355$name
#> [1] "Affymetrix GeneChip Rat Genome 230 2.0 Array"
#> 
#> $GPL1355$troubled
#> [1] FALSE
#> 
#> $GPL1355$lastUpdated
#> [1] 1.488129e+12
#> 
#> $GPL1355$lastTroubledEvent
#> NULL
#> 
#> $GPL1355$troubleDetails
#> [1] "No trouble details provided."
#> 
#> $GPL1355$lastNeedsAttentionEvent
#> NULL
#> 
#> $GPL1355$curationNote
#> NULL
#> 
#> $GPL1355$needsAttention
#> [1] FALSE
#> 
#> $GPL1355$lastNoteUpdateEvent
#> NULL
#> 
#> $GPL1355$id
#> [1] 2
#> 
#> 
#> $GPL1261
#> $GPL1261$shortName
#> [1] "GPL1261"
#> 
#> $GPL1261$taxon
#> [1] "mouse"
#> 
#> $GPL1261$technologyType
#> [1] "ONECOLOR"
#> 
#> $GPL1261$lastGeneMapping
#> NULL
#> 
#> $GPL1261$lastSequenceAnalysis
#> NULL
#> 
#> $GPL1261$lastSequenceUpdate
#> NULL
#> 
#> $GPL1261$numProbeSequences
#> NULL
#> 
#> $GPL1261$numProbeAlignments
#> NULL
#> 
#> $GPL1261$numProbesToGenes
#> NULL
#> 
#> $GPL1261$numGenes
#> NULL
#> 
#> $GPL1261$designElementCount
#> NULL
#> 
#> $GPL1261$dateCached
#> NULL
#> 
#> $GPL1261$color
#> [1] "ONECOLOR"
#> 
#> $GPL1261$expressionExperimentCount
#> [1] 807
#> 
#> $GPL1261$hasBlatAssociations
#> NULL
#> 
#> $GPL1261$hasGeneAssociations
#> NULL
#> 
#> $GPL1261$hasSequenceAssociations
#> NULL
#> 
#> $GPL1261$isMerged
#> NULL
#> 
#> $GPL1261$isMergee
#> [1] TRUE
#> 
#> $GPL1261$isSubsumed
#> NULL
#> 
#> $GPL1261$isSubsumer
#> NULL
#> 
#> $GPL1261$lastRepeatMask
#> NULL
#> 
#> $GPL1261$description
#> [1] " All probe sets represented on the GeneChip Mouse Expression Set 430 are included on the GeneChip Mouse Genome 430 2.0 Array. The sequences from which these probe sets were derived were selected from GenBank«, dbEST, and RefSeq. The sequence clusters were created from the UniGene database (Build 107, June 2002) and then refined by analysis and comparison with the publicly available draft assembly of the mouse genome from the Whitehead Institute for Genome Research (MGSC, April 2002). \nFrom GPL1261\nLast Updated: Oct 28 2005"
#> 
#> $GPL1261$name
#> [1] "Affymetrix GeneChip Mouse Genome 430 2.0 Array"
#> 
#> $GPL1261$troubled
#> [1] FALSE
#> 
#> $GPL1261$lastUpdated
#> [1] 1.48756e+12
#> 
#> $GPL1261$lastTroubledEvent
#> NULL
#> 
#> $GPL1261$troubleDetails
#> [1] "No trouble details provided."
#> 
#> $GPL1261$lastNeedsAttentionEvent
#> NULL
#> 
#> $GPL1261$curationNote
#> NULL
#> 
#> $GPL1261$needsAttention
#> [1] FALSE
#> 
#> $GPL1261$lastNoteUpdateEvent
#> NULL
#> 
#> $GPL1261$id
#> [1] 3
#> 
#> 
#> $GPL570
#> $GPL570$shortName
#> [1] "GPL570"
#> 
#> $GPL570$taxon
#> [1] "human"
#> 
#> $GPL570$technologyType
#> [1] "ONECOLOR"
#> 
#> $GPL570$lastGeneMapping
#> NULL
#> 
#> $GPL570$lastSequenceAnalysis
#> NULL
#> 
#> $GPL570$lastSequenceUpdate
#> NULL
#> 
#> $GPL570$numProbeSequences
#> NULL
#> 
#> $GPL570$numProbeAlignments
#> NULL
#> 
#> $GPL570$numProbesToGenes
#> NULL
#> 
#> $GPL570$numGenes
#> NULL
#> 
#> $GPL570$designElementCount
#> NULL
#> 
#> $GPL570$dateCached
#> NULL
#> 
#> $GPL570$color
#> [1] "ONECOLOR"
#> 
#> $GPL570$expressionExperimentCount
#> [1] 1214
#> 
#> $GPL570$hasBlatAssociations
#> NULL
#> 
#> $GPL570$hasGeneAssociations
#> NULL
#> 
#> $GPL570$hasSequenceAssociations
#> NULL
#> 
#> $GPL570$isMerged
#> NULL
#> 
#> $GPL570$isMergee
#> [1] TRUE
#> 
#> $GPL570$isSubsumed
#> NULL
#> 
#> $GPL570$isSubsumer
#> NULL
#> 
#> $GPL570$lastRepeatMask
#> NULL
#> 
#> $GPL570$description
#> [1] " Complete coverage of the Human Genome U133 Set plus 6,500 additional genes for analysis of over 47,000 transcripts All probe sets represented on the GeneChip Human Genome U133 Set are identically replicated on the GeneChip Human Genome U133 Plus 2.0 Array. The sequences from which these probe sets were derived were selected from GenBank®, dbEST, and RefSeq. The sequence clusters were created from the UniGene database (Build 133, April 20, 2001) and then refined by analysis and comparison with a number of other publicly available databases, including the Washington University EST trace repository and the University of California, Santa Cruz Golden-Path human genome database (April 2001 release). In addition, there are 9,921 new probe sets representing approximately 6,500 new genes. These gene sequences were selected from GenBank, dbEST, and RefSeq. Sequence clusters were created from the UniGene database (Build 159, January 25, 2003) and refined by analysis and comparison with a number of other publicly available databases, including the Washington University EST trace repository and the NCBI human genome assembly (Build 31). \nFrom GPL570\nLast Updated: Mar 09 2006"
#> 
#> $GPL570$name
#> [1] "Affymetrix GeneChip Human Genome U133 Plus 2.0 Array"
#> 
#> $GPL570$troubled
#> [1] FALSE
#> 
#> $GPL570$lastUpdated
#> [1] 1.487818e+12
#> 
#> $GPL570$lastTroubledEvent
#> NULL
#> 
#> $GPL570$troubleDetails
#> [1] "No trouble details provided."
#> 
#> $GPL570$lastNeedsAttentionEvent
#> NULL
#> 
#> $GPL570$curationNote
#> NULL
#> 
#> $GPL570$needsAttention
#> [1] FALSE
#> 
#> $GPL570$lastNoteUpdateEvent
#> NULL
#> 
#> $GPL570$id
#> [1] 4
#> 
#> 
#> $GPL81
#> $GPL81$shortName
#> [1] "GPL81"
#> 
#> $GPL81$taxon
#> [1] "mouse"
#> 
#> $GPL81$technologyType
#> [1] "ONECOLOR"
#> 
#> $GPL81$lastGeneMapping
#> NULL
#> 
#> $GPL81$lastSequenceAnalysis
#> NULL
#> 
#> $GPL81$lastSequenceUpdate
#> NULL
#> 
#> $GPL81$numProbeSequences
#> NULL
#> 
#> $GPL81$numProbeAlignments
#> NULL
#> 
#> $GPL81$numProbesToGenes
#> NULL
#> 
#> $GPL81$numGenes
#> NULL
#> 
#> $GPL81$designElementCount
#> NULL
#> 
#> $GPL81$dateCached
#> NULL
#> 
#> $GPL81$color
#> [1] "ONECOLOR"
#> 
#> $GPL81$expressionExperimentCount
#> [1] 196
#> 
#> $GPL81$hasBlatAssociations
#> NULL
#> 
#> $GPL81$hasGeneAssociations
#> NULL
#> 
#> $GPL81$hasSequenceAssociations
#> NULL
#> 
#> $GPL81$isMerged
#> NULL
#> 
#> $GPL81$isMergee
#> [1] TRUE
#> 
#> $GPL81$isSubsumed
#> NULL
#> 
#> $GPL81$isSubsumer
#> NULL
#> 
#> $GPL81$lastRepeatMask
#> NULL
#> 
#> $GPL81$description
#> [1] " The MG-U74 set includes 3 arrays with a total of 36899 entries and was indexed 29-Jan-2002. The set represents ~36,000 full length genes and EST clusters derived from sequence clusters in Build 74 of the Mouse Unigene Database. MG-U74A represents all sequences (~6,000) in the Mouse UniGene database (Build 74) that have been functionally characterized, as well as ~6,000 EST clusters.  This platform can also be used to submit data for the MG-U74A [version 1] array.  Keywords = high density oligonucleotide array\r\nFrom GPL81\r\nLast Updated: Nov 17 2005"
#> 
#> $GPL81$name
#> [1] "Affymetrix GeneChip Murine Genome U74 Version 2 Set MG-U74A"
#> 
#> $GPL81$troubled
#> [1] FALSE
#> 
#> $GPL81$lastUpdated
#> [1] 1.487561e+12
#> 
#> $GPL81$lastTroubledEvent
#> NULL
#> 
#> $GPL81$troubleDetails
#> [1] "No trouble details provided."
#> 
#> $GPL81$lastNeedsAttentionEvent
#> NULL
#> 
#> $GPL81$curationNote
#> NULL
#> 
#> $GPL81$needsAttention
#> [1] FALSE
#> 
#> $GPL81$lastNoteUpdateEvent
#> NULL
#> 
#> $GPL81$id
#> [1] 5
#> 
#> 
#> $GPL85
#> $GPL85$shortName
#> [1] "GPL85"
#> 
#> $GPL85$taxon
#> [1] "rat"
#> 
#> $GPL85$technologyType
#> [1] "ONECOLOR"
#> 
#> $GPL85$lastGeneMapping
#> NULL
#> 
#> $GPL85$lastSequenceAnalysis
#> NULL
#> 
#> $GPL85$lastSequenceUpdate
#> NULL
#> 
#> $GPL85$numProbeSequences
#> NULL
#> 
#> $GPL85$numProbeAlignments
#> NULL
#> 
#> $GPL85$numProbesToGenes
#> NULL
#> 
#> $GPL85$numGenes
#> NULL
#> 
#> $GPL85$designElementCount
#> NULL
#> 
#> $GPL85$dateCached
#> NULL
#> 
#> $GPL85$color
#> [1] "ONECOLOR"
#> 
#> $GPL85$expressionExperimentCount
#> [1] 83
#> 
#> $GPL85$hasBlatAssociations
#> NULL
#> 
#> $GPL85$hasGeneAssociations
#> NULL
#> 
#> $GPL85$hasSequenceAssociations
#> NULL
#> 
#> $GPL85$isMerged
#> NULL
#> 
#> $GPL85$isMergee
#> [1] FALSE
#> 
#> $GPL85$isSubsumed
#> NULL
#> 
#> $GPL85$isSubsumer
#> NULL
#> 
#> $GPL85$lastRepeatMask
#> NULL
#> 
#> $GPL85$description
#> [1] " The RG-U34 set includes 3 arrays with a total of 26379 entries and was indexed 29-Jan-2002. The set includes ~7,000 known genes and >17,000 EST clusters. The A array contains probes derived from all full-length or annotated genes as well as thousands of EST clusters. The B and C arrays contain only EST clusters.  Keywords = high density oligonucleotide array\nFrom GPL85\nLast Updated: Mar 09 2006"
#> 
#> $GPL85$name
#> [1] "Affymetrix GeneChip Rat Genome U34 Array Set RG-U34A"
#> 
#> $GPL85$troubled
#> [1] FALSE
#> 
#> $GPL85$lastUpdated
#> [1] 1.488129e+12
#> 
#> $GPL85$lastTroubledEvent
#> NULL
#> 
#> $GPL85$troubleDetails
#> [1] "No trouble details provided."
#> 
#> $GPL85$lastNeedsAttentionEvent
#> NULL
#> 
#> $GPL85$curationNote
#> NULL
#> 
#> $GPL85$needsAttention
#> [1] FALSE
#> 
#> $GPL85$lastNoteUpdateEvent
#> NULL
#> 
#> $GPL85$id
#> [1] 6
#> 
#> 
#> $GPL339
#> $GPL339$shortName
#> [1] "GPL339"
#> 
#> $GPL339$taxon
#> [1] "mouse"
#> 
#> $GPL339$technologyType
#> [1] "ONECOLOR"
#> 
#> $GPL339$lastGeneMapping
#> NULL
#> 
#> $GPL339$lastSequenceAnalysis
#> NULL
#> 
#> $GPL339$lastSequenceUpdate
#> NULL
#> 
#> $GPL339$numProbeSequences
#> NULL
#> 
#> $GPL339$numProbeAlignments
#> NULL
#> 
#> $GPL339$numProbesToGenes
#> NULL
#> 
#> $GPL339$numGenes
#> NULL
#> 
#> $GPL339$designElementCount
#> NULL
#> 
#> $GPL339$dateCached
#> NULL
#> 
#> $GPL339$color
#> [1] "ONECOLOR"
#> 
#> $GPL339$expressionExperimentCount
#> [1] 130
#> 
#> $GPL339$hasBlatAssociations
#> NULL
#> 
#> $GPL339$hasGeneAssociations
#> NULL
#> 
#> $GPL339$hasSequenceAssociations
#> NULL
#> 
#> $GPL339$isMerged
#> NULL
#> 
#> $GPL339$isMergee
#> [1] TRUE
#> 
#> $GPL339$isSubsumed
#> NULL
#> 
#> $GPL339$isSubsumer
#> NULL
#> 
#> $GPL339$lastRepeatMask
#> NULL
#> 
#> $GPL339$description
#> [1] "The GeneChip Mouse Genome 430A 2.0 Array is a single array representing approximately 14,000 well-characterized mouse genes that can be used to explore biology and disease processes.  Provides coverage of well-substantiated genes in the transcribed mouse genome on a single array.  Analyzes the expression level of approximately 14,000 well-characterized mouse genes.  All probe sets represented on the GeneChip Mouse Expression Array 430A are included on the GeneChip Mouse Genome 430A 2.0 Array  Sequences used in the design of the array were selected from GenBank®, dbEST, and RefSeq. The sequence clusters were created from the UniGene database (Build 107, June 2002) and then refined by analysis and comparison with the publicly available draft assembly of the mouse genome from the Whitehead Institute for Genome Research (MGSC, April 2002). \r\nFrom GPL339\r\nLast Updated: Oct 28 2005\r\n\t\t\t\t\t"
#> 
#> $GPL339$name
#> [1] "Affymetrix GeneChip Mouse Expression Array 430A and Mouse Genome 430A 2.0 Array"
#> 
#> $GPL339$troubled
#> [1] FALSE
#> 
#> $GPL339$lastUpdated
#> [1] 1.487563e+12
#> 
#> $GPL339$lastTroubledEvent
#> NULL
#> 
#> $GPL339$troubleDetails
#> [1] "No trouble details provided."
#> 
#> $GPL339$lastNeedsAttentionEvent
#> NULL
#> 
#> $GPL339$curationNote
#> NULL
#> 
#> $GPL339$needsAttention
#> [1] FALSE
#> 
#> $GPL339$lastNoteUpdateEvent
#> NULL
#> 
#> $GPL339$id
#> [1] 7
#> 
#> 
#> $GPL91
#> $GPL91$shortName
#> [1] "GPL91"
#> 
#> $GPL91$taxon
#> [1] "human"
#> 
#> $GPL91$technologyType
#> [1] "ONECOLOR"
#> 
#> $GPL91$lastGeneMapping
#> NULL
#> 
#> $GPL91$lastSequenceAnalysis
#> NULL
#> 
#> $GPL91$lastSequenceUpdate
#> NULL
#> 
#> $GPL91$numProbeSequences
#> NULL
#> 
#> $GPL91$numProbeAlignments
#> NULL
#> 
#> $GPL91$numProbesToGenes
#> NULL
#> 
#> $GPL91$numGenes
#> NULL
#> 
#> $GPL91$designElementCount
#> NULL
#> 
#> $GPL91$dateCached
#> NULL
#> 
#> $GPL91$color
#> [1] "ONECOLOR"
#> 
#> $GPL91$expressionExperimentCount
#> [1] 63
#> 
#> $GPL91$hasBlatAssociations
#> NULL
#> 
#> $GPL91$hasGeneAssociations
#> NULL
#> 
#> $GPL91$hasSequenceAssociations
#> NULL
#> 
#> $GPL91$isMerged
#> NULL
#> 
#> $GPL91$isMergee
#> [1] TRUE
#> 
#> $GPL91$isSubsumed
#> NULL
#> 
#> $GPL91$isSubsumer
#> NULL
#> 
#> $GPL91$lastRepeatMask
#> NULL
#> 
#> $GPL91$description
#> [1] " The Human Genome U95 (HG-U95) Set, consisting of five GeneChip arrays, contains almost 63,000 probe sets interrogating approximately 54,000 clusters derived from the UniGene database (Build 95). Based on this build and associated annotations, the HG-U95Av2 Array represents approximately 10,000 full-length genes, while arrays B through E represent EST clusters.  This platform can also be used to submit data for the HG_U95A [version 1] array.  Keywords = high density oligonucleotide array\nFrom GPL91\nLast Updated: Mar 09 2006"
#> 
#> $GPL91$name
#> [1] "Affymetrix GeneChip Human Genome U95 Version [1 or 2] Set HG-U95A"
#> 
#> $GPL91$troubled
#> [1] FALSE
#> 
#> $GPL91$lastUpdated
#> [1] 1.487819e+12
#> 
#> $GPL91$lastTroubledEvent
#> NULL
#> 
#> $GPL91$troubleDetails
#> [1] "No trouble details provided."
#> 
#> $GPL91$lastNeedsAttentionEvent
#> NULL
#> 
#> $GPL91$curationNote
#> NULL
#> 
#> $GPL91$needsAttention
#> [1] FALSE
#> 
#> $GPL91$lastNoteUpdateEvent
#> NULL
#> 
#> $GPL91$id
#> [1] 8
#> 
#> 
#> $GPL890
#> $GPL890$shortName
#> [1] "GPL890"
#> 
#> $GPL890$taxon
#> [1] "rat"
#> 
#> $GPL890$technologyType
#> [1] "TWOCOLOR"
#> 
#> $GPL890$lastGeneMapping
#> NULL
#> 
#> $GPL890$lastSequenceAnalysis
#> NULL
#> 
#> $GPL890$lastSequenceUpdate
#> NULL
#> 
#> $GPL890$numProbeSequences
#> NULL
#> 
#> $GPL890$numProbeAlignments
#> NULL
#> 
#> $GPL890$numProbesToGenes
#> NULL
#> 
#> $GPL890$numGenes
#> NULL
#> 
#> $GPL890$designElementCount
#> NULL
#> 
#> $GPL890$dateCached
#> NULL
#> 
#> $GPL890$color
#> [1] "TWOCOLOR"
#> 
#> $GPL890$expressionExperimentCount
#> [1] 14
#> 
#> $GPL890$hasBlatAssociations
#> NULL
#> 
#> $GPL890$hasGeneAssociations
#> NULL
#> 
#> $GPL890$hasSequenceAssociations
#> NULL
#> 
#> $GPL890$isMerged
#> NULL
#> 
#> $GPL890$isMergee
#> [1] FALSE
#> 
#> $GPL890$isSubsumed
#> NULL
#> 
#> $GPL890$isSubsumer
#> NULL
#> 
#> $GPL890$lastRepeatMask
#> NULL
#> 
#> $GPL890$description
#> [1] "  This microarray contains representative sequences derived from rat RefSeq, high-quality public rat mRNA accessions, homologs to human and mouse RefSeq, EST clusters determined by a leading biopharmaceutical company and the Incyte Zooseq database as well as toxicology markers determined by the National Institute of Environmental Health Sciences (NIEHS) and others. Many of the toxicology markers used in the design of this microarray were released to the public domain for the first time and will be used on this unique microarray product.  Arrays of this design have barcodes that begin with 16011868 or 2511868.  Orientation: Features are numbered numbered Left-to-Right, Top-to-Bottom as scanned by an Agilent scanner (barcode on the left, DNA on the back surface, scanned through the glass), matching the FeatureNum output from Agilent's Feature Extraction software.  The ID column represents the Agilent Feature Extraction feature number.  Rows and columns are numbered as scanned by an Axon Scanner (barcode on the bottom, DNA on the front surface).  To match data scanned on an Axon scanner, use the RefNumber column contained in the Agilent-provided GAL file as the ID_REF column in sample submissions. \nFrom GPL890\nLast Updated: Sep 22 2005"
#> 
#> $GPL890$name
#> [1] "Agilent-011868 Rat Oligo Microarray (G4130A)"
#> 
#> $GPL890$troubled
#> [1] FALSE
#> 
#> $GPL890$lastUpdated
#> [1] 1.48813e+12
#> 
#> $GPL890$lastTroubledEvent
#> NULL
#> 
#> $GPL890$troubleDetails
#> [1] "No trouble details provided."
#> 
#> $GPL890$lastNeedsAttentionEvent
#> NULL
#> 
#> $GPL890$curationNote
#> NULL
#> 
#> $GPL890$needsAttention
#> [1] FALSE
#> 
#> $GPL890$lastNoteUpdateEvent
#> NULL
#> 
#> $GPL890$id
#> [1] 9
#> 
#> 
#> $GPL1406
#> $GPL1406$shortName
#> [1] "GPL1406"
#> 
#> $GPL1406$taxon
#> [1] "mouse"
#> 
#> $GPL1406$technologyType
#> [1] "TWOCOLOR"
#> 
#> $GPL1406$lastGeneMapping
#> NULL
#> 
#> $GPL1406$lastSequenceAnalysis
#> NULL
#> 
#> $GPL1406$lastSequenceUpdate
#> NULL
#> 
#> $GPL1406$numProbeSequences
#> NULL
#> 
#> $GPL1406$numProbeAlignments
#> NULL
#> 
#> $GPL1406$numProbesToGenes
#> NULL
#> 
#> $GPL1406$numGenes
#> NULL
#> 
#> $GPL1406$designElementCount
#> NULL
#> 
#> $GPL1406$dateCached
#> NULL
#> 
#> $GPL1406$color
#> [1] "TWOCOLOR"
#> 
#> $GPL1406$expressionExperimentCount
#> [1] 2
#> 
#> $GPL1406$hasBlatAssociations
#> NULL
#> 
#> $GPL1406$hasGeneAssociations
#> NULL
#> 
#> $GPL1406$hasSequenceAssociations
#> NULL
#> 
#> $GPL1406$isMerged
#> NULL
#> 
#> $GPL1406$isMergee
#> [1] FALSE
#> 
#> $GPL1406$isSubsumed
#> NULL
#> 
#> $GPL1406$isSubsumer
#> NULL
#> 
#> $GPL1406$lastRepeatMask
#> NULL
#> 
#> $GPL1406$description
#> [1] " IncyteMouseGEM1 (127 cloneID)\nFrom GPL1406\nLast Updated: May 26 2005"
#> 
#> $GPL1406$name
#> [1] "IncyteMouseGEM1"
#> 
#> $GPL1406$troubled
#> [1] FALSE
#> 
#> $GPL1406$lastUpdated
#> [1] 1.487564e+12
#> 
#> $GPL1406$lastTroubledEvent
#> NULL
#> 
#> $GPL1406$troubleDetails
#> [1] "No trouble details provided."
#> 
#> $GPL1406$lastNeedsAttentionEvent
#> NULL
#> 
#> $GPL1406$curationNote
#> NULL
#> 
#> $GPL1406$needsAttention
#> [1] FALSE
#> 
#> $GPL1406$lastNoteUpdateEvent
#> NULL
#> 
#> $GPL1406$id
#> [1] 10
#> 
#> 
#> $GPL891
#> $GPL891$shortName
#> [1] "GPL891"
#> 
#> $GPL891$taxon
#> [1] "mouse"
#> 
#> $GPL891$technologyType
#> [1] "TWOCOLOR"
#> 
#> $GPL891$lastGeneMapping
#> NULL
#> 
#> $GPL891$lastSequenceAnalysis
#> NULL
#> 
#> $GPL891$lastSequenceUpdate
#> NULL
#> 
#> $GPL891$numProbeSequences
#> NULL
#> 
#> $GPL891$numProbeAlignments
#> NULL
#> 
#> $GPL891$numProbesToGenes
#> NULL
#> 
#> $GPL891$numGenes
#> NULL
#> 
#> $GPL891$designElementCount
#> NULL
#> 
#> $GPL891$dateCached
#> NULL
#> 
#> $GPL891$color
#> [1] "TWOCOLOR"
#> 
#> $GPL891$expressionExperimentCount
#> [1] 18
#> 
#> $GPL891$hasBlatAssociations
#> NULL
#> 
#> $GPL891$hasGeneAssociations
#> NULL
#> 
#> $GPL891$hasSequenceAssociations
#> NULL
#> 
#> $GPL891$isMerged
#> NULL
#> 
#> $GPL891$isMergee
#> [1] FALSE
#> 
#> $GPL891$isSubsumed
#> NULL
#> 
#> $GPL891$isSubsumer
#> NULL
#> 
#> $GPL891$lastRepeatMask
#> NULL
#> 
#> $GPL891$description
#> [1] "  Mouse is the most widely used mammalian model in biological research and target identification. This microarray will enable researchers to simultaneously characterize expression activities of many thousands of mouse genes associated with various biological functions and processes. Content for this empirically validated microarray includes a comprehensive set of toxicology markers determined through collaboration with leaders in toxicogenomics including NIEHS, Toxicogenomics Research Consortium, and Paradigm Genetics  Arrays of this design have barcodes that begin with 16011978 or 2511978.  Orientation: Features are numbered numbered Left-to-Right, Top-to-Bottom as scanned by an Agilent scanner (barcode on the left, DNA on the back surface, scanned through the glass), matching the FeatureNum output from Agilent's Feature Extraction software.  The ID column represents the Agilent Feature Extraction feature number.  Rows and columns are numbered as scanned by an Axon Scanner (barcode on the bottom, DNA on the front surface).  To match data scanned on an Axon scanner, use the RefNumber column contained in the Agilent-provided GAL file as the ID_REF column in sample submissions.   Mouse is the most widely used mammalian model in biological research and target identification. This microarray will enable researchers to simultaneously characterize expression activities of many thousands of mouse genes associated with various biological functions and processes. Content for this empirically validated microarray includes a comprehensive set of toxicology markers determined through collaboration with leaders in toxicogenomics including NIEHS, Toxicogenomics Research Consortium, and Paradigm Genetics  Arrays of this design have barcodes that begin with 16011978 or 2511978.  Orientation: Features are numbered numbered Left-to-Right, Top-to-Bottom as scanned by an Agilent scanner (barcode on the left, DNA on the back surface, scanned through the glass), matching the FeatureNum output from Agilent's Feature Extraction software.  The ID column represents the Agilent Feature Extraction feature number.  Rows and columns are numbered as scanned by an Axon Scanner (barcode on the bottom, DNA on the front surface).  To match data scanned on an Axon scanner, use the RefNumber column contained in the Agilent-provided GAL file as the ID_REF column in sample submissions. \nFrom GPL891\nLast Updated: Aug 17 2006"
#> 
#> $GPL891$name
#> [1] "Agilent-011978 Mouse Microarray G4121A"
#> 
#> $GPL891$troubled
#> [1] FALSE
#> 
#> $GPL891$lastUpdated
#> [1] 1.487566e+12
#> 
#> $GPL891$lastTroubledEvent
#> NULL
#> 
#> $GPL891$troubleDetails
#> [1] "No trouble details provided."
#> 
#> $GPL891$lastNeedsAttentionEvent
#> NULL
#> 
#> $GPL891$curationNote
#> NULL
#> 
#> $GPL891$needsAttention
#> [1] FALSE
#> 
#> $GPL891$lastNoteUpdateEvent
#> NULL
#> 
#> $GPL891$id
#> [1] 11
#> 
#> 
#> $GPL82
#> $GPL82$shortName
#> [1] "GPL82"
#> 
#> $GPL82$taxon
#> [1] "mouse"
#> 
#> $GPL82$technologyType
#> [1] "ONECOLOR"
#> 
#> $GPL82$lastGeneMapping
#> NULL
#> 
#> $GPL82$lastSequenceAnalysis
#> NULL
#> 
#> $GPL82$lastSequenceUpdate
#> NULL
#> 
#> $GPL82$numProbeSequences
#> NULL
#> 
#> $GPL82$numProbeAlignments
#> NULL
#> 
#> $GPL82$numProbesToGenes
#> NULL
#> 
#> $GPL82$numGenes
#> NULL
#> 
#> $GPL82$designElementCount
#> NULL
#> 
#> $GPL82$dateCached
#> NULL
#> 
#> $GPL82$color
#> [1] "ONECOLOR"
#> 
#> $GPL82$expressionExperimentCount
#> [1] 24
#> 
#> $GPL82$hasBlatAssociations
#> NULL
#> 
#> $GPL82$hasGeneAssociations
#> NULL
#> 
#> $GPL82$hasSequenceAssociations
#> NULL
#> 
#> $GPL82$isMerged
#> NULL
#> 
#> $GPL82$isMergee
#> [1] FALSE
#> 
#> $GPL82$isSubsumed
#> NULL
#> 
#> $GPL82$isSubsumer
#> NULL
#> 
#> $GPL82$lastRepeatMask
#> NULL
#> 
#> $GPL82$description
#> [1] " The MG-U74 set includes 3 arrays with a total of 36899 entries and was indexed 29-Jan-2002. The set represents ~36,000 full length genes and EST clusters derived from sequence clusters in Build 74 of the Mouse Unigene Database.  This platform can also be used to submit data for the MG-U74B [version 1] array.  Keywords = high density oligonucleotide array\nFrom GPL82\nLast Updated: Oct 25 2005"
#> 
#> $GPL82$name
#> [1] "Affymetrix GeneChip Murine Genome U74 Version 2 Set MG-U74B"
#> 
#> $GPL82$troubled
#> [1] FALSE
#> 
#> $GPL82$lastUpdated
#> [1] 1.487566e+12
#> 
#> $GPL82$lastTroubledEvent
#> NULL
#> 
#> $GPL82$troubleDetails
#> [1] "No trouble details provided."
#> 
#> $GPL82$lastNeedsAttentionEvent
#> NULL
#> 
#> $GPL82$curationNote
#> NULL
#> 
#> $GPL82$needsAttention
#> [1] FALSE
#> 
#> $GPL82$lastNoteUpdateEvent
#> NULL
#> 
#> $GPL82$id
#> [1] 12
#> 
#> 
#> $GPL560
#> $GPL560$shortName
#> [1] "GPL560"
#> 
#> $GPL560$taxon
#> [1] "mouse"
#> 
#> $GPL560$technologyType
#> [1] "TWOCOLOR"
#> 
#> $GPL560$lastGeneMapping
#> NULL
#> 
#> $GPL560$lastSequenceAnalysis
#> NULL
#> 
#> $GPL560$lastSequenceUpdate
#> NULL
#> 
#> $GPL560$numProbeSequences
#> NULL
#> 
#> $GPL560$numProbeAlignments
#> NULL
#> 
#> $GPL560$numProbesToGenes
#> NULL
#> 
#> $GPL560$numGenes
#> NULL
#> 
#> $GPL560$designElementCount
#> NULL
#> 
#> $GPL560$dateCached
#> NULL
#> 
#> $GPL560$color
#> [1] "TWOCOLOR"
#> 
#> $GPL560$expressionExperimentCount
#> [1] 8
#> 
#> $GPL560$hasBlatAssociations
#> NULL
#> 
#> $GPL560$hasGeneAssociations
#> NULL
#> 
#> $GPL560$hasSequenceAssociations
#> NULL
#> 
#> $GPL560$isMerged
#> NULL
#> 
#> $GPL560$isMergee
#> [1] FALSE
#> 
#> $GPL560$isSubsumed
#> NULL
#> 
#> $GPL560$isSubsumer
#> NULL
#> 
#> $GPL560$lastRepeatMask
#> NULL
#> 
#> $GPL560$description
#> [1] " Developmental Toxicity of the Mouse Embryo (DTME) includes expression data from early embryos at risk for teratogen-induced eye malformations. Early mouse embryos were exposed to various teratogens during neurulation stages with the aim of correlating large-scale changes in gene expression with a generic histopathological biomarker (p53 protein induction) across the critical period during exposure, and with the risk of malformations in term fetuses (the various exposure scenarios produce ~10% risk of malformations in term fetuses averaged across 50 conditions of the experiment). Microarray analysis was performed on RNA from the embryo, prosencephalon, or headfold. Each condition for microarray analysis was replicated with independent sampling and reversal of labeling assignments (100 samples total). DTME compares genes expressed between test vs. reference samples as evidenced by data from two-channel MICROMAX system from PerkinElmer Life Sciences (Wellesley, MA 02481-4078, USA). It also includes a series of comparisons of early optic development in the optic primordium from mouse and rat embryos harvested at equivalent morphogenetic stages. The microarray platform uses a novel integrative chemistry that culminates massive signal amplification and greatly enhanced detection. The method incorporates unique two-channel indirect labeling of target RNA products with biotinyl-11-dCTP and fluorescein-12-dCTP (or dinitrophenol-11-dCTP) with post-hybridization amplification through conventional immunodetection and tyramide signal amplification. The probe was spotted with 2.4K sequence-verified human gene elements (single or duplicate) from 50 different cDNA libraries and positive- and negative- control genes. The cDNAs represent more than 10 different tissue sources: 80% brain-derived and more than 40% full-length cDNA sequences and a range of human diseases, metabolic, and regulatory pathways. The probe has been shown effective with target RNA from human, mouse, or rat biosources. The dataset includes \"absolute\" and \"relative\" values. Absolute values represent the background-corrected signal. For single-spot 2400 arrays this means one measurement for every spotted element and each color channel. Relative values represent ratiometric data normalized as follows: transformation to log2-space; lowess correction of MxA plot; internal standardization by mean absolute deviation; and normalization between replica pairs in each condition. Keywords = mouse Keywords = embryo Keywords = teratogen Keywords = toxicity Keywords = development Keywords = eye\nFrom GPL560\nLast Updated: May 26 2005"
#> 
#> $GPL560$name
#> [1] "Developmental Toxicity of the Mouse Embryo (DTME)"
#> 
#> $GPL560$troubled
#> [1] FALSE
#> 
#> $GPL560$lastUpdated
#> [1] 1.487566e+12
#> 
#> $GPL560$lastTroubledEvent
#> NULL
#> 
#> $GPL560$troubleDetails
#> [1] "No trouble details provided."
#> 
#> $GPL560$lastNeedsAttentionEvent
#> NULL
#> 
#> $GPL560$curationNote
#> NULL
#> 
#> $GPL560$needsAttention
#> [1] FALSE
#> 
#> $GPL560$lastNoteUpdateEvent
#> NULL
#> 
#> $GPL560$id
#> [1] 13
#> 
#> 
#> $GPL1073
#> $GPL1073$shortName
#> [1] "GPL1073"
#> 
#> $GPL1073$taxon
#> [1] "mouse"
#> 
#> $GPL1073$technologyType
#> [1] "ONECOLOR"
#> 
#> $GPL1073$lastGeneMapping
#> NULL
#> 
#> $GPL1073$lastSequenceAnalysis
#> NULL
#> 
#> $GPL1073$lastSequenceUpdate
#> NULL
#> 
#> $GPL1073$numProbeSequences
#> NULL
#> 
#> $GPL1073$numProbeAlignments
#> NULL
#> 
#> $GPL1073$numProbesToGenes
#> NULL
#> 
#> $GPL1073$numGenes
#> NULL
#> 
#> $GPL1073$designElementCount
#> NULL
#> 
#> $GPL1073$dateCached
#> NULL
#> 
#> $GPL1073$color
#> [1] "ONECOLOR"
#> 
#> $GPL1073$expressionExperimentCount
#> [1] 5
#> 
#> $GPL1073$hasBlatAssociations
#> NULL
#> 
#> $GPL1073$hasGeneAssociations
#> NULL
#> 
#> $GPL1073$hasSequenceAssociations
#> NULL
#> 
#> $GPL1073$isMerged
#> NULL
#> 
#> $GPL1073$isMergee
#> [1] FALSE
#> 
#> $GPL1073$isSubsumed
#> NULL
#> 
#> $GPL1073$isSubsumer
#> NULL
#> 
#> $GPL1073$lastRepeatMask
#> NULL
#> 
#> $GPL1073$description
#> [1] " GNF1M We identified a non-redundant set of target sequences for the mouse using the following sources: RefSeq (12,029 sequences), Celera (29,331 sequences), and RIKEN (46,299 sequences).  First, all sequences were RepeatMasked to remove repetitive elements.  Next, sequence identity between individual sequences was established using pairwise BLAT or BLAST and sim4.  The results from single-linkage clustering were further triaged to produce a final target set of 36,182 targets with the highest degree of confidence of computational prediction (biasing toward sequences containing Interpro domains and away from non-coding RNAs).  The final target sets were submitted to the Affymetrix's chip design pipeline for fabrication of the arrays. For data submission here, only probe sets that mapped to public sequences are given.\nFrom GPL1073\nLast Updated: Jun 16 2005"
#> 
#> $GPL1073$name
#> [1] "GNF1M"
#> 
#> $GPL1073$troubled
#> [1] FALSE
#> 
#> $GPL1073$lastUpdated
#> [1] 1.487569e+12
#> 
#> $GPL1073$lastTroubledEvent
#> NULL
#> 
#> $GPL1073$troubleDetails
#> [1] "No trouble details provided."
#> 
#> $GPL1073$lastNeedsAttentionEvent
#> NULL
#> 
#> $GPL1073$curationNote
#> NULL
#> 
#> $GPL1073$needsAttention
#> [1] FALSE
#> 
#> $GPL1073$lastNoteUpdateEvent
#> NULL
#> 
#> $GPL1073$id
#> [1] 14
#> 
#> 
#> $GPL369
#> $GPL369$shortName
#> [1] "GPL369"
#> 
#> $GPL369$taxon
#> [1] "mouse"
#> 
#> $GPL369$technologyType
#> [1] "TWOCOLOR"
#> 
#> $GPL369$lastGeneMapping
#> NULL
#> 
#> $GPL369$lastSequenceAnalysis
#> NULL
#> 
#> $GPL369$lastSequenceUpdate
#> NULL
#> 
#> $GPL369$numProbeSequences
#> NULL
#> 
#> $GPL369$numProbeAlignments
#> NULL
#> 
#> $GPL369$numProbesToGenes
#> NULL
#> 
#> $GPL369$numGenes
#> NULL
#> 
#> $GPL369$designElementCount
#> NULL
#> 
#> $GPL369$dateCached
#> NULL
#> 
#> $GPL369$color
#> [1] "TWOCOLOR"
#> 
#> $GPL369$expressionExperimentCount
#> [1] 1
#> 
#> $GPL369$hasBlatAssociations
#> NULL
#> 
#> $GPL369$hasGeneAssociations
#> NULL
#> 
#> $GPL369$hasSequenceAssociations
#> NULL
#> 
#> $GPL369$isMerged
#> NULL
#> 
#> $GPL369$isMergee
#> [1] FALSE
#> 
#> $GPL369$isSubsumed
#> NULL
#> 
#> $GPL369$isSubsumer
#> NULL
#> 
#> $GPL369$lastRepeatMask
#> NULL
#> 
#> $GPL369$description
#> [1] " 9k mouse cDNA microarray printed at Albert Einstein College of Medicine\nFrom GPL369\nLast Updated: May 26 2005"
#> 
#> $GPL369$name
#> [1] "AECOM MOUSE 9K CHIP"
#> 
#> $GPL369$troubled
#> [1] FALSE
#> 
#> $GPL369$lastUpdated
#> [1] 1.48757e+12
#> 
#> $GPL369$lastTroubledEvent
#> NULL
#> 
#> $GPL369$troubleDetails
#> [1] "No trouble details provided."
#> 
#> $GPL369$lastNeedsAttentionEvent
#> NULL
#> 
#> $GPL369$curationNote
#> NULL
#> 
#> $GPL369$needsAttention
#> [1] FALSE
#> 
#> $GPL369$lastNoteUpdateEvent
#> NULL
#> 
#> $GPL369$id
#> [1] 16
#> 
#> 
#> $GPL83
#> $GPL83$shortName
#> [1] "GPL83"
#> 
#> $GPL83$taxon
#> [1] "mouse"
#> 
#> $GPL83$technologyType
#> [1] "ONECOLOR"
#> 
#> $GPL83$lastGeneMapping
#> NULL
#> 
#> $GPL83$lastSequenceAnalysis
#> NULL
#> 
#> $GPL83$lastSequenceUpdate
#> NULL
#> 
#> $GPL83$numProbeSequences
#> NULL
#> 
#> $GPL83$numProbeAlignments
#> NULL
#> 
#> $GPL83$numProbesToGenes
#> NULL
#> 
#> $GPL83$numGenes
#> NULL
#> 
#> $GPL83$designElementCount
#> NULL
#> 
#> $GPL83$dateCached
#> NULL
#> 
#> $GPL83$color
#> [1] "ONECOLOR"
#> 
#> $GPL83$expressionExperimentCount
#> [1] 19
#> 
#> $GPL83$hasBlatAssociations
#> NULL
#> 
#> $GPL83$hasGeneAssociations
#> NULL
#> 
#> $GPL83$hasSequenceAssociations
#> NULL
#> 
#> $GPL83$isMerged
#> NULL
#> 
#> $GPL83$isMergee
#> [1] FALSE
#> 
#> $GPL83$isSubsumed
#> NULL
#> 
#> $GPL83$isSubsumer
#> NULL
#> 
#> $GPL83$lastRepeatMask
#> NULL
#> 
#> $GPL83$description
#> [1] " The MG-U74 set includes 3 arrays with a total of 36899 entries and was indexed 29-Jan-2002. The set represents ~36,000 full length genes and EST clusters derived from sequence clusters in Build 74 of the Mouse Unigene Database.  This platform can also be used to submit data for the MG-U74C [version 1] array.  Keywords = high density oligonucleotide array\nFrom GPL83\nLast Updated: Jul 28 2006"
#> 
#> $GPL83$name
#> [1] "Affymetrix GeneChip Murine Genome U74 Version 2 Set MG-U74C"
#> 
#> $GPL83$troubled
#> [1] FALSE
#> 
#> $GPL83$lastUpdated
#> [1] 1.48757e+12
#> 
#> $GPL83$lastTroubledEvent
#> NULL
#> 
#> $GPL83$troubleDetails
#> [1] "No trouble details provided."
#> 
#> $GPL83$lastNeedsAttentionEvent
#> NULL
#> 
#> $GPL83$curationNote
#> NULL
#> 
#> $GPL83$needsAttention
#> [1] FALSE
#> 
#> $GPL83$lastNoteUpdateEvent
#> NULL
#> 
#> $GPL83$id
#> [1] 17
#> 
#> 
#> $GPL226
#> $GPL226$shortName
#> [1] "GPL226"
#> 
#> $GPL226$taxon
#> [1] "mouse"
#> 
#> $GPL226$technologyType
#> [1] "TWOCOLOR"
#> 
#> $GPL226$lastGeneMapping
#> NULL
#> 
#> $GPL226$lastSequenceAnalysis
#> NULL
#> 
#> $GPL226$lastSequenceUpdate
#> NULL
#> 
#> $GPL226$numProbeSequences
#> NULL
#> 
#> $GPL226$numProbeAlignments
#> NULL
#> 
#> $GPL226$numProbesToGenes
#> NULL
#> 
#> $GPL226$numGenes
#> NULL
#> 
#> $GPL226$designElementCount
#> NULL
#> 
#> $GPL226$dateCached
#> NULL
#> 
#> $GPL226$color
#> [1] "TWOCOLOR"
#> 
#> $GPL226$expressionExperimentCount
#> [1] 1
#> 
#> $GPL226$hasBlatAssociations
#> NULL
#> 
#> $GPL226$hasGeneAssociations
#> NULL
#> 
#> $GPL226$hasSequenceAssociations
#> NULL
#> 
#> $GPL226$isMerged
#> NULL
#> 
#> $GPL226$isMergee
#> [1] FALSE
#> 
#> $GPL226$isSubsumed
#> NULL
#> 
#> $GPL226$isSubsumer
#> NULL
#> 
#> $GPL226$lastRepeatMask
#> NULL
#> 
#> $GPL226$description
#> [1] " 956 clones, mostly from the NIA 15k library were spotted in triplicates\nFrom GPL226\nLast Updated: May 26 2005"
#> 
#> $GPL226$name
#> [1] "MS1: M1K array"
#> 
#> $GPL226$troubled
#> [1] FALSE
#> 
#> $GPL226$lastUpdated
#> [1] 1.487571e+12
#> 
#> $GPL226$lastTroubledEvent
#> NULL
#> 
#> $GPL226$troubleDetails
#> [1] "No trouble details provided."
#> 
#> $GPL226$lastNeedsAttentionEvent
#> NULL
#> 
#> $GPL226$curationNote
#> NULL
#> 
#> $GPL226$needsAttention
#> [1] FALSE
#> 
#> $GPL226$lastNoteUpdateEvent
#> NULL
#> 
#> $GPL226$id
#> [1] 18
#> 
#> 
#> $GPL76
#> $GPL76$shortName
#> [1] "GPL76"
#> 
#> $GPL76$taxon
#> [1] "mouse"
#> 
#> $GPL76$technologyType
#> [1] "ONECOLOR"
#> 
#> $GPL76$lastGeneMapping
#> NULL
#> 
#> $GPL76$lastSequenceAnalysis
#> NULL
#> 
#> $GPL76$lastSequenceUpdate
#> NULL
#> 
#> $GPL76$numProbeSequences
#> NULL
#> 
#> $GPL76$numProbeAlignments
#> NULL
#> 
#> $GPL76$numProbesToGenes
#> NULL
#> 
#> $GPL76$numGenes
#> NULL
#> 
#> $GPL76$designElementCount
#> NULL
#> 
#> $GPL76$dateCached
#> NULL
#> 
#> $GPL76$color
#> [1] "ONECOLOR"
#> 
#> $GPL76$expressionExperimentCount
#> [1] 9
#> 
#> $GPL76$hasBlatAssociations
#> NULL
#> 
#> $GPL76$hasGeneAssociations
#> NULL
#> 
#> $GPL76$hasSequenceAssociations
#> NULL
#> 
#> $GPL76$isMerged
#> NULL
#> 
#> $GPL76$isMergee
#> [1] TRUE
#> 
#> $GPL76$isSubsumed
#> NULL
#> 
#> $GPL76$isSubsumer
#> NULL
#> 
#> $GPL76$lastRepeatMask
#> NULL
#> 
#> $GPL76$description
#> [1] " Has 13179 entries and was indexed 29-Jan-2002. Keywords = high density oligonucleotide array\nFrom GPL76\nLast Updated: Jul 28 2006"
#> 
#> $GPL76$name
#> [1] "Affymetrix GeneChip Murine 11K SubB Array Mu11K-B"
#> 
#> $GPL76$troubled
#> [1] FALSE
#> 
#> $GPL76$lastUpdated
#> [1] 1.487571e+12
#> 
#> $GPL76$lastTroubledEvent
#> NULL
#> 
#> $GPL76$troubleDetails
#> [1] "No trouble details provided."
#> 
#> $GPL76$lastNeedsAttentionEvent
#> NULL
#> 
#> $GPL76$curationNote
#> NULL
#> 
#> $GPL76$needsAttention
#> [1] FALSE
#> 
#> $GPL76$lastNoteUpdateEvent
#> NULL
#> 
#> $GPL76$id
#> [1] 19
#> 
#> 
#> $GPL75
#> $GPL75$shortName
#> [1] "GPL75"
#> 
#> $GPL75$taxon
#> [1] "mouse"
#> 
#> $GPL75$technologyType
#> [1] "ONECOLOR"
#> 
#> $GPL75$lastGeneMapping
#> NULL
#> 
#> $GPL75$lastSequenceAnalysis
#> NULL
#> 
#> $GPL75$lastSequenceUpdate
#> NULL
#> 
#> $GPL75$numProbeSequences
#> NULL
#> 
#> $GPL75$numProbeAlignments
#> NULL
#> 
#> $GPL75$numProbesToGenes
#> NULL
#> 
#> $GPL75$numGenes
#> NULL
#> 
#> $GPL75$designElementCount
#> NULL
#> 
#> $GPL75$dateCached
#> NULL
#> 
#> $GPL75$color
#> [1] "ONECOLOR"
#> 
#> $GPL75$expressionExperimentCount
#> [1] 8
#> 
#> $GPL75$hasBlatAssociations
#> NULL
#> 
#> $GPL75$hasGeneAssociations
#> NULL
#> 
#> $GPL75$hasSequenceAssociations
#> NULL
#> 
#> $GPL75$isMerged
#> NULL
#> 
#> $GPL75$isMergee
#> [1] TRUE
#> 
#> $GPL75$isSubsumed
#> NULL
#> 
#> $GPL75$isSubsumer
#> NULL
#> 
#> $GPL75$lastRepeatMask
#> NULL
#> 
#> $GPL75$description
#> [1] " Has 13179 entries and was indexed 29-Jan-2002.  Keywords = high density oligonucleotide array\nFrom GPL75\nLast Updated: Jul 28 2006"
#> 
#> $GPL75$name
#> [1] "Affymetrix GeneChip Murine 11K SubA Array Mu11K-A"
#> 
#> $GPL75$troubled
#> [1] FALSE
#> 
#> $GPL75$lastUpdated
#> [1] 1.487572e+12
#> 
#> $GPL75$lastTroubledEvent
#> NULL
#> 
#> $GPL75$troubleDetails
#> [1] "No trouble details provided."
#> 
#> $GPL75$lastNeedsAttentionEvent
#> NULL
#> 
#> $GPL75$curationNote
#> NULL
#> 
#> $GPL75$needsAttention
#> [1] FALSE
#> 
#> $GPL75$lastNoteUpdateEvent
#> NULL
#> 
#> $GPL75$id
#> [1] 20
#> 
#> 
#> $GPL8
#> $GPL8$shortName
#> [1] "GPL8"
#> 
#> $GPL8$taxon
#> [1] "mouse"
#> 
#> $GPL8$technologyType
#> [1] "ONECOLOR"
#> 
#> $GPL8$lastGeneMapping
#> NULL
#> 
#> $GPL8$lastSequenceAnalysis
#> NULL
#> 
#> $GPL8$lastSequenceUpdate
#> NULL
#> 
#> $GPL8$numProbeSequences
#> NULL
#> 
#> $GPL8$numProbeAlignments
#> NULL
#> 
#> $GPL8$numProbesToGenes
#> NULL
#> 
#> $GPL8$numGenes
#> NULL
#> 
#> $GPL8$designElementCount
#> NULL
#> 
#> $GPL8$dateCached
#> NULL
#> 
#> $GPL8$color
#> [1] "ONECOLOR"
#> 
#> $GPL8$expressionExperimentCount
#> [1] 1
#> 
#> $GPL8$hasBlatAssociations
#> NULL
#> 
#> $GPL8$hasGeneAssociations
#> NULL
#> 
#> $GPL8$hasSequenceAssociations
#> NULL
#> 
#> $GPL8$isMerged
#> NULL
#> 
#> $GPL8$isMergee
#> [1] FALSE
#> 
#> $GPL8$isSubsumed
#> NULL
#> 
#> $GPL8$isSubsumer
#> NULL
#> 
#> $GPL8$lastRepeatMask
#> NULL
#> 
#> $GPL8$description
#> [1] "(Editor's Note: this array design is NOT the Mu11k) The Affymetrix GeneChip Mu11K was used to analyze the gene expression profile in developing mouse cerebellum (two GeneChips per E18, P7, P14, P21, and P56) to assist in the understanding of the genetic basis of cerebellar development in mice.  The analysis showed 81.6% (10,321/12,654) of the genes represented on the GeneChip were expressed in the postnatal cerebellum, and among those, 8.7% (897/10,321) were differentially expressed with more than a two-fold change in their maximum and minimum expression levels during the developmental time course.  All of these 897 differentially expressed genes were listed in this platform.  The explanation of gene function classification is as follows:  CGDD (cellular growth, differentiation, and death), CLAM (carbohydrate, lipid, and amino acid metabolism), CSC (cell structure and communication), IMTN (intra/intercellular molecular transport and neurotransmission), NNM (nucleotide and nucleic acid metabolism), ST (signal transduction), and TTPM (transcription, translation, and protein modification). Keywords = cerebellum, postnatal development, high-density oligonucleotide array\r\nFrom GPL8\r\nLast Updated: Mar 06 2006\r\n\t\t\t\t\t"
#> 
#> $GPL8$name
#> [1] "Cerebellargenexp"
#> 
#> $GPL8$troubled
#> [1] FALSE
#> 
#> $GPL8$lastUpdated
#> [1] 1.487572e+12
#> 
#> $GPL8$lastTroubledEvent
#> NULL
#> 
#> $GPL8$troubleDetails
#> [1] "No trouble details provided."
#> 
#> $GPL8$lastNeedsAttentionEvent
#> NULL
#> 
#> $GPL8$curationNote
#> NULL
#> 
#> $GPL8$needsAttention
#> [1] FALSE
#> 
#> $GPL8$lastNoteUpdateEvent
#> NULL
#> 
#> $GPL8$id
#> [1] 21
#> 
#> 

allPlatforms(c('GPL1355','GPL570'))

#> $GPL1355
#> $GPL1355$shortName
#> [1] "GPL1355"
#> 
#> $GPL1355$taxon
#> [1] "rat"
#> 
#> $GPL1355$technologyType
#> [1] "ONECOLOR"
#> 
#> $GPL1355$lastGeneMapping
#> NULL
#> 
#> $GPL1355$lastSequenceAnalysis
#> NULL
#> 
#> $GPL1355$lastSequenceUpdate
#> NULL
#> 
#> $GPL1355$numProbeSequences
#> NULL
#> 
#> $GPL1355$numProbeAlignments
#> NULL
#> 
#> $GPL1355$numProbesToGenes
#> NULL
#> 
#> $GPL1355$numGenes
#> NULL
#> 
#> $GPL1355$designElementCount
#> NULL
#> 
#> $GPL1355$dateCached
#> NULL
#> 
#> $GPL1355$color
#> [1] "ONECOLOR"
#> 
#> $GPL1355$expressionExperimentCount
#> [1] 197
#> 
#> $GPL1355$hasBlatAssociations
#> NULL
#> 
#> $GPL1355$hasGeneAssociations
#> NULL
#> 
#> $GPL1355$hasSequenceAssociations
#> NULL
#> 
#> $GPL1355$isMerged
#> NULL
#> 
#> $GPL1355$isMergee
#> [1] FALSE
#> 
#> $GPL1355$isSubsumed
#> NULL
#> 
#> $GPL1355$isSubsumer
#> NULL
#> 
#> $GPL1355$lastRepeatMask
#> NULL
#> 
#> $GPL1355$description
#> [1] " The GeneChip Rat Genome 230 2.0 Array is a powerful tool for toxicology, neurobiology, and other applications using rat as a model organism.  - Provides comprehensive coverage of the transcribed rat genome on a single array - Comprised of more than 31,000 probe sets, analyzing over 30,000 transcripts and variants from over 28,000 well-substantiated rat genes - The publicly available draft of the rat genome and leading public rat databases were used to refine sequences and provide a higher quality of data output  All probe sets represented on the GeneChip Rat Expression Set 230 are included on the GeneChip Rat Genome 230 2.0 Array. Sequences used in the design of the GeneChip Rat Genome 230 2.0 Array were selected from GenBank, dbEST, and RefSeq. The sequence clusters were created from the UniGene database (Build 99, June 2002) and then refined by analysis and comparison with the publicly available draft assembly of the rat genome from the Baylor College of Medicine Human Genome Sequencing Center (June 2002).  The GeneChip Rat Genome 230 2.0 Array includes representation of the RefSeq database sequences and probe sets related to sequences and refined EST clusters previously represented on the GeneChip Rat Genome U34 Set.  Oligonucleotide probes complementary to each corresponding sequence are synthesized in situ on the arrays. Eleven pairs of oligonucleotide probes are used to measure the level of transcription of each sequence represented on the GeneChip Rat Genome 230 2.0 Array.  Annotations derived from Affymetrix CSV file dated 6/23/2004\nFrom GPL1355\nLast Updated: May 31 2005"
#> 
#> $GPL1355$name
#> [1] "Affymetrix GeneChip Rat Genome 230 2.0 Array"
#> 
#> $GPL1355$troubled
#> [1] FALSE
#> 
#> $GPL1355$lastUpdated
#> [1] 1.488129e+12
#> 
#> $GPL1355$lastTroubledEvent
#> NULL
#> 
#> $GPL1355$troubleDetails
#> [1] "No trouble details provided."
#> 
#> $GPL1355$lastNeedsAttentionEvent
#> NULL
#> 
#> $GPL1355$curationNote
#> NULL
#> 
#> $GPL1355$needsAttention
#> [1] FALSE
#> 
#> $GPL1355$lastNoteUpdateEvent
#> NULL
#> 
#> $GPL1355$id
#> [1] 2
#> 
#> 
#> $GPL570
#> $GPL570$shortName
#> [1] "GPL570"
#> 
#> $GPL570$taxon
#> [1] "human"
#> 
#> $GPL570$technologyType
#> [1] "ONECOLOR"
#> 
#> $GPL570$lastGeneMapping
#> NULL
#> 
#> $GPL570$lastSequenceAnalysis
#> NULL
#> 
#> $GPL570$lastSequenceUpdate
#> NULL
#> 
#> $GPL570$numProbeSequences
#> NULL
#> 
#> $GPL570$numProbeAlignments
#> NULL
#> 
#> $GPL570$numProbesToGenes
#> NULL
#> 
#> $GPL570$numGenes
#> NULL
#> 
#> $GPL570$designElementCount
#> NULL
#> 
#> $GPL570$dateCached
#> NULL
#> 
#> $GPL570$color
#> [1] "ONECOLOR"
#> 
#> $GPL570$expressionExperimentCount
#> [1] 1214
#> 
#> $GPL570$hasBlatAssociations
#> NULL
#> 
#> $GPL570$hasGeneAssociations
#> NULL
#> 
#> $GPL570$hasSequenceAssociations
#> NULL
#> 
#> $GPL570$isMerged
#> NULL
#> 
#> $GPL570$isMergee
#> [1] TRUE
#> 
#> $GPL570$isSubsumed
#> NULL
#> 
#> $GPL570$isSubsumer
#> NULL
#> 
#> $GPL570$lastRepeatMask
#> NULL
#> 
#> $GPL570$description
#> [1] " Complete coverage of the Human Genome U133 Set plus 6,500 additional genes for analysis of over 47,000 transcripts All probe sets represented on the GeneChip Human Genome U133 Set are identically replicated on the GeneChip Human Genome U133 Plus 2.0 Array. The sequences from which these probe sets were derived were selected from GenBank®, dbEST, and RefSeq. The sequence clusters were created from the UniGene database (Build 133, April 20, 2001) and then refined by analysis and comparison with a number of other publicly available databases, including the Washington University EST trace repository and the University of California, Santa Cruz Golden-Path human genome database (April 2001 release). In addition, there are 9,921 new probe sets representing approximately 6,500 new genes. These gene sequences were selected from GenBank, dbEST, and RefSeq. Sequence clusters were created from the UniGene database (Build 159, January 25, 2003) and refined by analysis and comparison with a number of other publicly available databases, including the Washington University EST trace repository and the NCBI human genome assembly (Build 31). \nFrom GPL570\nLast Updated: Mar 09 2006"
#> 
#> $GPL570$name
#> [1] "Affymetrix GeneChip Human Genome U133 Plus 2.0 Array"
#> 
#> $GPL570$troubled
#> [1] FALSE
#> 
#> $GPL570$lastUpdated
#> [1] 1.487818e+12
#> 
#> $GPL570$lastTroubledEvent
#> NULL
#> 
#> $GPL570$troubleDetails
#> [1] "No trouble details provided."
#> 
#> $GPL570$lastNeedsAttentionEvent
#> NULL
#> 
#> $GPL570$curationNote
#> NULL
#> 
#> $GPL570$needsAttention
#> [1] FALSE
#> 
#> $GPL570$lastNoteUpdateEvent
#> NULL
#> 
#> $GPL570$id
#> [1] 4
#> 
#> 

# return all platforms it is slower and prone to connection interruptions
# alternative is to loop using offset and limit

# NOT RUN {
allPlatforms(limit = 0)
# }