All taxa

Lists all available taxa.

allTaxa(
  taxa = NULL,
  file = NULL,
  return = TRUE,
  overwrite = FALSE,
  memoised = FALSE
)

Arguments

taxa
Limits the result to entities with given identifiers.

A list of identifiers, separated by commas (e.g: human, mouse, fly). Identifiers can be the any of the following:

taxon ID

scientific name

common name

abbreviation

file
Character. File path. If provided, response will be saved to file

return
Logical. If the response should be returned. Set to false when you only want to save a file

overwrite
Logical. If TRUE, existing files will be overwritten. If FALSE a warning will be thrown and no action is taken.

memoised
Logical. If TRUE a memoised version of the function will be used which is faster for repeated requests. Use forgetGemmaMemoised to clear memory.

Value

List of lists containing experiment object.

Examples

allTaxa()

#> $`Homo sapiens`
#> $`Homo sapiens`$externalDatabase
#> $`Homo sapiens`$externalDatabase$checked
#> [1] FALSE
#> 
#> $`Homo sapiens`$externalDatabase$name
#> [1] "hg38"
#> 
#> $`Homo sapiens`$externalDatabase$id
#> [1] 87
#> 
#> 
#> $`Homo sapiens`$commonName
#> [1] "human"
#> 
#> $`Homo sapiens`$scientificName
#> [1] "Homo sapiens"
#> 
#> $`Homo sapiens`$isGenesUsable
#> [1] TRUE
#> 
#> $`Homo sapiens`$ncbiId
#> [1] 9606
#> 
#> $`Homo sapiens`$isSpecies
#> NULL
#> 
#> $`Homo sapiens`$`_totalInQuery`
#> [1] 0
#> 
#> $`Homo sapiens`$id
#> [1] 1
#> 
#> 
#> $`Mus musculus`
#> $`Mus musculus`$externalDatabase
#> $`Mus musculus`$externalDatabase$checked
#> [1] FALSE
#> 
#> $`Mus musculus`$externalDatabase$name
#> [1] "mm10"
#> 
#> $`Mus musculus`$externalDatabase$id
#> [1] 81
#> 
#> 
#> $`Mus musculus`$commonName
#> [1] "mouse"
#> 
#> $`Mus musculus`$scientificName
#> [1] "Mus musculus"
#> 
#> $`Mus musculus`$isGenesUsable
#> [1] TRUE
#> 
#> $`Mus musculus`$ncbiId
#> [1] 10090
#> 
#> $`Mus musculus`$isSpecies
#> NULL
#> 
#> $`Mus musculus`$`_totalInQuery`
#> [1] 0
#> 
#> $`Mus musculus`$id
#> [1] 2
#> 
#> 
#> $`Rattus norvegicus`
#> $`Rattus norvegicus`$externalDatabase
#> $`Rattus norvegicus`$externalDatabase$checked
#> [1] FALSE
#> 
#> $`Rattus norvegicus`$externalDatabase$name
#> [1] "rn6"
#> 
#> $`Rattus norvegicus`$externalDatabase$id
#> [1] 86
#> 
#> 
#> $`Rattus norvegicus`$commonName
#> [1] "rat"
#> 
#> $`Rattus norvegicus`$scientificName
#> [1] "Rattus norvegicus"
#> 
#> $`Rattus norvegicus`$isGenesUsable
#> [1] TRUE
#> 
#> $`Rattus norvegicus`$ncbiId
#> [1] 10116
#> 
#> $`Rattus norvegicus`$isSpecies
#> NULL
#> 
#> $`Rattus norvegicus`$`_totalInQuery`
#> [1] 0
#> 
#> $`Rattus norvegicus`$id
#> [1] 3
#> 
#> 
#> $`Saccharomyces cerevisiae`
#> $`Saccharomyces cerevisiae`$externalDatabase
#> NULL
#> 
#> $`Saccharomyces cerevisiae`$commonName
#> [1] "yeast"
#> 
#> $`Saccharomyces cerevisiae`$scientificName
#> [1] "Saccharomyces cerevisiae"
#> 
#> $`Saccharomyces cerevisiae`$isGenesUsable
#> [1] TRUE
#> 
#> $`Saccharomyces cerevisiae`$ncbiId
#> [1] 4932
#> 
#> $`Saccharomyces cerevisiae`$isSpecies
#> NULL
#> 
#> $`Saccharomyces cerevisiae`$`_totalInQuery`
#> [1] 0
#> 
#> $`Saccharomyces cerevisiae`$id
#> [1] 11
#> 
#> 
#> $`Danio rerio`
#> $`Danio rerio`$externalDatabase
#> NULL
#> 
#> $`Danio rerio`$commonName
#> [1] "zebrafish"
#> 
#> $`Danio rerio`$scientificName
#> [1] "Danio rerio"
#> 
#> $`Danio rerio`$isGenesUsable
#> [1] TRUE
#> 
#> $`Danio rerio`$ncbiId
#> [1] 7955
#> 
#> $`Danio rerio`$isSpecies
#> NULL
#> 
#> $`Danio rerio`$`_totalInQuery`
#> [1] 0
#> 
#> $`Danio rerio`$id
#> [1] 12
#> 
#> 
#> $`Drosophila melanogaster`
#> $`Drosophila melanogaster`$externalDatabase
#> NULL
#> 
#> $`Drosophila melanogaster`$commonName
#> [1] "fly"
#> 
#> $`Drosophila melanogaster`$scientificName
#> [1] "Drosophila melanogaster"
#> 
#> $`Drosophila melanogaster`$isGenesUsable
#> [1] TRUE
#> 
#> $`Drosophila melanogaster`$ncbiId
#> [1] 7227
#> 
#> $`Drosophila melanogaster`$isSpecies
#> NULL
#> 
#> $`Drosophila melanogaster`$`_totalInQuery`
#> [1] 0
#> 
#> $`Drosophila melanogaster`$id
#> [1] 13
#> 
#> 
#> $`Caenorhabditis elegans`
#> $`Caenorhabditis elegans`$externalDatabase
#> NULL
#> 
#> $`Caenorhabditis elegans`$commonName
#> [1] "worm"
#> 
#> $`Caenorhabditis elegans`$scientificName
#> [1] "Caenorhabditis elegans"
#> 
#> $`Caenorhabditis elegans`$isGenesUsable
#> [1] TRUE
#> 
#> $`Caenorhabditis elegans`$ncbiId
#> [1] 6239
#> 
#> $`Caenorhabditis elegans`$isSpecies
#> NULL
#> 
#> $`Caenorhabditis elegans`$`_totalInQuery`
#> [1] 0
#> 
#> $`Caenorhabditis elegans`$id
#> [1] 14
#> 
#> 
#> $`synthetic construct`
#> $`synthetic construct`$externalDatabase
#> NULL
#> 
#> $`synthetic construct`$commonName
#> [1] "synthetic construct"
#> 
#> $`synthetic construct`$scientificName
#> [1] "synthetic construct"
#> 
#> $`synthetic construct`$isGenesUsable
#> [1] FALSE
#> 
#> $`synthetic construct`$ncbiId
#> NULL
#> 
#> $`synthetic construct`$isSpecies
#> NULL
#> 
#> $`synthetic construct`$`_totalInQuery`
#> [1] 0
#> 
#> $`synthetic construct`$id
#> [1] 19
#> 
#> 
#> $`Schizosaccharomyces pombe`
#> $`Schizosaccharomyces pombe`$externalDatabase
#> NULL
#> 
#> $`Schizosaccharomyces pombe`$commonName
#> [1] "fission yeast"
#> 
#> $`Schizosaccharomyces pombe`$scientificName
#> [1] "Schizosaccharomyces pombe"
#> 
#> $`Schizosaccharomyces pombe`$isGenesUsable
#> [1] FALSE
#> 
#> $`Schizosaccharomyces pombe`$ncbiId
#> NULL
#> 
#> $`Schizosaccharomyces pombe`$isSpecies
#> NULL
#> 
#> $`Schizosaccharomyces pombe`$`_totalInQuery`
#> [1] 0
#> 
#> $`Schizosaccharomyces pombe`$id
#> [1] 33
#> 
#> 
#> $`Rattus rattus`
#> $`Rattus rattus`$externalDatabase
#> NULL
#> 
#> $`Rattus rattus`$commonName
#> NULL
#> 
#> $`Rattus rattus`$scientificName
#> [1] "Rattus rattus"
#> 
#> $`Rattus rattus`$isGenesUsable
#> [1] FALSE
#> 
#> $`Rattus rattus`$ncbiId
#> NULL
#> 
#> $`Rattus rattus`$isSpecies
#> NULL
#> 
#> $`Rattus rattus`$`_totalInQuery`
#> [1] 0
#> 
#> $`Rattus rattus`$id
#> [1] 79
#> 
#> 
#> $Rattus
#> $Rattus$externalDatabase
#> NULL
#> 
#> $Rattus$commonName
#> NULL
#> 
#> $Rattus$scientificName
#> [1] "Rattus"
#> 
#> $Rattus$isGenesUsable
#> [1] TRUE
#> 
#> $Rattus$ncbiId
#> NULL
#> 
#> $Rattus$isSpecies
#> NULL
#> 
#> $Rattus$`_totalInQuery`
#> [1] 0
#> 
#> $Rattus$id
#> [1] 80
#> 
#> 
#> $`Tetraodon nigroviridis`
#> $`Tetraodon nigroviridis`$externalDatabase
#> NULL
#> 
#> $`Tetraodon nigroviridis`$commonName
#> NULL
#> 
#> $`Tetraodon nigroviridis`$scientificName
#> [1] "Tetraodon nigroviridis"
#> 
#> $`Tetraodon nigroviridis`$isGenesUsable
#> [1] FALSE
#> 
#> $`Tetraodon nigroviridis`$ncbiId
#> NULL
#> 
#> $`Tetraodon nigroviridis`$isSpecies
#> NULL
#> 
#> $`Tetraodon nigroviridis`$`_totalInQuery`
#> [1] 0
#> 
#> $`Tetraodon nigroviridis`$id
#> [1] 81
#> 
#> 
#> $`Human gammaherpesvirus 4`
#> $`Human gammaherpesvirus 4`$externalDatabase
#> NULL
#> 
#> $`Human gammaherpesvirus 4`$commonName
#> NULL
#> 
#> $`Human gammaherpesvirus 4`$scientificName
#> [1] "Human gammaherpesvirus 4"
#> 
#> $`Human gammaherpesvirus 4`$isGenesUsable
#> [1] FALSE
#> 
#> $`Human gammaherpesvirus 4`$ncbiId
#> NULL
#> 
#> $`Human gammaherpesvirus 4`$isSpecies
#> NULL
#> 
#> $`Human gammaherpesvirus 4`$`_totalInQuery`
#> [1] 0
#> 
#> $`Human gammaherpesvirus 4`$id
#> [1] 82
#> 
#> 
#> $`Takifugu rubripes`
#> $`Takifugu rubripes`$externalDatabase
#> NULL
#> 
#> $`Takifugu rubripes`$commonName
#> NULL
#> 
#> $`Takifugu rubripes`$scientificName
#> [1] "Takifugu rubripes"
#> 
#> $`Takifugu rubripes`$isGenesUsable
#> [1] FALSE
#> 
#> $`Takifugu rubripes`$ncbiId
#> NULL
#> 
#> $`Takifugu rubripes`$isSpecies
#> NULL
#> 
#> $`Takifugu rubripes`$`_totalInQuery`
#> [1] 0
#> 
#> $`Takifugu rubripes`$id
#> [1] 83
#> 
#> 
#> $`Human gammaherpesvirus 8`
#> $`Human gammaherpesvirus 8`$externalDatabase
#> NULL
#> 
#> $`Human gammaherpesvirus 8`$commonName
#> NULL
#> 
#> $`Human gammaherpesvirus 8`$scientificName
#> [1] "Human gammaherpesvirus 8"
#> 
#> $`Human gammaherpesvirus 8`$isGenesUsable
#> [1] FALSE
#> 
#> $`Human gammaherpesvirus 8`$ncbiId
#> NULL
#> 
#> $`Human gammaherpesvirus 8`$isSpecies
#> NULL
#> 
#> $`Human gammaherpesvirus 8`$`_totalInQuery`
#> [1] 0
#> 
#> $`Human gammaherpesvirus 8`$id
#> [1] 84
#> 
#> 
#> $`Bombyx mori`
#> $`Bombyx mori`$externalDatabase
#> NULL
#> 
#> $`Bombyx mori`$commonName
#> NULL
#> 
#> $`Bombyx mori`$scientificName
#> [1] "Bombyx mori"
#> 
#> $`Bombyx mori`$isGenesUsable
#> [1] FALSE
#> 
#> $`Bombyx mori`$ncbiId
#> NULL
#> 
#> $`Bombyx mori`$isSpecies
#> NULL
#> 
#> $`Bombyx mori`$`_totalInQuery`
#> [1] 0
#> 
#> $`Bombyx mori`$id
#> [1] 85
#> 
#> 
#> $`Macaca mulatta`
#> $`Macaca mulatta`$externalDatabase
#> NULL
#> 
#> $`Macaca mulatta`$commonName
#> NULL
#> 
#> $`Macaca mulatta`$scientificName
#> [1] "Macaca mulatta"
#> 
#> $`Macaca mulatta`$isGenesUsable
#> [1] FALSE
#> 
#> $`Macaca mulatta`$ncbiId
#> NULL
#> 
#> $`Macaca mulatta`$isSpecies
#> NULL
#> 
#> $`Macaca mulatta`$`_totalInQuery`
#> [1] 0
#> 
#> $`Macaca mulatta`$id
#> [1] 86
#> 
#> 
#> $`Populus trichocarpa`
#> $`Populus trichocarpa`$externalDatabase
#> NULL
#> 
#> $`Populus trichocarpa`$commonName
#> NULL
#> 
#> $`Populus trichocarpa`$scientificName
#> [1] "Populus trichocarpa"
#> 
#> $`Populus trichocarpa`$isGenesUsable
#> [1] FALSE
#> 
#> $`Populus trichocarpa`$ncbiId
#> NULL
#> 
#> $`Populus trichocarpa`$isSpecies
#> NULL
#> 
#> $`Populus trichocarpa`$`_totalInQuery`
#> [1] 0
#> 
#> $`Populus trichocarpa`$id
#> [1] 87
#> 
#> 
#> $`Gorilla gorilla`
#> $`Gorilla gorilla`$externalDatabase
#> NULL
#> 
#> $`Gorilla gorilla`$commonName
#> NULL
#> 
#> $`Gorilla gorilla`$scientificName
#> [1] "Gorilla gorilla"
#> 
#> $`Gorilla gorilla`$isGenesUsable
#> [1] FALSE
#> 
#> $`Gorilla gorilla`$ncbiId
#> NULL
#> 
#> $`Gorilla gorilla`$isSpecies
#> NULL
#> 
#> $`Gorilla gorilla`$`_totalInQuery`
#> [1] 0
#> 
#> $`Gorilla gorilla`$id
#> [1] 88
#> 
#> 
#> $`Arabidopsis thaliana`
#> $`Arabidopsis thaliana`$externalDatabase
#> NULL
#> 
#> $`Arabidopsis thaliana`$commonName
#> NULL
#> 
#> $`Arabidopsis thaliana`$scientificName
#> [1] "Arabidopsis thaliana"
#> 
#> $`Arabidopsis thaliana`$isGenesUsable
#> [1] FALSE
#> 
#> $`Arabidopsis thaliana`$ncbiId
#> NULL
#> 
#> $`Arabidopsis thaliana`$isSpecies
#> NULL
#> 
#> $`Arabidopsis thaliana`$`_totalInQuery`
#> [1] 0
#> 
#> $`Arabidopsis thaliana`$id
#> [1] 89
#> 
#> 
#> $`Apis mellifera`
#> $`Apis mellifera`$externalDatabase
#> NULL
#> 
#> $`Apis mellifera`$commonName
#> NULL
#> 
#> $`Apis mellifera`$scientificName
#> [1] "Apis mellifera"
#> 
#> $`Apis mellifera`$isGenesUsable
#> [1] FALSE
#> 
#> $`Apis mellifera`$ncbiId
#> NULL
#> 
#> $`Apis mellifera`$isSpecies
#> NULL
#> 
#> $`Apis mellifera`$`_totalInQuery`
#> [1] 0
#> 
#> $`Apis mellifera`$id
#> [1] 90
#> 
#> 
#> $`Oryza sativa`
#> $`Oryza sativa`$externalDatabase
#> NULL
#> 
#> $`Oryza sativa`$commonName
#> NULL
#> 
#> $`Oryza sativa`$scientificName
#> [1] "Oryza sativa"
#> 
#> $`Oryza sativa`$isGenesUsable
#> [1] FALSE
#> 
#> $`Oryza sativa`$ncbiId
#> NULL
#> 
#> $`Oryza sativa`$isSpecies
#> NULL
#> 
#> $`Oryza sativa`$`_totalInQuery`
#> [1] 0
#> 
#> $`Oryza sativa`$id
#> [1] 91
#> 
#> 
#> $`Sorghum bicolor`
#> $`Sorghum bicolor`$externalDatabase
#> NULL
#> 
#> $`Sorghum bicolor`$commonName
#> NULL
#> 
#> $`Sorghum bicolor`$scientificName
#> [1] "Sorghum bicolor"
#> 
#> $`Sorghum bicolor`$isGenesUsable
#> [1] FALSE
#> 
#> $`Sorghum bicolor`$ncbiId
#> NULL
#> 
#> $`Sorghum bicolor`$isSpecies
#> NULL
#> 
#> $`Sorghum bicolor`$`_totalInQuery`
#> [1] 0
#> 
#> $`Sorghum bicolor`$id
#> [1] 92
#> 
#> 
#> $`Macacine gammaherpesvirus 4`
#> $`Macacine gammaherpesvirus 4`$externalDatabase
#> NULL
#> 
#> $`Macacine gammaherpesvirus 4`$commonName
#> NULL
#> 
#> $`Macacine gammaherpesvirus 4`$scientificName
#> [1] "Macacine gammaherpesvirus 4"
#> 
#> $`Macacine gammaherpesvirus 4`$isGenesUsable
#> [1] FALSE
#> 
#> $`Macacine gammaherpesvirus 4`$ncbiId
#> NULL
#> 
#> $`Macacine gammaherpesvirus 4`$isSpecies
#> NULL
#> 
#> $`Macacine gammaherpesvirus 4`$`_totalInQuery`
#> [1] 0
#> 
#> $`Macacine gammaherpesvirus 4`$id
#> [1] 93
#> 
#> 
#> $`Anopheles gambiae`
#> $`Anopheles gambiae`$externalDatabase
#> NULL
#> 
#> $`Anopheles gambiae`$commonName
#> NULL
#> 
#> $`Anopheles gambiae`$scientificName
#> [1] "Anopheles gambiae"
#> 
#> $`Anopheles gambiae`$isGenesUsable
#> [1] FALSE
#> 
#> $`Anopheles gambiae`$ncbiId
#> NULL
#> 
#> $`Anopheles gambiae`$isSpecies
#> NULL
#> 
#> $`Anopheles gambiae`$`_totalInQuery`
#> [1] 0
#> 
#> $`Anopheles gambiae`$id
#> [1] 94
#> 
#> 
#> $`Canis lupus familiaris`
#> $`Canis lupus familiaris`$externalDatabase
#> NULL
#> 
#> $`Canis lupus familiaris`$commonName
#> NULL
#> 
#> $`Canis lupus familiaris`$scientificName
#> [1] "Canis lupus familiaris"
#> 
#> $`Canis lupus familiaris`$isGenesUsable
#> [1] FALSE
#> 
#> $`Canis lupus familiaris`$ncbiId
#> NULL
#> 
#> $`Canis lupus familiaris`$isSpecies
#> NULL
#> 
#> $`Canis lupus familiaris`$`_totalInQuery`
#> [1] 0
#> 
#> $`Canis lupus familiaris`$id
#> [1] 95
#> 
#> 
#> $`Lemur catta`
#> $`Lemur catta`$externalDatabase
#> NULL
#> 
#> $`Lemur catta`$commonName
#> NULL
#> 
#> $`Lemur catta`$scientificName
#> [1] "Lemur catta"
#> 
#> $`Lemur catta`$isGenesUsable
#> [1] FALSE
#> 
#> $`Lemur catta`$ncbiId
#> NULL
#> 
#> $`Lemur catta`$isSpecies
#> NULL
#> 
#> $`Lemur catta`$`_totalInQuery`
#> [1] 0
#> 
#> $`Lemur catta`$id
#> [1] 96
#> 
#> 
#> $`Xenopus tropicalis`
#> $`Xenopus tropicalis`$externalDatabase
#> NULL
#> 
#> $`Xenopus tropicalis`$commonName
#> NULL
#> 
#> $`Xenopus tropicalis`$scientificName
#> [1] "Xenopus tropicalis"
#> 
#> $`Xenopus tropicalis`$isGenesUsable
#> [1] FALSE
#> 
#> $`Xenopus tropicalis`$ncbiId
#> NULL
#> 
#> $`Xenopus tropicalis`$isSpecies
#> NULL
#> 
#> $`Xenopus tropicalis`$`_totalInQuery`
#> [1] 0
#> 
#> $`Xenopus tropicalis`$id
#> [1] 97
#> 
#> 
#> $`Monodelphis domestica`
#> $`Monodelphis domestica`$externalDatabase
#> NULL
#> 
#> $`Monodelphis domestica`$commonName
#> NULL
#> 
#> $`Monodelphis domestica`$scientificName
#> [1] "Monodelphis domestica"
#> 
#> $`Monodelphis domestica`$isGenesUsable
#> [1] FALSE
#> 
#> $`Monodelphis domestica`$ncbiId
#> NULL
#> 
#> $`Monodelphis domestica`$isSpecies
#> NULL
#> 
#> $`Monodelphis domestica`$`_totalInQuery`
#> [1] 0
#> 
#> $`Monodelphis domestica`$id
#> [1] 98
#> 
#> 
#> $`Gallus gallus`
#> $`Gallus gallus`$externalDatabase
#> NULL
#> 
#> $`Gallus gallus`$commonName
#> NULL
#> 
#> $`Gallus gallus`$scientificName
#> [1] "Gallus gallus"
#> 
#> $`Gallus gallus`$isGenesUsable
#> [1] FALSE
#> 
#> $`Gallus gallus`$ncbiId
#> NULL
#> 
#> $`Gallus gallus`$isSpecies
#> NULL
#> 
#> $`Gallus gallus`$`_totalInQuery`
#> [1] 0
#> 
#> $`Gallus gallus`$id
#> [1] 99
#> 
#> 
#> $`Medicago truncatula`
#> $`Medicago truncatula`$externalDatabase
#> NULL
#> 
#> $`Medicago truncatula`$commonName
#> NULL
#> 
#> $`Medicago truncatula`$scientificName
#> [1] "Medicago truncatula"
#> 
#> $`Medicago truncatula`$isGenesUsable
#> [1] FALSE
#> 
#> $`Medicago truncatula`$ncbiId
#> NULL
#> 
#> $`Medicago truncatula`$isSpecies
#> NULL
#> 
#> $`Medicago truncatula`$`_totalInQuery`
#> [1] 0
#> 
#> $`Medicago truncatula`$id
#> [1] 100
#> 
#> 
#> $`Macaca nemestrina`
#> $`Macaca nemestrina`$externalDatabase
#> NULL
#> 
#> $`Macaca nemestrina`$commonName
#> NULL
#> 
#> $`Macaca nemestrina`$scientificName
#> [1] "Macaca nemestrina"
#> 
#> $`Macaca nemestrina`$isGenesUsable
#> [1] FALSE
#> 
#> $`Macaca nemestrina`$ncbiId
#> NULL
#> 
#> $`Macaca nemestrina`$isSpecies
#> NULL
#> 
#> $`Macaca nemestrina`$`_totalInQuery`
#> [1] 0
#> 
#> $`Macaca nemestrina`$id
#> [1] 101
#> 
#> 
#> $`Saguinus labiatus`
#> $`Saguinus labiatus`$externalDatabase
#> NULL
#> 
#> $`Saguinus labiatus`$commonName
#> NULL
#> 
#> $`Saguinus labiatus`$scientificName
#> [1] "Saguinus labiatus"
#> 
#> $`Saguinus labiatus`$isGenesUsable
#> [1] FALSE
#> 
#> $`Saguinus labiatus`$ncbiId
#> NULL
#> 
#> $`Saguinus labiatus`$isSpecies
#> NULL
#> 
#> $`Saguinus labiatus`$`_totalInQuery`
#> [1] 0
#> 
#> $`Saguinus labiatus`$id
#> [1] 102
#> 
#> 
#> $`Zea mays`
#> $`Zea mays`$externalDatabase
#> NULL
#> 
#> $`Zea mays`$commonName
#> NULL
#> 
#> $`Zea mays`$scientificName
#> [1] "Zea mays"
#> 
#> $`Zea mays`$isGenesUsable
#> [1] FALSE
#> 
#> $`Zea mays`$ncbiId
#> NULL
#> 
#> $`Zea mays`$isSpecies
#> NULL
#> 
#> $`Zea mays`$`_totalInQuery`
#> [1] 0
#> 
#> $`Zea mays`$id
#> [1] 103
#> 
#> 
#> $`Caenorhabditis briggsae`
#> $`Caenorhabditis briggsae`$externalDatabase
#> NULL
#> 
#> $`Caenorhabditis briggsae`$commonName
#> NULL
#> 
#> $`Caenorhabditis briggsae`$scientificName
#> [1] "Caenorhabditis briggsae"
#> 
#> $`Caenorhabditis briggsae`$isGenesUsable
#> [1] FALSE
#> 
#> $`Caenorhabditis briggsae`$ncbiId
#> NULL
#> 
#> $`Caenorhabditis briggsae`$isSpecies
#> NULL
#> 
#> $`Caenorhabditis briggsae`$`_totalInQuery`
#> [1] 0
#> 
#> $`Caenorhabditis briggsae`$id
#> [1] 104
#> 
#> 
#> $`Xenopus laevis`
#> $`Xenopus laevis`$externalDatabase
#> NULL
#> 
#> $`Xenopus laevis`$commonName
#> NULL
#> 
#> $`Xenopus laevis`$scientificName
#> [1] "Xenopus laevis"
#> 
#> $`Xenopus laevis`$isGenesUsable
#> [1] FALSE
#> 
#> $`Xenopus laevis`$ncbiId
#> NULL
#> 
#> $`Xenopus laevis`$isSpecies
#> NULL
#> 
#> $`Xenopus laevis`$`_totalInQuery`
#> [1] 0
#> 
#> $`Xenopus laevis`$id
#> [1] 105
#> 
#> 
#> $`Pan paniscus`
#> $`Pan paniscus`$externalDatabase
#> NULL
#> 
#> $`Pan paniscus`$commonName
#> NULL
#> 
#> $`Pan paniscus`$scientificName
#> [1] "Pan paniscus"
#> 
#> $`Pan paniscus`$isGenesUsable
#> [1] FALSE
#> 
#> $`Pan paniscus`$ncbiId
#> NULL
#> 
#> $`Pan paniscus`$isSpecies
#> NULL
#> 
#> $`Pan paniscus`$`_totalInQuery`
#> [1] 0
#> 
#> $`Pan paniscus`$id
#> [1] 106
#> 
#> 
#> $`Pan troglodytes`
#> $`Pan troglodytes`$externalDatabase
#> NULL
#> 
#> $`Pan troglodytes`$commonName
#> NULL
#> 
#> $`Pan troglodytes`$scientificName
#> [1] "Pan troglodytes"
#> 
#> $`Pan troglodytes`$isGenesUsable
#> [1] FALSE
#> 
#> $`Pan troglodytes`$ncbiId
#> NULL
#> 
#> $`Pan troglodytes`$isSpecies
#> NULL
#> 
#> $`Pan troglodytes`$`_totalInQuery`
#> [1] 0
#> 
#> $`Pan troglodytes`$id
#> [1] 107
#> 
#> 
#> $`Bos taurus`
#> $`Bos taurus`$externalDatabase
#> NULL
#> 
#> $`Bos taurus`$commonName
#> NULL
#> 
#> $`Bos taurus`$scientificName
#> [1] "Bos taurus"
#> 
#> $`Bos taurus`$isGenesUsable
#> [1] FALSE
#> 
#> $`Bos taurus`$ncbiId
#> NULL
#> 
#> $`Bos taurus`$isSpecies
#> NULL
#> 
#> $`Bos taurus`$`_totalInQuery`
#> [1] 0
#> 
#> $`Bos taurus`$id
#> [1] 108
#> 
#> 
#> $`Glycine max`
#> $`Glycine max`$externalDatabase
#> NULL
#> 
#> $`Glycine max`$commonName
#> NULL
#> 
#> $`Glycine max`$scientificName
#> [1] "Glycine max"
#> 
#> $`Glycine max`$isGenesUsable
#> [1] FALSE
#> 
#> $`Glycine max`$ncbiId
#> NULL
#> 
#> $`Glycine max`$isSpecies
#> NULL
#> 
#> $`Glycine max`$`_totalInQuery`
#> [1] 0
#> 
#> $`Glycine max`$id
#> [1] 109
#> 
#> 
#> $`Gallid alphaherpesvirus 2`
#> $`Gallid alphaherpesvirus 2`$externalDatabase
#> NULL
#> 
#> $`Gallid alphaherpesvirus 2`$commonName
#> NULL
#> 
#> $`Gallid alphaherpesvirus 2`$scientificName
#> [1] "Gallid alphaherpesvirus 2"
#> 
#> $`Gallid alphaherpesvirus 2`$isGenesUsable
#> [1] FALSE
#> 
#> $`Gallid alphaherpesvirus 2`$ncbiId
#> NULL
#> 
#> $`Gallid alphaherpesvirus 2`$isSpecies
#> NULL
#> 
#> $`Gallid alphaherpesvirus 2`$`_totalInQuery`
#> [1] 0
#> 
#> $`Gallid alphaherpesvirus 2`$id
#> [1] 110
#> 
#> 
#> $`Macaca mulatta polyomavirus 1`
#> $`Macaca mulatta polyomavirus 1`$externalDatabase
#> NULL
#> 
#> $`Macaca mulatta polyomavirus 1`$commonName
#> NULL
#> 
#> $`Macaca mulatta polyomavirus 1`$scientificName
#> [1] "Macaca mulatta polyomavirus 1"
#> 
#> $`Macaca mulatta polyomavirus 1`$isGenesUsable
#> [1] FALSE
#> 
#> $`Macaca mulatta polyomavirus 1`$ncbiId
#> NULL
#> 
#> $`Macaca mulatta polyomavirus 1`$isSpecies
#> NULL
#> 
#> $`Macaca mulatta polyomavirus 1`$`_totalInQuery`
#> [1] 0
#> 
#> $`Macaca mulatta polyomavirus 1`$id
#> [1] 111
#> 
#> 
#> $`Lagothrix lagotricha`
#> $`Lagothrix lagotricha`$externalDatabase
#> NULL
#> 
#> $`Lagothrix lagotricha`$commonName
#> NULL
#> 
#> $`Lagothrix lagotricha`$scientificName
#> [1] "Lagothrix lagotricha"
#> 
#> $`Lagothrix lagotricha`$isGenesUsable
#> [1] FALSE
#> 
#> $`Lagothrix lagotricha`$ncbiId
#> NULL
#> 
#> $`Lagothrix lagotricha`$isSpecies
#> NULL
#> 
#> $`Lagothrix lagotricha`$`_totalInQuery`
#> [1] 0
#> 
#> $`Lagothrix lagotricha`$id
#> [1] 112
#> 
#> 
#> $`Ovis aries`
#> $`Ovis aries`$externalDatabase
#> NULL
#> 
#> $`Ovis aries`$commonName
#> NULL
#> 
#> $`Ovis aries`$scientificName
#> [1] "Ovis aries"
#> 
#> $`Ovis aries`$isGenesUsable
#> [1] FALSE
#> 
#> $`Ovis aries`$ncbiId
#> NULL
#> 
#> $`Ovis aries`$isSpecies
#> NULL
#> 
#> $`Ovis aries`$`_totalInQuery`
#> [1] 0
#> 
#> $`Ovis aries`$id
#> [1] 113
#> 
#> 
#> $`Saccharum officinarum`
#> $`Saccharum officinarum`$externalDatabase
#> NULL
#> 
#> $`Saccharum officinarum`$commonName
#> NULL
#> 
#> $`Saccharum officinarum`$scientificName
#> [1] "Saccharum officinarum"
#> 
#> $`Saccharum officinarum`$isGenesUsable
#> [1] FALSE
#> 
#> $`Saccharum officinarum`$ncbiId
#> NULL
#> 
#> $`Saccharum officinarum`$isSpecies
#> NULL
#> 
#> $`Saccharum officinarum`$`_totalInQuery`
#> [1] 0
#> 
#> $`Saccharum officinarum`$id
#> [1] 114
#> 
#> 
#> $`Physcomitrella patens`
#> $`Physcomitrella patens`$externalDatabase
#> NULL
#> 
#> $`Physcomitrella patens`$commonName
#> NULL
#> 
#> $`Physcomitrella patens`$scientificName
#> [1] "Physcomitrella patens"
#> 
#> $`Physcomitrella patens`$isGenesUsable
#> [1] FALSE
#> 
#> $`Physcomitrella patens`$ncbiId
#> NULL
#> 
#> $`Physcomitrella patens`$isSpecies
#> NULL
#> 
#> $`Physcomitrella patens`$`_totalInQuery`
#> [1] 0
#> 
#> $`Physcomitrella patens`$id
#> [1] 115
#> 
#> 
#> $`Drosophila pseudoobscura`
#> $`Drosophila pseudoobscura`$externalDatabase
#> NULL
#> 
#> $`Drosophila pseudoobscura`$commonName
#> NULL
#> 
#> $`Drosophila pseudoobscura`$scientificName
#> [1] "Drosophila pseudoobscura"
#> 
#> $`Drosophila pseudoobscura`$isGenesUsable
#> [1] FALSE
#> 
#> $`Drosophila pseudoobscura`$ncbiId
#> NULL
#> 
#> $`Drosophila pseudoobscura`$isSpecies
#> NULL
#> 
#> $`Drosophila pseudoobscura`$`_totalInQuery`
#> [1] 0
#> 
#> $`Drosophila pseudoobscura`$id
#> [1] 116
#> 
#> 
#> $`Ateles geoffroyi`
#> $`Ateles geoffroyi`$externalDatabase
#> NULL
#> 
#> $`Ateles geoffroyi`$commonName
#> NULL
#> 
#> $`Ateles geoffroyi`$scientificName
#> [1] "Ateles geoffroyi"
#> 
#> $`Ateles geoffroyi`$isGenesUsable
#> [1] FALSE
#> 
#> $`Ateles geoffroyi`$ncbiId
#> NULL
#> 
#> $`Ateles geoffroyi`$isSpecies
#> NULL
#> 
#> $`Ateles geoffroyi`$`_totalInQuery`
#> [1] 0
#> 
#> $`Ateles geoffroyi`$id
#> [1] 117
#> 
#> 
#> $`Murid gammaherpesvirus 4`
#> $`Murid gammaherpesvirus 4`$externalDatabase
#> NULL
#> 
#> $`Murid gammaherpesvirus 4`$commonName
#> NULL
#> 
#> $`Murid gammaherpesvirus 4`$scientificName
#> [1] "Murid gammaherpesvirus 4"
#> 
#> $`Murid gammaherpesvirus 4`$isGenesUsable
#> [1] FALSE
#> 
#> $`Murid gammaherpesvirus 4`$ncbiId
#> NULL
#> 
#> $`Murid gammaherpesvirus 4`$isSpecies
#> NULL
#> 
#> $`Murid gammaherpesvirus 4`$`_totalInQuery`
#> [1] 0
#> 
#> $`Murid gammaherpesvirus 4`$id
#> [1] 118
#> 
#> 
#> $`Schmidtea mediterranea`
#> $`Schmidtea mediterranea`$externalDatabase
#> NULL
#> 
#> $`Schmidtea mediterranea`$commonName
#> NULL
#> 
#> $`Schmidtea mediterranea`$scientificName
#> [1] "Schmidtea mediterranea"
#> 
#> $`Schmidtea mediterranea`$isGenesUsable
#> [1] FALSE
#> 
#> $`Schmidtea mediterranea`$ncbiId
#> NULL
#> 
#> $`Schmidtea mediterranea`$isSpecies
#> NULL
#> 
#> $`Schmidtea mediterranea`$`_totalInQuery`
#> [1] 0
#> 
#> $`Schmidtea mediterranea`$id
#> [1] 119
#> 
#> 
#> $`Sus scrofa`
#> $`Sus scrofa`$externalDatabase
#> NULL
#> 
#> $`Sus scrofa`$commonName
#> NULL
#> 
#> $`Sus scrofa`$scientificName
#> [1] "Sus scrofa"
#> 
#> $`Sus scrofa`$isGenesUsable
#> [1] FALSE
#> 
#> $`Sus scrofa`$ncbiId
#> NULL
#> 
#> $`Sus scrofa`$isSpecies
#> NULL
#> 
#> $`Sus scrofa`$`_totalInQuery`
#> [1] 0
#> 
#> $`Sus scrofa`$id
#> [1] 120
#> 
#> 
#> $`Pongo pygmaeus`
#> $`Pongo pygmaeus`$externalDatabase
#> NULL
#> 
#> $`Pongo pygmaeus`$commonName
#> NULL
#> 
#> $`Pongo pygmaeus`$scientificName
#> [1] "Pongo pygmaeus"
#> 
#> $`Pongo pygmaeus`$isGenesUsable
#> [1] FALSE
#> 
#> $`Pongo pygmaeus`$ncbiId
#> NULL
#> 
#> $`Pongo pygmaeus`$isSpecies
#> NULL
#> 
#> $`Pongo pygmaeus`$`_totalInQuery`
#> [1] 0
#> 
#> $`Pongo pygmaeus`$id
#> [1] 121
#> 
#> 
#> $`Human betaherpesvirus 5`
#> $`Human betaherpesvirus 5`$externalDatabase
#> NULL
#> 
#> $`Human betaherpesvirus 5`$commonName
#> NULL
#> 
#> $`Human betaherpesvirus 5`$scientificName
#> [1] "Human betaherpesvirus 5"
#> 
#> $`Human betaherpesvirus 5`$isGenesUsable
#> [1] FALSE
#> 
#> $`Human betaherpesvirus 5`$ncbiId
#> NULL
#> 
#> $`Human betaherpesvirus 5`$isSpecies
#> NULL
#> 
#> $`Human betaherpesvirus 5`$`_totalInQuery`
#> [1] 0
#> 
#> $`Human betaherpesvirus 5`$id
#> [1] 122
#> 
#> 
#> $`Ornithorhynchus anatinus`
#> $`Ornithorhynchus anatinus`$externalDatabase
#> NULL
#> 
#> $`Ornithorhynchus anatinus`$commonName
#> NULL
#> 
#> $`Ornithorhynchus anatinus`$scientificName
#> [1] "Ornithorhynchus anatinus"
#> 
#> $`Ornithorhynchus anatinus`$isGenesUsable
#> [1] FALSE
#> 
#> $`Ornithorhynchus anatinus`$ncbiId
#> NULL
#> 
#> $`Ornithorhynchus anatinus`$isSpecies
#> NULL
#> 
#> $`Ornithorhynchus anatinus`$`_totalInQuery`
#> [1] 0
#> 
#> $`Ornithorhynchus anatinus`$id
#> [1] 123
#> 
#> 
#> $`Plasmodium berghei`
#> $`Plasmodium berghei`$externalDatabase
#> NULL
#> 
#> $`Plasmodium berghei`$commonName
#> NULL
#> 
#> $`Plasmodium berghei`$scientificName
#> [1] "Plasmodium berghei"
#> 
#> $`Plasmodium berghei`$isGenesUsable
#> [1] FALSE
#> 
#> $`Plasmodium berghei`$ncbiId
#> NULL
#> 
#> $`Plasmodium berghei`$isSpecies
#> NULL
#> 
#> $`Plasmodium berghei`$`_totalInQuery`
#> [1] 0
#> 
#> $`Plasmodium berghei`$id
#> [1] 124
#> 
#> 
#> $`Streptococcus pyogenes`
#> $`Streptococcus pyogenes`$externalDatabase
#> NULL
#> 
#> $`Streptococcus pyogenes`$commonName
#> NULL
#> 
#> $`Streptococcus pyogenes`$scientificName
#> [1] "Streptococcus pyogenes"
#> 
#> $`Streptococcus pyogenes`$isGenesUsable
#> [1] FALSE
#> 
#> $`Streptococcus pyogenes`$ncbiId
#> NULL
#> 
#> $`Streptococcus pyogenes`$isSpecies
#> NULL
#> 
#> $`Streptococcus pyogenes`$`_totalInQuery`
#> [1] 0
#> 
#> $`Streptococcus pyogenes`$id
#> [1] 125
#> 
#> 
#> $`Chlamydia trachomatis`
#> $`Chlamydia trachomatis`$externalDatabase
#> NULL
#> 
#> $`Chlamydia trachomatis`$commonName
#> NULL
#> 
#> $`Chlamydia trachomatis`$scientificName
#> [1] "Chlamydia trachomatis"
#> 
#> $`Chlamydia trachomatis`$isGenesUsable
#> [1] FALSE
#> 
#> $`Chlamydia trachomatis`$ncbiId
#> NULL
#> 
#> $`Chlamydia trachomatis`$isSpecies
#> NULL
#> 
#> $`Chlamydia trachomatis`$`_totalInQuery`
#> [1] 0
#> 
#> $`Chlamydia trachomatis`$id
#> [1] 126
#> 
#> 
#> $`Borreliella burgdorferi`
#> $`Borreliella burgdorferi`$externalDatabase
#> NULL
#> 
#> $`Borreliella burgdorferi`$commonName
#> NULL
#> 
#> $`Borreliella burgdorferi`$scientificName
#> [1] "Borreliella burgdorferi"
#> 
#> $`Borreliella burgdorferi`$isGenesUsable
#> [1] FALSE
#> 
#> $`Borreliella burgdorferi`$ncbiId
#> NULL
#> 
#> $`Borreliella burgdorferi`$isSpecies
#> NULL
#> 
#> $`Borreliella burgdorferi`$`_totalInQuery`
#> [1] 0
#> 
#> $`Borreliella burgdorferi`$id
#> [1] 127
#> 
#> 
#> $`Coxiella burnetii`
#> $`Coxiella burnetii`$externalDatabase
#> NULL
#> 
#> $`Coxiella burnetii`$commonName
#> NULL
#> 
#> $`Coxiella burnetii`$scientificName
#> [1] "Coxiella burnetii"
#> 
#> $`Coxiella burnetii`$isGenesUsable
#> [1] FALSE
#> 
#> $`Coxiella burnetii`$ncbiId
#> NULL
#> 
#> $`Coxiella burnetii`$isSpecies
#> NULL
#> 
#> $`Coxiella burnetii`$`_totalInQuery`
#> [1] 0
#> 
#> $`Coxiella burnetii`$id
#> [1] 128
#> 
#> 
#> $`Staphylococcus aureus`
#> $`Staphylococcus aureus`$externalDatabase
#> NULL
#> 
#> $`Staphylococcus aureus`$commonName
#> NULL
#> 
#> $`Staphylococcus aureus`$scientificName
#> [1] "Staphylococcus aureus"
#> 
#> $`Staphylococcus aureus`$isGenesUsable
#> [1] FALSE
#> 
#> $`Staphylococcus aureus`$ncbiId
#> NULL
#> 
#> $`Staphylococcus aureus`$isSpecies
#> NULL
#> 
#> $`Staphylococcus aureus`$`_totalInQuery`
#> [1] 0
#> 
#> $`Staphylococcus aureus`$id
#> [1] 129
#> 
#> 
#> $`Yersinia pestis`
#> $`Yersinia pestis`$externalDatabase
#> NULL
#> 
#> $`Yersinia pestis`$commonName
#> NULL
#> 
#> $`Yersinia pestis`$scientificName
#> [1] "Yersinia pestis"
#> 
#> $`Yersinia pestis`$isGenesUsable
#> [1] FALSE
#> 
#> $`Yersinia pestis`$ncbiId
#> NULL
#> 
#> $`Yersinia pestis`$isSpecies
#> NULL
#> 
#> $`Yersinia pestis`$`_totalInQuery`
#> [1] 0
#> 
#> $`Yersinia pestis`$id
#> [1] 130
#> 
#> 
#> $`Nannospalax galili`
#> $`Nannospalax galili`$externalDatabase
#> NULL
#> 
#> $`Nannospalax galili`$commonName
#> NULL
#> 
#> $`Nannospalax galili`$scientificName
#> [1] "Nannospalax galili"
#> 
#> $`Nannospalax galili`$isGenesUsable
#> [1] FALSE
#> 
#> $`Nannospalax galili`$ncbiId
#> NULL
#> 
#> $`Nannospalax galili`$isSpecies
#> NULL
#> 
#> $`Nannospalax galili`$`_totalInQuery`
#> [1] 0
#> 
#> $`Nannospalax galili`$id
#> [1] 131
#> 
#> 
#> $`Crocosphaera subtropica ATCC 51142`
#> $`Crocosphaera subtropica ATCC 51142`$externalDatabase
#> NULL
#> 
#> $`Crocosphaera subtropica ATCC 51142`$commonName
#> NULL
#> 
#> $`Crocosphaera subtropica ATCC 51142`$scientificName
#> [1] "Crocosphaera subtropica ATCC 51142"
#> 
#> $`Crocosphaera subtropica ATCC 51142`$isGenesUsable
#> [1] FALSE
#> 
#> $`Crocosphaera subtropica ATCC 51142`$ncbiId
#> NULL
#> 
#> $`Crocosphaera subtropica ATCC 51142`$isSpecies
#> NULL
#> 
#> $`Crocosphaera subtropica ATCC 51142`$`_totalInQuery`
#> [1] 0
#> 
#> $`Crocosphaera subtropica ATCC 51142`$id
#> [1] 132
#> 
#> 
#> $`Salmonella enterica subsp. enterica serovar Typhimurium str. LT2`
#> $`Salmonella enterica subsp. enterica serovar Typhimurium str. LT2`$externalDatabase
#> NULL
#> 
#> $`Salmonella enterica subsp. enterica serovar Typhimurium str. LT2`$commonName
#> NULL
#> 
#> $`Salmonella enterica subsp. enterica serovar Typhimurium str. LT2`$scientificName
#> [1] "Salmonella enterica subsp. enterica serovar Typhimurium str. LT2"
#> 
#> $`Salmonella enterica subsp. enterica serovar Typhimurium str. LT2`$isGenesUsable
#> [1] FALSE
#> 
#> $`Salmonella enterica subsp. enterica serovar Typhimurium str. LT2`$ncbiId
#> NULL
#> 
#> $`Salmonella enterica subsp. enterica serovar Typhimurium str. LT2`$isSpecies
#> NULL
#> 
#> $`Salmonella enterica subsp. enterica serovar Typhimurium str. LT2`$`_totalInQuery`
#> [1] 0
#> 
#> $`Salmonella enterica subsp. enterica serovar Typhimurium str. LT2`$id
#> [1] 133
#> 
#> 
#> $`Salmonella enterica subsp. enterica serovar Typhimurium str. UK-1`
#> $`Salmonella enterica subsp. enterica serovar Typhimurium str. UK-1`$externalDatabase
#> NULL
#> 
#> $`Salmonella enterica subsp. enterica serovar Typhimurium str. UK-1`$commonName
#> NULL
#> 
#> $`Salmonella enterica subsp. enterica serovar Typhimurium str. UK-1`$scientificName
#> [1] "Salmonella enterica subsp. enterica serovar Typhimurium str. UK-1"
#> 
#> $`Salmonella enterica subsp. enterica serovar Typhimurium str. UK-1`$isGenesUsable
#> [1] TRUE
#> 
#> $`Salmonella enterica subsp. enterica serovar Typhimurium str. UK-1`$ncbiId
#> NULL
#> 
#> $`Salmonella enterica subsp. enterica serovar Typhimurium str. UK-1`$isSpecies
#> NULL
#> 
#> $`Salmonella enterica subsp. enterica serovar Typhimurium str. UK-1`$`_totalInQuery`
#> [1] 0
#> 
#> $`Salmonella enterica subsp. enterica serovar Typhimurium str. UK-1`$id
#> [1] 134
#> 
#> 
#> $`JC polyomavirus`
#> $`JC polyomavirus`$externalDatabase
#> NULL
#> 
#> $`JC polyomavirus`$commonName
#> NULL
#> 
#> $`JC polyomavirus`$scientificName
#> [1] "JC polyomavirus"
#> 
#> $`JC polyomavirus`$isGenesUsable
#> [1] FALSE
#> 
#> $`JC polyomavirus`$ncbiId
#> NULL
#> 
#> $`JC polyomavirus`$isSpecies
#> NULL
#> 
#> $`JC polyomavirus`$`_totalInQuery`
#> [1] 0
#> 
#> $`JC polyomavirus`$id
#> [1] 135
#> 
#> 
#> $`Human alphaherpesvirus 1`
#> $`Human alphaherpesvirus 1`$externalDatabase
#> NULL
#> 
#> $`Human alphaherpesvirus 1`$commonName
#> NULL
#> 
#> $`Human alphaherpesvirus 1`$scientificName
#> [1] "Human alphaherpesvirus 1"
#> 
#> $`Human alphaherpesvirus 1`$isGenesUsable
#> [1] FALSE
#> 
#> $`Human alphaherpesvirus 1`$ncbiId
#> NULL
#> 
#> $`Human alphaherpesvirus 1`$isSpecies
#> NULL
#> 
#> $`Human alphaherpesvirus 1`$`_totalInQuery`
#> [1] 0
#> 
#> $`Human alphaherpesvirus 1`$id
#> [1] 136
#> 
#> 
#> $`Murid betaherpesvirus 1`
#> $`Murid betaherpesvirus 1`$externalDatabase
#> NULL
#> 
#> $`Murid betaherpesvirus 1`$commonName
#> NULL
#> 
#> $`Murid betaherpesvirus 1`$scientificName
#> [1] "Murid betaherpesvirus 1"
#> 
#> $`Murid betaherpesvirus 1`$isGenesUsable
#> [1] FALSE
#> 
#> $`Murid betaherpesvirus 1`$ncbiId
#> NULL
#> 
#> $`Murid betaherpesvirus 1`$isSpecies
#> NULL
#> 
#> $`Murid betaherpesvirus 1`$`_totalInQuery`
#> [1] 0
#> 
#> $`Murid betaherpesvirus 1`$id
#> [1] 137
#> 
#> 
#> $`Human immunodeficiency virus 1`
#> $`Human immunodeficiency virus 1`$externalDatabase
#> NULL
#> 
#> $`Human immunodeficiency virus 1`$commonName
#> NULL
#> 
#> $`Human immunodeficiency virus 1`$scientificName
#> [1] "Human immunodeficiency virus 1"
#> 
#> $`Human immunodeficiency virus 1`$isGenesUsable
#> [1] FALSE
#> 
#> $`Human immunodeficiency virus 1`$ncbiId
#> NULL
#> 
#> $`Human immunodeficiency virus 1`$isSpecies
#> NULL
#> 
#> $`Human immunodeficiency virus 1`$`_totalInQuery`
#> [1] 0
#> 
#> $`Human immunodeficiency virus 1`$id
#> [1] 138
#> 
#> 
#> $`Human polyomavirus 1`
#> $`Human polyomavirus 1`$externalDatabase
#> NULL
#> 
#> $`Human polyomavirus 1`$commonName
#> NULL
#> 
#> $`Human polyomavirus 1`$scientificName
#> [1] "Human polyomavirus 1"
#> 
#> $`Human polyomavirus 1`$isGenesUsable
#> [1] FALSE
#> 
#> $`Human polyomavirus 1`$ncbiId
#> NULL
#> 
#> $`Human polyomavirus 1`$isSpecies
#> NULL
#> 
#> $`Human polyomavirus 1`$`_totalInQuery`
#> [1] 0
#> 
#> $`Human polyomavirus 1`$id
#> [1] 139
#> 
#> 
#> $`Oncorhynchus mykiss`
#> $`Oncorhynchus mykiss`$externalDatabase
#> NULL
#> 
#> $`Oncorhynchus mykiss`$commonName
#> NULL
#> 
#> $`Oncorhynchus mykiss`$scientificName
#> [1] "Oncorhynchus mykiss"
#> 
#> $`Oncorhynchus mykiss`$isGenesUsable
#> [1] FALSE
#> 
#> $`Oncorhynchus mykiss`$ncbiId
#> NULL
#> 
#> $`Oncorhynchus mykiss`$isSpecies
#> NULL
#> 
#> $`Oncorhynchus mykiss`$`_totalInQuery`
#> [1] 0
#> 
#> $`Oncorhynchus mykiss`$id
#> [1] 140
#> 
#> 
#> $`Salmonella enterica subsp. enterica serovar Typhimurium str. SL1344`
#> $`Salmonella enterica subsp. enterica serovar Typhimurium str. SL1344`$externalDatabase
#> NULL
#> 
#> $`Salmonella enterica subsp. enterica serovar Typhimurium str. SL1344`$commonName
#> NULL
#> 
#> $`Salmonella enterica subsp. enterica serovar Typhimurium str. SL1344`$scientificName
#> [1] "Salmonella enterica subsp. enterica serovar Typhimurium str. SL1344"
#> 
#> $`Salmonella enterica subsp. enterica serovar Typhimurium str. SL1344`$isGenesUsable
#> [1] FALSE
#> 
#> $`Salmonella enterica subsp. enterica serovar Typhimurium str. SL1344`$ncbiId
#> NULL
#> 
#> $`Salmonella enterica subsp. enterica serovar Typhimurium str. SL1344`$isSpecies
#> NULL
#> 
#> $`Salmonella enterica subsp. enterica serovar Typhimurium str. SL1344`$`_totalInQuery`
#> [1] 0
#> 
#> $`Salmonella enterica subsp. enterica serovar Typhimurium str. SL1344`$id
#> [1] 141
#> 
#> 
#> $`Salmonella enterica subsp. enterica serovar Typhimurium`
#> $`Salmonella enterica subsp. enterica serovar Typhimurium`$externalDatabase
#> NULL
#> 
#> $`Salmonella enterica subsp. enterica serovar Typhimurium`$commonName
#> NULL
#> 
#> $`Salmonella enterica subsp. enterica serovar Typhimurium`$scientificName
#> [1] "Salmonella enterica subsp. enterica serovar Typhimurium"
#> 
#> $`Salmonella enterica subsp. enterica serovar Typhimurium`$isGenesUsable
#> [1] TRUE
#> 
#> $`Salmonella enterica subsp. enterica serovar Typhimurium`$ncbiId
#> NULL
#> 
#> $`Salmonella enterica subsp. enterica serovar Typhimurium`$isSpecies
#> NULL
#> 
#> $`Salmonella enterica subsp. enterica serovar Typhimurium`$`_totalInQuery`
#> [1] 0
#> 
#> $`Salmonella enterica subsp. enterica serovar Typhimurium`$id
#> [1] 142
#> 
#> 
#> $`Phaseolus vulgaris`
#> $`Phaseolus vulgaris`$externalDatabase
#> NULL
#> 
#> $`Phaseolus vulgaris`$commonName
#> NULL
#> 
#> $`Phaseolus vulgaris`$scientificName
#> [1] "Phaseolus vulgaris"
#> 
#> $`Phaseolus vulgaris`$isGenesUsable
#> [1] FALSE
#> 
#> $`Phaseolus vulgaris`$ncbiId
#> NULL
#> 
#> $`Phaseolus vulgaris`$isSpecies
#> NULL
#> 
#> $`Phaseolus vulgaris`$`_totalInQuery`
#> [1] 0
#> 
#> $`Phaseolus vulgaris`$id
#> [1] 143
#> 
#> 
#> $`Callithrix jacchus`
#> $`Callithrix jacchus`$externalDatabase
#> NULL
#> 
#> $`Callithrix jacchus`$commonName
#> NULL
#> 
#> $`Callithrix jacchus`$scientificName
#> [1] "Callithrix jacchus"
#> 
#> $`Callithrix jacchus`$isGenesUsable
#> [1] FALSE
#> 
#> $`Callithrix jacchus`$ncbiId
#> NULL
#> 
#> $`Callithrix jacchus`$isSpecies
#> NULL
#> 
#> $`Callithrix jacchus`$`_totalInQuery`
#> [1] 0
#> 
#> $`Callithrix jacchus`$id
#> [1] 144
#> 
#> 
#> $`Mustela putorius furo`
#> $`Mustela putorius furo`$externalDatabase
#> NULL
#> 
#> $`Mustela putorius furo`$commonName
#> NULL
#> 
#> $`Mustela putorius furo`$scientificName
#> [1] "Mustela putorius furo"
#> 
#> $`Mustela putorius furo`$isGenesUsable
#> [1] FALSE
#> 
#> $`Mustela putorius furo`$ncbiId
#> NULL
#> 
#> $`Mustela putorius furo`$isSpecies
#> NULL
#> 
#> $`Mustela putorius furo`$`_totalInQuery`
#> [1] 0
#> 
#> $`Mustela putorius furo`$id
#> [1] 145
#> 
#> 

allTaxa(c('human','mouse'))

#> $`Homo sapiens`
#> $`Homo sapiens`$externalDatabase
#> $`Homo sapiens`$externalDatabase$checked
#> [1] FALSE
#> 
#> $`Homo sapiens`$externalDatabase$name
#> [1] "hg38"
#> 
#> $`Homo sapiens`$externalDatabase$id
#> [1] 87
#> 
#> 
#> $`Homo sapiens`$commonName
#> [1] "human"
#> 
#> $`Homo sapiens`$scientificName
#> [1] "Homo sapiens"
#> 
#> $`Homo sapiens`$isGenesUsable
#> [1] TRUE
#> 
#> $`Homo sapiens`$ncbiId
#> [1] 9606
#> 
#> $`Homo sapiens`$isSpecies
#> NULL
#> 
#> $`Homo sapiens`$`_totalInQuery`
#> [1] 0
#> 
#> $`Homo sapiens`$id
#> [1] 1
#> 
#> 
#> $`Mus musculus`
#> $`Mus musculus`$externalDatabase
#> $`Mus musculus`$externalDatabase$checked
#> [1] FALSE
#> 
#> $`Mus musculus`$externalDatabase$name
#> [1] "mm10"
#> 
#> $`Mus musculus`$externalDatabase$id
#> [1] 81
#> 
#> 
#> $`Mus musculus`$commonName
#> [1] "mouse"
#> 
#> $`Mus musculus`$scientificName
#> [1] "Mus musculus"
#> 
#> $`Mus musculus`$isGenesUsable
#> [1] TRUE
#> 
#> $`Mus musculus`$ncbiId
#> [1] 10090
#> 
#> $`Mus musculus`$isSpecies
#> NULL
#> 
#> $`Mus musculus`$`_totalInQuery`
#> [1] 0
#> 
#> $`Mus musculus`$id
#> [1] 2
#> 
#>