annotationInfo

annotationInfo

annotationInfo(
  annotation,
  request = NULL,
  file = NULL,
  return = TRUE,
  overwrite = FALSE,
  memoised = FALSE
)

Arguments

annotation
Character. Either plain text, ontology term ID or an ontology term URI.

request
Character. If null retreeives the annotation object. Otherwise

datasets: Does a search for datasets containing annotations matching the given string

file
Character. File path. If provided, response will be saved to file

return
Logical. If the response should be returned. Set to false when you only want to save a file

overwrite
Logical. If TRUE, existing files will be overwritten. If FALSE a warning will be thrown and no action is taken.

memoised
Logical. If TRUE a memoised version of the function will be used which is faster for repeated requests. Use forgetGemmaMemoised to clear memory.

Value

A list

Examples

annotationInfo('http://purl.obolibrary.org/obo/OBI_0000105')

#> $transplantation
#> $transplantation$value
#> [1] "transplantation"
#> 
#> $transplantation$valueUri
#> [1] "http://purl.obolibrary.org/obo/OBI_0000105"
#> 
#> $transplantation$category
#> [1] "treatment"
#> 
#> $transplantation$categoryUri
#> [1] "http://www.ebi.ac.uk/efo/EFO_0000727"
#> 
#> 

annotationInfo('transplantation')

#> $transplantation
#> $transplantation$value
#> [1] "transplantation"
#> 
#> $transplantation$valueUri
#> [1] "http://purl.obolibrary.org/obo/OBI_0000105"
#> 
#> $transplantation$category
#> [1] "treatment"
#> 
#> $transplantation$categoryUri
#> [1] "http://www.ebi.ac.uk/efo/EFO_0000727"
#> 
#> 
#> $xenotransplantation
#> $xenotransplantation$value
#> [1] "xenotransplantation"
#> 
#> $xenotransplantation$valueUri
#> [1] "http://purl.obolibrary.org/obo/OBI_0000043"
#> 
#> $xenotransplantation$category
#> [1] "environmental history"
#> 
#> $xenotransplantation$categoryUri
#> [1] "http://www.ebi.ac.uk/efo/EFO_0004444"
#> 
#> 
#> $`patient with myelodysplastic syndrome refractory anemia with excess of blasts type 2 in remission after transplantation with allogeneic hematopoietic stem cells`
#> $`patient with myelodysplastic syndrome refractory anemia with excess of blasts type 2 in remission after transplantation with allogeneic hematopoietic stem cells`$value
#> [1] "patient with myelodysplastic syndrome refractory anemia with excess of blasts type 2 in remission after transplantation with allogeneic hematopoietic stem cells"
#> 
#> $`patient with myelodysplastic syndrome refractory anemia with excess of blasts type 2 in remission after transplantation with allogeneic hematopoietic stem cells`$valueUri
#> NULL
#> 
#> $`patient with myelodysplastic syndrome refractory anemia with excess of blasts type 2 in remission after transplantation with allogeneic hematopoietic stem cells`$category
#> [1] "disease"
#> 
#> $`patient with myelodysplastic syndrome refractory anemia with excess of blasts type 2 in remission after transplantation with allogeneic hematopoietic stem cells`$categoryUri
#> [1] "http://www.ebi.ac.uk/efo/EFO_0000408"
#> 
#> 
#> $`In vivo (14 after transplantation into the adult mouse brain) Neural Stem Cell (NSC) induced from Mouse Embryonic Fibroblast (MEF)`
#> $`In vivo (14 after transplantation into the adult mouse brain) Neural Stem Cell (NSC) induced from Mouse Embryonic Fibroblast (MEF)`$value
#> [1] "In vivo (14 after transplantation into the adult mouse brain) Neural Stem Cell (NSC) induced from Mouse Embryonic Fibroblast (MEF)"
#> 
#> $`In vivo (14 after transplantation into the adult mouse brain) Neural Stem Cell (NSC) induced from Mouse Embryonic Fibroblast (MEF)`$valueUri
#> NULL
#> 
#> $`In vivo (14 after transplantation into the adult mouse brain) Neural Stem Cell (NSC) induced from Mouse Embryonic Fibroblast (MEF)`$category
#> [1] "cell type"
#> 
#> $`In vivo (14 after transplantation into the adult mouse brain) Neural Stem Cell (NSC) induced from Mouse Embryonic Fibroblast (MEF)`$categoryUri
#> [1] "http://www.ebi.ac.uk/efo/EFO_0000324"
#> 
#> 
#> $`4 months after transduction/transplantation with RUNX1(41-214)-expressing retroviral vector`
#> $`4 months after transduction/transplantation with RUNX1(41-214)-expressing retroviral vector`$value
#> [1] "4 months after transduction/transplantation with RUNX1(41-214)-expressing retroviral vector"
#> 
#> $`4 months after transduction/transplantation with RUNX1(41-214)-expressing retroviral vector`$valueUri
#> NULL
#> 
#> $`4 months after transduction/transplantation with RUNX1(41-214)-expressing retroviral vector`$category
#> NULL
#> 
#> $`4 months after transduction/transplantation with RUNX1(41-214)-expressing retroviral vector`$categoryUri
#> NULL
#> 
#> 
#> $`patient with chronic myeloid leukemia in remission after transplantation with allogeneic hematopoietic stem cells`
#> $`patient with chronic myeloid leukemia in remission after transplantation with allogeneic hematopoietic stem cells`$value
#> [1] "patient with chronic myeloid leukemia in remission after transplantation with allogeneic hematopoietic stem cells"
#> 
#> $`patient with chronic myeloid leukemia in remission after transplantation with allogeneic hematopoietic stem cells`$valueUri
#> NULL
#> 
#> $`patient with chronic myeloid leukemia in remission after transplantation with allogeneic hematopoietic stem cells`$category
#> [1] "disease"
#> 
#> $`patient with chronic myeloid leukemia in remission after transplantation with allogeneic hematopoietic stem cells`$categoryUri
#> [1] "http://www.ebi.ac.uk/efo/EFO_0000408"
#> 
#> 
#> $`12month protocol bx (post-renal transplantation)`
#> $`12month protocol bx (post-renal transplantation)`$value
#> [1] "12month protocol bx (post-renal transplantation)"
#> 
#> $`12month protocol bx (post-renal transplantation)`$valueUri
#> NULL
#> 
#> $`12month protocol bx (post-renal transplantation)`$category
#> NULL
#> 
#> $`12month protocol bx (post-renal transplantation)`$categoryUri
#> NULL
#> 
#> 
#> $`liver transplantation`
#> $`liver transplantation`$value
#> [1] "liver transplantation"
#> 
#> $`liver transplantation`$valueUri
#> NULL
#> 
#> $`liver transplantation`$category
#> [1] "treatment"
#> 
#> $`liver transplantation`$categoryUri
#> [1] "http://www.ebi.ac.uk/efo/EFO_0000727"
#> 
#> 
#> $`hematopoietic cell transplantation`
#> $`hematopoietic cell transplantation`$value
#> [1] "hematopoietic cell transplantation"
#> 
#> $`hematopoietic cell transplantation`$valueUri
#> NULL
#> 
#> $`hematopoietic cell transplantation`$category
#> [1] "treatment"
#> 
#> $`hematopoietic cell transplantation`$categoryUri
#> [1] "http://www.ebi.ac.uk/efo/EFO_0000727"
#> 
#> 
#> $`rejection time (days after transplantation): 39`
#> $`rejection time (days after transplantation): 39`$value
#> [1] "rejection time (days after transplantation): 39"
#> 
#> $`rejection time (days after transplantation): 39`$valueUri
#> NULL
#> 
#> $`rejection time (days after transplantation): 39`$category
#> NULL
#> 
#> $`rejection time (days after transplantation): 39`$categoryUri
#> NULL
#> 
#> 
#> $`patient with acute myeloid leukemia in remission after transplantation with allogeneic hematopoietic stem cells`
#> $`patient with acute myeloid leukemia in remission after transplantation with allogeneic hematopoietic stem cells`$value
#> [1] "patient with acute myeloid leukemia in remission after transplantation with allogeneic hematopoietic stem cells"
#> 
#> $`patient with acute myeloid leukemia in remission after transplantation with allogeneic hematopoietic stem cells`$valueUri
#> NULL
#> 
#> $`patient with acute myeloid leukemia in remission after transplantation with allogeneic hematopoietic stem cells`$category
#> [1] "disease"
#> 
#> $`patient with acute myeloid leukemia in remission after transplantation with allogeneic hematopoietic stem cells`$categoryUri
#> [1] "http://www.ebi.ac.uk/efo/EFO_0000408"
#> 
#> 
#> $`Donor dendritic cells from medulla of renal allograft (3 days post-transplantation)`
#> $`Donor dendritic cells from medulla of renal allograft (3 days post-transplantation)`$value
#> [1] "Donor dendritic cells from medulla of renal allograft (3 days post-transplantation)"
#> 
#> $`Donor dendritic cells from medulla of renal allograft (3 days post-transplantation)`$valueUri
#> NULL
#> 
#> $`Donor dendritic cells from medulla of renal allograft (3 days post-transplantation)`$category
#> [1] "BioSource"
#> 
#> $`Donor dendritic cells from medulla of renal allograft (3 days post-transplantation)`$categoryUri
#> [1] "http://www.ebi.ac.uk/efo/EFO_0000635"
#> 
#> 
#> $`rejection time (days after transplantation): 3`
#> $`rejection time (days after transplantation): 3`$value
#> [1] "rejection time (days after transplantation): 3"
#> 
#> $`rejection time (days after transplantation): 3`$valueUri
#> NULL
#> 
#> $`rejection time (days after transplantation): 3`$category
#> NULL
#> 
#> $`rejection time (days after transplantation): 3`$categoryUri
#> NULL
#> 
#> 
#> $`rejection time (days after transplantation): 73`
#> $`rejection time (days after transplantation): 73`$value
#> [1] "rejection time (days after transplantation): 73"
#> 
#> $`rejection time (days after transplantation): 73`$valueUri
#> NULL
#> 
#> $`rejection time (days after transplantation): 73`$category
#> NULL
#> 
#> $`rejection time (days after transplantation): 73`$categoryUri
#> NULL
#> 
#> 
#> $`time after transplantation: ?2 months`
#> $`time after transplantation: ?2 months`$value
#> [1] "time after transplantation: ?2 months"
#> 
#> $`time after transplantation: ?2 months`$valueUri
#> NULL
#> 
#> $`time after transplantation: ?2 months`$category
#> NULL
#> 
#> $`time after transplantation: ?2 months`$categoryUri
#> NULL
#> 
#> 
#> $`rejection time (days after transplantation): 35`
#> $`rejection time (days after transplantation): 35`$value
#> [1] "rejection time (days after transplantation): 35"
#> 
#> $`rejection time (days after transplantation): 35`$valueUri
#> NULL
#> 
#> $`rejection time (days after transplantation): 35`$category
#> NULL
#> 
#> $`rejection time (days after transplantation): 35`$categoryUri
#> NULL
#> 
#> 
#> $`advanced heart failure-DCM-serum-at LVAD explantation for heart transplantation`
#> $`advanced heart failure-DCM-serum-at LVAD explantation for heart transplantation`$value
#> [1] "advanced heart failure-DCM-serum-at LVAD explantation for heart transplantation"
#> 
#> $`advanced heart failure-DCM-serum-at LVAD explantation for heart transplantation`$valueUri
#> NULL
#> 
#> $`advanced heart failure-DCM-serum-at LVAD explantation for heart transplantation`$category
#> [1] "BioSource"
#> 
#> $`advanced heart failure-DCM-serum-at LVAD explantation for heart transplantation`$categoryUri
#> [1] "http://www.ebi.ac.uk/efo/EFO_0000635"
#> 
#> 
#> $`differentiated for two months in vivo after transplantation`
#> $`differentiated for two months in vivo after transplantation`$value
#> [1] "differentiated for two months in vivo after transplantation"
#> 
#> $`differentiated for two months in vivo after transplantation`$valueUri
#> NULL
#> 
#> $`differentiated for two months in vivo after transplantation`$category
#> [1] "cell type"
#> 
#> $`differentiated for two months in vivo after transplantation`$categoryUri
#> [1] "http://www.ebi.ac.uk/efo/EFO_0000324"
#> 
#> 
#> $`advanced heart failure-ICM-plasma-at LVAD explantation for heart transplantation`
#> $`advanced heart failure-ICM-plasma-at LVAD explantation for heart transplantation`$value
#> [1] "advanced heart failure-ICM-plasma-at LVAD explantation for heart transplantation"
#> 
#> $`advanced heart failure-ICM-plasma-at LVAD explantation for heart transplantation`$valueUri
#> NULL
#> 
#> $`advanced heart failure-ICM-plasma-at LVAD explantation for heart transplantation`$category
#> [1] "BioSource"
#> 
#> $`advanced heart failure-ICM-plasma-at LVAD explantation for heart transplantation`$categoryUri
#> [1] "http://www.ebi.ac.uk/efo/EFO_0000635"
#> 
#> 
#> $`advanced heart failure-ICM-serum-at LVAD explantation for heart transplantation`
#> $`advanced heart failure-ICM-serum-at LVAD explantation for heart transplantation`$value
#> [1] "advanced heart failure-ICM-serum-at LVAD explantation for heart transplantation"
#> 
#> $`advanced heart failure-ICM-serum-at LVAD explantation for heart transplantation`$valueUri
#> NULL
#> 
#> $`advanced heart failure-ICM-serum-at LVAD explantation for heart transplantation`$category
#> [1] "BioSource"
#> 
#> $`advanced heart failure-ICM-serum-at LVAD explantation for heart transplantation`$categoryUri
#> [1] "http://www.ebi.ac.uk/efo/EFO_0000635"
#> 
#> 
#> $`gxbdataset_biirinternal: P3 TRANSPLANTATION-BUMC (26) 11Jun08 Liver Transplant for Modules`
#> $`gxbdataset_biirinternal: P3 TRANSPLANTATION-BUMC (26) 11Jun08 Liver Transplant for Modules`$value
#> [1] "gxbdataset_biirinternal: P3 TRANSPLANTATION-BUMC (26) 11Jun08 Liver Transplant for Modules"
#> 
#> $`gxbdataset_biirinternal: P3 TRANSPLANTATION-BUMC (26) 11Jun08 Liver Transplant for Modules`$valueUri
#> NULL
#> 
#> $`gxbdataset_biirinternal: P3 TRANSPLANTATION-BUMC (26) 11Jun08 Liver Transplant for Modules`$category
#> NULL
#> 
#> $`gxbdataset_biirinternal: P3 TRANSPLANTATION-BUMC (26) 11Jun08 Liver Transplant for Modules`$categoryUri
#> NULL
#> 
#> 
#> $`rejection time (days after transplantation): 7`
#> $`rejection time (days after transplantation): 7`$value
#> [1] "rejection time (days after transplantation): 7"
#> 
#> $`rejection time (days after transplantation): 7`$valueUri
#> NULL
#> 
#> $`rejection time (days after transplantation): 7`$category
#> NULL
#> 
#> $`rejection time (days after transplantation): 7`$categoryUri
#> NULL
#> 
#> 
#> $`rejection time (days after transplantation): 8`
#> $`rejection time (days after transplantation): 8`$value
#> [1] "rejection time (days after transplantation): 8"
#> 
#> $`rejection time (days after transplantation): 8`$valueUri
#> NULL
#> 
#> $`rejection time (days after transplantation): 8`$category
#> NULL
#> 
#> $`rejection time (days after transplantation): 8`$categoryUri
#> NULL
#> 
#> 
#> $`rejection time (days after transplantation): 9`
#> $`rejection time (days after transplantation): 9`$value
#> [1] "rejection time (days after transplantation): 9"
#> 
#> $`rejection time (days after transplantation): 9`$valueUri
#> NULL
#> 
#> $`rejection time (days after transplantation): 9`$category
#> NULL
#> 
#> $`rejection time (days after transplantation): 9`$categoryUri
#> NULL
#> 
#> 
#> $`Host dendritic cells from cortex of renal allograft (3 days post-transplantation)`
#> $`Host dendritic cells from cortex of renal allograft (3 days post-transplantation)`$value
#> [1] "Host dendritic cells from cortex of renal allograft (3 days post-transplantation)"
#> 
#> $`Host dendritic cells from cortex of renal allograft (3 days post-transplantation)`$valueUri
#> NULL
#> 
#> $`Host dendritic cells from cortex of renal allograft (3 days post-transplantation)`$category
#> [1] "BioSource"
#> 
#> $`Host dendritic cells from cortex of renal allograft (3 days post-transplantation)`$categoryUri
#> [1] "http://www.ebi.ac.uk/efo/EFO_0000635"
#> 
#> 
#> $`rejection time (days after transplantation): 69`
#> $`rejection time (days after transplantation): 69`$value
#> [1] "rejection time (days after transplantation): 69"
#> 
#> $`rejection time (days after transplantation): 69`$valueUri
#> NULL
#> 
#> $`rejection time (days after transplantation): 69`$category
#> NULL
#> 
#> $`rejection time (days after transplantation): 69`$categoryUri
#> NULL
#> 
#> 
#> $`deterioration of graft function within 2 years after transplantation: yes`
#> $`deterioration of graft function within 2 years after transplantation: yes`$value
#> [1] "deterioration of graft function within 2 years after transplantation: yes"
#> 
#> $`deterioration of graft function within 2 years after transplantation: yes`$valueUri
#> NULL
#> 
#> $`deterioration of graft function within 2 years after transplantation: yes`$category
#> NULL
#> 
#> $`deterioration of graft function within 2 years after transplantation: yes`$categoryUri
#> NULL
#> 
#> 
#> $`rejection time (days after transplantation): 28`
#> $`rejection time (days after transplantation): 28`$value
#> [1] "rejection time (days after transplantation): 28"
#> 
#> $`rejection time (days after transplantation): 28`$valueUri
#> NULL
#> 
#> $`rejection time (days after transplantation): 28`$category
#> NULL
#> 
#> $`rejection time (days after transplantation): 28`$categoryUri
#> NULL
#> 
#> 
#> $`rejection time (days after transplantation): 61`
#> $`rejection time (days after transplantation): 61`$value
#> [1] "rejection time (days after transplantation): 61"
#> 
#> $`rejection time (days after transplantation): 61`$valueUri
#> NULL
#> 
#> $`rejection time (days after transplantation): 61`$category
#> NULL
#> 
#> $`rejection time (days after transplantation): 61`$categoryUri
#> NULL
#> 
#> 
#> $`rejection time (days after transplantation): 22`
#> $`rejection time (days after transplantation): 22`$value
#> [1] "rejection time (days after transplantation): 22"
#> 
#> $`rejection time (days after transplantation): 22`$valueUri
#> NULL
#> 
#> $`rejection time (days after transplantation): 22`$category
#> NULL
#> 
#> $`rejection time (days after transplantation): 22`$categoryUri
#> NULL
#> 
#> 
#> $`advanced heart failure-DCM-plasma-at LVAD explantation for heart transplantation`
#> $`advanced heart failure-DCM-plasma-at LVAD explantation for heart transplantation`$value
#> [1] "advanced heart failure-DCM-plasma-at LVAD explantation for heart transplantation"
#> 
#> $`advanced heart failure-DCM-plasma-at LVAD explantation for heart transplantation`$valueUri
#> NULL
#> 
#> $`advanced heart failure-DCM-plasma-at LVAD explantation for heart transplantation`$category
#> [1] "BioSource"
#> 
#> $`advanced heart failure-DCM-plasma-at LVAD explantation for heart transplantation`$categoryUri
#> [1] "http://www.ebi.ac.uk/efo/EFO_0000635"
#> 
#> 
#> $`time after transplantation: 3 months`
#> $`time after transplantation: 3 months`$value
#> [1] "time after transplantation: 3 months"
#> 
#> $`time after transplantation: 3 months`$valueUri
#> NULL
#> 
#> $`time after transplantation: 3 months`$category
#> NULL
#> 
#> $`time after transplantation: 3 months`$categoryUri
#> NULL
#> 
#> 
#> $`Hematopoietic progenitor cells (HPC) were enriched within the CD34+ cell fraction (selection by magnetic beads and subsequent flow cytometric separation) from G-CSF mobilized blood from healthy donors for allogeneic blood stem cell transplantation.`
#> $`Hematopoietic progenitor cells (HPC) were enriched within the CD34+ cell fraction (selection by magnetic beads and subsequent flow cytometric separation) from G-CSF mobilized blood from healthy donors for allogeneic blood stem cell transplantation.`$value
#> [1] "Hematopoietic progenitor cells (HPC) were enriched within the CD34+ cell fraction (selection by magnetic beads and subsequent flow cytometric separation) from G-CSF mobilized blood from healthy donors for allogeneic blood stem cell transplantation."
#> 
#> $`Hematopoietic progenitor cells (HPC) were enriched within the CD34+ cell fraction (selection by magnetic beads and subsequent flow cytometric separation) from G-CSF mobilized blood from healthy donors for allogeneic blood stem cell transplantation.`$...
#> NULL
#> 
#> $`Hematopoietic progenitor cells (HPC) were enriched within the CD34+ cell fraction (selection by magnetic beads and subsequent flow cytometric separation) from G-CSF mobilized blood from healthy donors for allogeneic blood stem cell transplantation.`$...
#> NULL
#> 
#> $`Hematopoietic progenitor cells (HPC) were enriched within the CD34+ cell fraction (selection by magnetic beads and subsequent flow cytometric separation) from G-CSF mobilized blood from healthy donors for allogeneic blood stem cell transplantation.`$...
#> NULL
#> 
#> 
#> $`kidney, transplanted from Dark Agouti rats into Lewis rats, high weight difference between donor and recipient, treatment with anti-IL6R antibody immediately after transplantation`
#> $`kidney, transplanted from Dark Agouti rats into Lewis rats, high weight difference between donor and recipient, treatment with anti-IL6R antibody immediately after transplantation`$value
#> [1] "kidney, transplanted from Dark Agouti rats into Lewis rats, high weight difference between donor and recipient, treatment with anti-IL6R antibody immediately after transplantation"
#> 
#> $`kidney, transplanted from Dark Agouti rats into Lewis rats, high weight difference between donor and recipient, treatment with anti-IL6R antibody immediately after transplantation`$valueUri
#> NULL
#> 
#> $`kidney, transplanted from Dark Agouti rats into Lewis rats, high weight difference between donor and recipient, treatment with anti-IL6R antibody immediately after transplantation`$category
#> [1] "organism part"
#> 
#> $`kidney, transplanted from Dark Agouti rats into Lewis rats, high weight difference between donor and recipient, treatment with anti-IL6R antibody immediately after transplantation`$categoryUri
#> [1] "http://www.ebi.ac.uk/efo/EFO_0000635"
#> 
#> 
#> $`rejection time (days after transplantation): 15`
#> $`rejection time (days after transplantation): 15`$value
#> [1] "rejection time (days after transplantation): 15"
#> 
#> $`rejection time (days after transplantation): 15`$valueUri
#> NULL
#> 
#> $`rejection time (days after transplantation): 15`$category
#> NULL
#> 
#> $`rejection time (days after transplantation): 15`$categoryUri
#> NULL
#> 
#> 
#> $`explanted heart tissue at time of cardiac transplantation`
#> $`explanted heart tissue at time of cardiac transplantation`$value
#> [1] "explanted heart tissue at time of cardiac transplantation"
#> 
#> $`explanted heart tissue at time of cardiac transplantation`$valueUri
#> NULL
#> 
#> $`explanted heart tissue at time of cardiac transplantation`$category
#> [1] "BioSource"
#> 
#> $`explanted heart tissue at time of cardiac transplantation`$categoryUri
#> [1] "http://www.ebi.ac.uk/efo/EFO_0000635"
#> 
#> 
#> $`olfactory ensheathing cell transplantation `
#> $`olfactory ensheathing cell transplantation `$value
#> [1] "olfactory ensheathing cell transplantation "
#> 
#> $`olfactory ensheathing cell transplantation `$valueUri
#> NULL
#> 
#> $`olfactory ensheathing cell transplantation `$category
#> [1] "treatment"
#> 
#> $`olfactory ensheathing cell transplantation `$categoryUri
#> [1] "http://www.ebi.ac.uk/efo/EFO_0000727"
#> 
#> 
#> $`rejection time (days after transplantation): 16`
#> $`rejection time (days after transplantation): 16`$value
#> [1] "rejection time (days after transplantation): 16"
#> 
#> $`rejection time (days after transplantation): 16`$valueUri
#> NULL
#> 
#> $`rejection time (days after transplantation): 16`$category
#> NULL
#> 
#> $`rejection time (days after transplantation): 16`$categoryUri
#> NULL
#> 
#> 
#> $`rejection time (days after transplantation): 18`
#> $`rejection time (days after transplantation): 18`$value
#> [1] "rejection time (days after transplantation): 18"
#> 
#> $`rejection time (days after transplantation): 18`$valueUri
#> NULL
#> 
#> $`rejection time (days after transplantation): 18`$category
#> NULL
#> 
#> $`rejection time (days after transplantation): 18`$categoryUri
#> NULL
#> 
#> 
#> $`rejection time (days after transplantation): 53`
#> $`rejection time (days after transplantation): 53`$value
#> [1] "rejection time (days after transplantation): 53"
#> 
#> $`rejection time (days after transplantation): 53`$valueUri
#> NULL
#> 
#> $`rejection time (days after transplantation): 53`$category
#> NULL
#> 
#> $`rejection time (days after transplantation): 53`$categoryUri
#> NULL
#> 
#> 
#> $`rejection time (days after transplantation): 99`
#> $`rejection time (days after transplantation): 99`$value
#> [1] "rejection time (days after transplantation): 99"
#> 
#> $`rejection time (days after transplantation): 99`$valueUri
#> NULL
#> 
#> $`rejection time (days after transplantation): 99`$category
#> NULL
#> 
#> $`rejection time (days after transplantation): 99`$categoryUri
#> NULL
#> 
#> 
#> $`kidney transplantation`
#> $`kidney transplantation`$value
#> [1] "kidney transplantation"
#> 
#> $`kidney transplantation`$valueUri
#> NULL
#> 
#> $`kidney transplantation`$category
#> [1] "treatment"
#> 
#> $`kidney transplantation`$categoryUri
#> [1] "http://www.ebi.ac.uk/efo/EFO_0000727"
#> 
#> 
#> $`rejection time (days after transplantation): 10`
#> $`rejection time (days after transplantation): 10`$value
#> [1] "rejection time (days after transplantation): 10"
#> 
#> $`rejection time (days after transplantation): 10`$valueUri
#> NULL
#> 
#> $`rejection time (days after transplantation): 10`$category
#> NULL
#> 
#> $`rejection time (days after transplantation): 10`$categoryUri
#> NULL
#> 
#> 
#> $`rejection time (days after transplantation): 13`
#> $`rejection time (days after transplantation): 13`$value
#> [1] "rejection time (days after transplantation): 13"
#> 
#> $`rejection time (days after transplantation): 13`$valueUri
#> NULL
#> 
#> $`rejection time (days after transplantation): 13`$category
#> NULL
#> 
#> $`rejection time (days after transplantation): 13`$categoryUri
#> NULL
#> 
#> 
#> $`rejection time (days after transplantation): 12`
#> $`rejection time (days after transplantation): 12`$value
#> [1] "rejection time (days after transplantation): 12"
#> 
#> $`rejection time (days after transplantation): 12`$valueUri
#> NULL
#> 
#> $`rejection time (days after transplantation): 12`$category
#> NULL
#> 
#> $`rejection time (days after transplantation): 12`$categoryUri
#> NULL
#> 
#> 
#> $`mesenchymal stem cell transplantation`
#> $`mesenchymal stem cell transplantation`$value
#> [1] "mesenchymal stem cell transplantation"
#> 
#> $`mesenchymal stem cell transplantation`$valueUri
#> NULL
#> 
#> $`mesenchymal stem cell transplantation`$category
#> [1] "treatment"
#> 
#> $`mesenchymal stem cell transplantation`$categoryUri
#> [1] "http://www.ebi.ac.uk/efo/EFO_0000727"
#> 
#> 
#> $`Host dendritic cells from medulla of renal allograft (3 days post-transplantation)`
#> $`Host dendritic cells from medulla of renal allograft (3 days post-transplantation)`$value
#> [1] "Host dendritic cells from medulla of renal allograft (3 days post-transplantation)"
#> 
#> $`Host dendritic cells from medulla of renal allograft (3 days post-transplantation)`$valueUri
#> NULL
#> 
#> $`Host dendritic cells from medulla of renal allograft (3 days post-transplantation)`$category
#> [1] "BioSource"
#> 
#> $`Host dendritic cells from medulla of renal allograft (3 days post-transplantation)`$categoryUri
#> [1] "http://www.ebi.ac.uk/efo/EFO_0000635"
#> 
#> 
#> $`bone marrow transplantation`
#> $`bone marrow transplantation`$value
#> [1] "bone marrow transplantation"
#> 
#> $`bone marrow transplantation`$valueUri
#> NULL
#> 
#> $`bone marrow transplantation`$category
#> [1] "treatment"
#> 
#> $`bone marrow transplantation`$categoryUri
#> NULL
#> 
#> 
#> $`status: differentiated in vivo for two months upon transplantation in the lesioned spinal cord.`
#> $`status: differentiated in vivo for two months upon transplantation in the lesioned spinal cord.`$value
#> [1] "status: differentiated in vivo for two months upon transplantation in the lesioned spinal cord."
#> 
#> $`status: differentiated in vivo for two months upon transplantation in the lesioned spinal cord.`$valueUri
#> NULL
#> 
#> $`status: differentiated in vivo for two months upon transplantation in the lesioned spinal cord.`$category
#> NULL
#> 
#> $`status: differentiated in vivo for two months upon transplantation in the lesioned spinal cord.`$categoryUri
#> NULL
#> 
#> 
#> $`microbiota transplantation`
#> $`microbiota transplantation`$value
#> [1] "microbiota transplantation"
#> 
#> $`microbiota transplantation`$valueUri
#> NULL
#> 
#> $`microbiota transplantation`$category
#> [1] "treatment"
#> 
#> $`microbiota transplantation`$categoryUri
#> [1] "http://www.ebi.ac.uk/efo/EFO_0000727"
#> 
#> 
#> $`Donor dendritic cells from cortex of renal allograft (3 days post-transplantation)`
#> $`Donor dendritic cells from cortex of renal allograft (3 days post-transplantation)`$value
#> [1] "Donor dendritic cells from cortex of renal allograft (3 days post-transplantation)"
#> 
#> $`Donor dendritic cells from cortex of renal allograft (3 days post-transplantation)`$valueUri
#> NULL
#> 
#> $`Donor dendritic cells from cortex of renal allograft (3 days post-transplantation)`$category
#> [1] "BioSource"
#> 
#> $`Donor dendritic cells from cortex of renal allograft (3 days post-transplantation)`$categoryUri
#> [1] "http://www.ebi.ac.uk/efo/EFO_0000635"
#> 
#> 
#> $`patient with myelofibrose in remission after transplantation with allogeneic hematopoietic stem cells`
#> $`patient with myelofibrose in remission after transplantation with allogeneic hematopoietic stem cells`$value
#> [1] "patient with myelofibrose in remission after transplantation with allogeneic hematopoietic stem cells"
#> 
#> $`patient with myelofibrose in remission after transplantation with allogeneic hematopoietic stem cells`$valueUri
#> NULL
#> 
#> $`patient with myelofibrose in remission after transplantation with allogeneic hematopoietic stem cells`$category
#> [1] "disease"
#> 
#> $`patient with myelofibrose in remission after transplantation with allogeneic hematopoietic stem cells`$categoryUri
#> [1] "http://www.ebi.ac.uk/efo/EFO_0000408"
#> 
#> 
#> $`kidney, transplanted from Dark Agouti rats into Lewis rats, high weight difference between donor and recipient, treatment with prednisolone immediately after transplantation`
#> $`kidney, transplanted from Dark Agouti rats into Lewis rats, high weight difference between donor and recipient, treatment with prednisolone immediately after transplantation`$value
#> [1] "kidney, transplanted from Dark Agouti rats into Lewis rats, high weight difference between donor and recipient, treatment with prednisolone immediately after transplantation"
#> 
#> $`kidney, transplanted from Dark Agouti rats into Lewis rats, high weight difference between donor and recipient, treatment with prednisolone immediately after transplantation`$valueUri
#> NULL
#> 
#> $`kidney, transplanted from Dark Agouti rats into Lewis rats, high weight difference between donor and recipient, treatment with prednisolone immediately after transplantation`$category
#> [1] "organism part"
#> 
#> $`kidney, transplanted from Dark Agouti rats into Lewis rats, high weight difference between donor and recipient, treatment with prednisolone immediately after transplantation`$categoryUri
#> [1] "http://www.ebi.ac.uk/efo/EFO_0000635"
#> 
#> 
#> $`Surgery due to:\t\tlung transplantation`
#> $`Surgery due to:\t\tlung transplantation`$value
#> [1] "Surgery due to:\t\tlung transplantation"
#> 
#> $`Surgery due to:\t\tlung transplantation`$valueUri
#> NULL
#> 
#> $`Surgery due to:\t\tlung transplantation`$category
#> NULL
#> 
#> $`Surgery due to:\t\tlung transplantation`$categoryUri
#> NULL
#> 
#> 
#> $`patient with chronic myeloid leukemia in remission prior to transplantation with allogeneic hematopoietic stem cells`
#> $`patient with chronic myeloid leukemia in remission prior to transplantation with allogeneic hematopoietic stem cells`$value
#> [1] "patient with chronic myeloid leukemia in remission prior to transplantation with allogeneic hematopoietic stem cells"
#> 
#> $`patient with chronic myeloid leukemia in remission prior to transplantation with allogeneic hematopoietic stem cells`$valueUri
#> NULL
#> 
#> $`patient with chronic myeloid leukemia in remission prior to transplantation with allogeneic hematopoietic stem cells`$category
#> [1] "disease"
#> 
#> $`patient with chronic myeloid leukemia in remission prior to transplantation with allogeneic hematopoietic stem cells`$categoryUri
#> [1] "http://www.ebi.ac.uk/efo/EFO_0000408"
#> 
#> 
#> $`deterioration of graft function within 2 years after transplantation: no`
#> $`deterioration of graft function within 2 years after transplantation: no`$value
#> [1] "deterioration of graft function within 2 years after transplantation: no"
#> 
#> $`deterioration of graft function within 2 years after transplantation: no`$valueUri
#> NULL
#> 
#> $`deterioration of graft function within 2 years after transplantation: no`$category
#> NULL
#> 
#> $`deterioration of graft function within 2 years after transplantation: no`$categoryUri
#> NULL
#> 
#> 

annotationInfo('http://purl.obolibrary.org/obo/OBI_0000105', request = 'datasets')

#> named list()