datasetInfo

Retrieves information about a single dataset based on the given dataset identifier. Combines several API calls

datasetInfo(
  dataset,
  request = NULL,
  ...,
  file = NULL,
  return = TRUE,
  overwrite = FALSE,
  memoised = FALSE
)

Arguments

dataset
Character. Can either be the dataset ID or its short name (e.g. GSE1234). If a vector of length>1 is provided return all matching dataset objects similar to allDatasets but without access to additional parameters. request parameter cannot be specified for inputs of length>1 unless specifiet otherwise in request description.

request
Character. If NULL retrieves the dataset object. Otherwise

platforms: Retrieves platforms for the given dataset

samples: Retrieves samples for the given dataset

annotations: Retrieves the annotations for the given dataset

design: Retrieves the design for the given dataset

data: Retrieves the data for the given dataset. Parameters:

filter: Optional, defaults to FALSE. If TRUE, call returns filtered expression data.

IdColnames: Optional. defaults to FALSE. If true shortens data column names to only include the bioAssayId which is unique to samples in Gemma. Makes it easier to match to samples acqured from the "samples" request.

differential: Retrieves available differential expression tests for the given dataset.

offset: Optional, defaults to 0. Skips the specified amount of objects when retrieving them from the database.

limit: Optional, defaults to 20. Limits the result to specified amount of objects. Use 0 for no limit.

degs: Retrieves the differential expression results for the given dataset.

differential: Differential id of the differential expression. Can be acquired from the differential endpoint.

diffExExpr: Retrieves expression values for differential expression subsets for the given datasets. Parameters:

diffExSet: Result set id of the differential expression. Can be acquired from the differential endpoint:

datasetInfo('GSE43364',request = 'differential')$resultSets[[1]]$id

keepNonSpecific: Optional, defaults to FALSE.

If set to false, the response will only include elements that map exclusively to each gene

If set to true, the response will include all elements that map to each gene, even if they also map to other genes.

threshold: Optional, defaults to 100. The threshold that the differential expression has to meet to be included in the response.

limit: Optional, defaults to 100. Maximum amount of returned gene-probe expression level pairs to include in the response.

consolidate: Optional. Defaults no NULL.

Whether genes with multiple elements should consolidate the information. If the 'keepNonSpecific' parameter is set to true, then all gene non-specific vectors are excluded from the chosen procedure.

The options are:

NULL:list all vectors separately.

"pickmax": only return the vector that has the highest expression (mean over all its bioAssays).

"pickvar": only return the vector with highest variance of expression across its bioAssays

"average": create a new vector that will average the bioAssay values from all vectors

geneExpression: Retrieves the expression levels of given genes for given datasets. Can be used with multiple datasets. Parameters:

genes: Required. A list of identifiers, separated by commas (e.g: 1859, 5728).

Can either be the NCBI ID (1859), Ensembl ID (ENSG00000157540) or official symbol (DYRK1A) of the gene.

NCBI ID is the most efficient (and guaranteed to be unique) identifier.

Official symbol represents a gene homologue for a random taxon, unless used in a specific taxon (see the Taxa Endpoints).

If the gene taxon does not match the taxon of the given datasets, expression levels for that gene will be missing from the response

You can combine various identifiers in one query, but any invalid identifier will cause the call to yield an error. Duplicate identifiers of the same gene will result in duplicates in the response.

keepNonSpecific: Optional. Defaults to FALSE.

If set to false, the response will only include elements that map exclusively to each queried gene

If set to true, the response will include all elements that map to each queried gene, even if they also map to other genes.

consolidate: Optional. Defaults no NULL.

Whether genes with multiple elements should consolidate the information. If the 'keepNonSpecific' parameter is set to true, then all gene non-specific vectors are excluded from the chosen procedure.

The options are:

NULL:list all vectors separately.

"pickmax": only return the vector that has the highest expression (mean over all its bioAssays).

"pickvar": only return the vector with highest variance of expression across its bioAssays

"average": create a new vector that will average the bioAssay values from all vectors

...
Use if the specified request has additional parameters.

file
Character. File path. If provided, response will be saved to file

return
Logical. If the response should be returned. Set to false when you only want to save a file

overwrite
Logical. If TRUE, existing files will be overwritten. If FALSE a warning will be thrown and no action is taken.

memoised
Logical. If TRUE a memoised version of the function will be used which is faster for repeated requests. Use forgetGemmaMemoised to clear memory.

Value

A data.frame or a list depending on the request

Examples

datasetInfo('GSE81454')

#> $GSE81454
#> $GSE81454$shortName
#> [1] "GSE81454"
#> 
#> $GSE81454$metadata
#> NULL
#> 
#> $GSE81454$userOwned
#> [1] FALSE
#> 
#> $GSE81454$bioMaterialCount
#> NULL
#> 
#> $GSE81454$processedExpressionVectorCount
#> [1] 49395
#> 
#> $GSE81454$batchConfound
#> NULL
#> 
#> $GSE81454$batchEffect
#> NULL
#> 
#> $GSE81454$accession
#> [1] "GSE81454"
#> 
#> $GSE81454$externalDatabase
#> [1] "GEO"
#> 
#> $GSE81454$taxon
#> [1] "human"
#> 
#> $GSE81454$taxonId
#> [1] 1
#> 
#> $GSE81454$experimentalDesign
#> [1] 9321
#> 
#> $GSE81454$technologyType
#> [1] "ONECOLOR"
#> 
#> $GSE81454$isPublic
#> [1] TRUE
#> 
#> $GSE81454$geeq
#> $GSE81454$geeq$sScorePublication
#> [1] 1
#> 
#> $GSE81454$geeq$sScorePlatformAmount
#> [1] 1
#> 
#> $GSE81454$geeq$sScorePlatformsTechMulti
#> [1] 1
#> 
#> $GSE81454$geeq$sScoreAvgPlatformPopularity
#> [1] -0.5
#> 
#> $GSE81454$geeq$sScoreAvgPlatformSize
#> [1] 1
#> 
#> $GSE81454$geeq$sScoreSampleSize
#> [1] 1
#> 
#> $GSE81454$geeq$sScoreRawData
#> [1] 1
#> 
#> $GSE81454$geeq$sScoreMissingValues
#> [1] 1
#> 
#> $GSE81454$geeq$qScoreOutliers
#> [1] 1
#> 
#> $GSE81454$geeq$qScorePlatformsTech
#> [1] 1
#> 
#> $GSE81454$geeq$qScoreReplicates
#> [1] 0
#> 
#> $GSE81454$geeq$qScoreBatchInfo
#> [1] -1
#> 
#> $GSE81454$geeq$qScoreSampleMeanCorrelation
#> [1] 0.9267995
#> 
#> $GSE81454$geeq$qScoreSampleMedianCorrelation
#> [1] 0.9372003
#> 
#> $GSE81454$geeq$qScoreSampleCorrelationVariance
#> [1] 0.001940028
#> 
#> $GSE81454$geeq$noVectors
#> [1] FALSE
#> 
#> $GSE81454$geeq$corrMatIssues
#> [1] 2
#> 
#> $GSE81454$geeq$replicatesIssues
#> [1] 2
#> 
#> $GSE81454$geeq$batchCorrected
#> [1] FALSE
#> 
#> $GSE81454$geeq$publicQualityScore
#> [1] 0.2767429
#> 
#> $GSE81454$geeq$publicSuitabilityScore
#> [1] 0.8125
#> 
#> $GSE81454$geeq$qScorePublicBatchEffect
#> [1] 0
#> 
#> $GSE81454$geeq$qScorePublicBatchConfound
#> [1] 0
#> 
#> $GSE81454$geeq$`_totalInQuery`
#> [1] 0
#> 
#> $GSE81454$geeq$id
#> [1] 7697
#> 
#> 
#> $GSE81454$arrayDesignCount
#> [1] 1
#> 
#> $GSE81454$bioAssayCount
#> [1] 31
#> 
#> $GSE81454$currentUserHasWritePermission
#> [1] FALSE
#> 
#> $GSE81454$currentUserIsOwner
#> [1] FALSE
#> 
#> $GSE81454$externalUri
#> [1] "http://www.ncbi.nlm.nih.gov/geo/"
#> 
#> $GSE81454$isShared
#> [1] FALSE
#> 
#> $GSE81454$securableClass
#> [1] "ubic.gemma.model.expression.experiment.ExpressionExperiment"
#> 
#> $GSE81454$userCanWrite
#> [1] FALSE
#> 
#> $GSE81454$description
#> [1] "Immune profiles were performed retrospectively in highly sensitized kidney transplant candidates Our hypothesis was that baseline differences in immune profiles could help identify candidates that respond to desensitization therapy.\nLast Updated (by provider): Apr 20 2017\nContributors:  Julie Yabu"
#> 
#> $GSE81454$source
#> NULL
#> 
#> $GSE81454$name
#> [1] "HLA sensitization study"
#> 
#> $GSE81454$id
#> [1] 10634
#> 
#> $GSE81454$lastUpdated
#> [1] 1.585531e+12
#> 
#> $GSE81454$troubled
#> [1] FALSE
#> 
#> $GSE81454$lastTroubledEvent
#> NULL
#> 
#> $GSE81454$needsAttention
#> [1] FALSE
#> 
#> $GSE81454$troubleDetails
#> [1] "No trouble details provided."
#> 
#> $GSE81454$lastNoteUpdateEvent
#> NULL
#> 
#> $GSE81454$lastNeedsAttentionEvent
#> NULL
#> 
#> $GSE81454$curationNote
#> NULL
#> 
#> $GSE81454$`_totalInQuery`
#> [1] 1
#> 
#> 

datasetInfo('GSE81454', request = 'platforms')

#> $GPL15207
#> $GPL15207$shortName
#> [1] "GPL15207"
#> 
#> $GPL15207$dateCached
#> NULL
#> 
#> $GPL15207$numProbeSequences
#> NULL
#> 
#> $GPL15207$numProbeAlignments
#> NULL
#> 
#> $GPL15207$numProbesToGenes
#> NULL
#> 
#> $GPL15207$numGenes
#> NULL
#> 
#> $GPL15207$designElementCount
#> NULL
#> 
#> $GPL15207$taxon
#> [1] "human"
#> 
#> $GPL15207$technologyType
#> [1] "ONECOLOR"
#> 
#> $GPL15207$lastGeneMapping
#> NULL
#> 
#> $GPL15207$lastSequenceAnalysis
#> NULL
#> 
#> $GPL15207$lastSequenceUpdate
#> NULL
#> 
#> $GPL15207$blackListed
#> [1] FALSE
#> 
#> $GPL15207$taxonID
#> NULL
#> 
#> $GPL15207$color
#> [1] "ONECOLOR"
#> 
#> $GPL15207$expressionExperimentCount
#> [1] 20
#> 
#> $GPL15207$hasBlatAssociations
#> NULL
#> 
#> $GPL15207$hasGeneAssociations
#> NULL
#> 
#> $GPL15207$hasSequenceAssociations
#> NULL
#> 
#> $GPL15207$isAffymetrixAltCdf
#> [1] FALSE
#> 
#> $GPL15207$isMerged
#> NULL
#> 
#> $GPL15207$isMergee
#> [1] FALSE
#> 
#> $GPL15207$isSubsumed
#> NULL
#> 
#> $GPL15207$isSubsumer
#> NULL
#> 
#> $GPL15207$lastRepeatMask
#> NULL
#> 
#> $GPL15207$description
#> [1] ""
#> 
#> $GPL15207$name
#> [1] "[PrimeView] Affymetrix Human Gene Expression Array"
#> 
#> $GPL15207$lastUpdated
#> [1] 1.575781e+12
#> 
#> $GPL15207$troubled
#> [1] FALSE
#> 
#> $GPL15207$lastTroubledEvent
#> NULL
#> 
#> $GPL15207$needsAttention
#> [1] FALSE
#> 
#> $GPL15207$troubleDetails
#> [1] "No trouble details provided."
#> 
#> $GPL15207$lastNoteUpdateEvent
#> NULL
#> 
#> $GPL15207$lastNeedsAttentionEvent
#> NULL
#> 
#> $GPL15207$curationNote
#> NULL
#> 
#> $GPL15207$`_totalInQuery`
#> [1] 0
#> 
#> $GPL15207$id
#> [1] 783
#> 
#> 

datasetInfo('GSE81454', request='data',filter = FALSE)

#> # A tibble: 49,395 x 37
#>    Probe Sequence GeneSymbol GeneName GemmaId NCBIid `GSE81454_Bioma~
#>    <chr> <chr>    <chr>      <chr>    <chr>   <chr>             <dbl>
#>  1 1171~ 1171510~ NA         NA       NA      NA                 5.57
#>  2 1171~ 1171510~ NA         NA       NA      NA                 5.49
#>  3 1171~ 1171510~ NA         NA       NA      NA                 5.26
#>  4 1171~ 1171510~ TNFAIP8L1  TNF alp~ 383466  126282             6.95
#>  5 1171~ 1171510~ OTOP2      otopetr~ 371285  92736              4.90
#>  6 1171~ 1171510~ C17orf78   chromos~ 418711  284099             3.85
#>  7 1171~ 1171510~ CTAGE6|CT~ CTAGE f~ 425383~ 34030~             5.08
#>  8 1171~ 1171510~ F8A1|F8A3~ coagula~ 176266~ 8263|~             5.22
#>  9 1171~ 1171510~ LINC01098  long in~ 421233  285501             3.67
#> 10 1171~ 1171510~ SAMD7      sterile~ 426827  344658             4.54
#> # ... with 49,385 more rows, and 30 more variables:
#> #   `GSE81454_Biomat_17___BioAssayId=414332Name=patient1258.030[38792]` <dbl>,
#> #   `GSE81454_Biomat_19___BioAssayId=414331Name=patient1258.030[38848]` <dbl>,
#> #   `GSE81454_Biomat_3___BioAssayId=414330Name=patient1258.030[39011]` <dbl>,
#> #   `GSE81454_Biomat_1___BioAssayId=414329Name=patient1258.030[39105]` <dbl>,
#> #   `GSE81454_Biomat_29___BioAssayId=414345Name=patient1258.343[39400]` <dbl>,
#> #   `GSE81454_Biomat_5___BioAssayId=414344Name=patient1258.343[39428]` <dbl>,
#> #   `GSE81454_Biomat_2___BioAssayId=414343Name=patient1258.343[39456]` <dbl>,
#> #   `GSE81454_Biomat_9___BioAssayId=414342Name=patient1258.343[39540]` <dbl>,
#> #   `GSE81454_Biomat_7___BioAssayId=414341Name=patient1258.343[39576]` <dbl>,
#> #   `GSE81454_Biomat_18___BioAssayId=414340Name=patient1258.343[39581]` <dbl>,
#> #   `GSE81454_Biomat_20___BioAssayId=414339Name=patient1258.343[39602]` <dbl>,
#> #   `GSE81454_Biomat_21___BioAssayId=414338Name=patient1258.343[39624]` <dbl>,
#> #   `GSE81454_Biomat_22___BioAssayId=414337Name=patient1258.587[39450]` <dbl>,
#> #   `GSE81454_Biomat_24___BioAssayId=414335Name=patient1258.587[39506]` <dbl>,
#> #   `GSE81454_Biomat_23___BioAssayId=414336Name=patient1258.587[39513]` <dbl>,
#> #   `GSE81454_Biomat_25___BioAssayId=414334Name=patient1258.587[39590]` <dbl>,
#> #   `GSE81454_Biomat_4___BioAssayId=414359Name=patient1258.896[39226]` <dbl>,
#> #   `GSE81454_Biomat_6___BioAssayId=414358Name=patient1258.896[39254]` <dbl>,
#> #   `GSE81454_Biomat_8___BioAssayId=414357Name=patient1258.896[39288]` <dbl>,
#> #   `GSE81454_Biomat_10___BioAssayId=414356Name=patient1258.896[39367]` <dbl>,
#> #   `GSE81454_Biomat_11___BioAssayId=414355Name=patient1258.896[39548]` <dbl>,
#> #   `GSE81454_Biomat_12___BioAssayId=414354Name=patient1258.896[39562]` <dbl>,
#> #   `GSE81454_Biomat_13___BioAssayId=414353Name=patient1258.896[39591]` <dbl>,
#> #   `GSE81454_Biomat_14___BioAssayId=414352Name=patient1258.896[39623]` <dbl>,
#> #   `GSE81454_Biomat_15___BioAssayId=414351Name=patient1258.914[39345]` <dbl>,
#> #   `GSE81454_Biomat_26___BioAssayId=414350Name=patient1258.914[39373]` <dbl>,
#> #   `GSE81454_Biomat_27___BioAssayId=414349Name=patient1258.914[39457]` <dbl>,
#> #   `GSE81454_Biomat_30___BioAssayId=414348Name=patient1258.914[39547]` <dbl>,
#> #   `GSE81454_Biomat_31___BioAssayId=414347Name=patient1258.914[39571]` <dbl>,
#> #   `GSE81454_Biomat_28___BioAssayId=414346Name=patient1258.914[39583]` <dbl>