Prototype function body

This should not be called directly, but is called from the API functions.

.body(
  memoised,
  fname,
  validators,
  endpoint,
  envWhere,
  isFile,
  raw,
  overwrite,
  file,
  async,
  .call
)

Arguments

memoised
Whether or not to memoise results

fname
The function name

validators
Validators for parameters

endpoint
Formatted endpoint URL

envWhere
Environment to evaluate in

isFile
Whether or not the endpoint is expect to return a file

raw
Whether to return JSON (TRUE) or data.table (FALSE)

overwrite
Whether or not to overwrite the file if @param file is specified

file
A filename to save results to

async
Whether or not to run asynchronously

.call
The original function call