expressionSubset

Uses gene expression endpoint to get expression of some genes

expressionSubset(dataset, genes, keepNonSpecific = FALSE,
  consolidate = NULL, memoised = FALSE)

Arguments

dataset
Can either be the dataset ID or its short name (e.g. GSE1234)

genes
A list of identifiers, separated by commas (e.g: 1859, 5728).

Can either be the NCBI ID (1859), Ensembl ID (ENSG00000157540) or official symbol (DYRK1A) of the gene.

NCBI ID is the most efficient (and guaranteed to be unique) identifier.

Official symbol represents a gene homologue for a random taxon, unless used in a specific taxon (see the Taxa Endpoints).

If the gene taxon does not match the taxon of the given datasets, expression levels for that gene will be missing from the response

You can combine various identifiers in one query, but any invalid identifier will cause the call to yield an error. Duplicate identifiers of the same gene will result in duplicates in the response.

keepNonSpecific
Defaults to FALSE.

If set to false, the response will only include elements that map exclusively to each queried gene

If set to true, the response will include all elements that map to each queried gene, even if they also map to other genes.

consolidate
Optional. Defaults no NULL. Optional, defaults to empty.

Whether genes with multiple elements should consolidate the information. If the 'keepNonSpecific' parameter is set to true, then all gene non-specific vectors are excluded from the chosen procedure.

The options are:

NULL:list all vectors separately.

"pickmax": only return the vector that has the highest expression (mean over all its bioAssays).

"pickvar": only return the vector with highest variance of expression across its bioAssays

"average": create a new vector that will average the bioAssay values from all vectors