Dataset annotations

Retrieves the annotations for the given dataset

getDatasetAnnotations(
  dataset = NA_character_,
  raw = getOption("gemma.raw", FALSE),
  async = getOption("gemma.async", FALSE),
  memoised = getOption("gemma.memoise", FALSE),
  file = getOption("gemma.file", NA_character_),
  overwrite = getOption("gemma.overwrite", FALSE)
)

Arguments

dataset
Required, part of the URL path. Can either be the dataset ID or its short name (e.g. GSE1234). Retrieval by ID is more efficient. Only datasets that user has access to will be available

raw
TRUE to receive results as-is from Gemma, or FALSE to enable parsing.

async
TRUE to run the API query on a separate worker, or FALSE to run synchronously. See the async package for details.

memoised
Whether or not to cache results so future requests for the same data will be faster. Use forgetGemmaMemoised to clear the cache.

file
The name of a file to save the results to, or NULL to not write results to a file. If raw == TRUE, the output will be a JSON file. Otherwise, it will be a RDS file.

overwrite
Whether or not to overwrite if a file exists at the specified filename.

Value

An array of annotations (annotation value objects) attached to the given dataset. A 404 error if the given identifier does not map to any object.

Examples

getDatasetAnnotations("GSE2018")

#>       class.Type    class.Name                            class.URL
#> 1: ExperimentTag organism part http://www.ebi.ac.uk/efo/EFO_0000635
#> 2: ExperimentTag     cell type http://www.ebi.ac.uk/efo/EFO_0000324
#> 3:   FactorValue       disease http://www.ebi.ac.uk/efo/EFO_0000408
#>    evidence.Code                       term.Name
#> 1:            IC                            lung
#> 2:            IC     respiratory epithelial cell
#> 3:               Transplanted Organ Complication
#>                                         term.URL
#> 1: http://purl.obolibrary.org/obo/UBERON_0002048
#> 2:     http://purl.obolibrary.org/obo/CL_0002368
#> 3:       http://purl.obolibrary.org/obo/DOID_556