Dataset data

Retrieves the data for the given dataset

getDatasetData(
  dataset = NA_character_,
  filter = FALSE,
  raw = getOption("gemma.raw", FALSE),
  async = getOption("gemma.async", FALSE),
  memoised = getOption("gemma.memoise", FALSE),
  file = getOption("gemma.file", NA_character_),
  overwrite = getOption("gemma.overwrite", FALSE)
)

Arguments

dataset
Required, part of the URL path. Can either be the dataset ID or its short name (e.g. GSE1234). Retrieval by ID is more efficient. Only datasets that user has access to will be available

filter
Optional, defaults to empty. Filtering can be done on any* property or nested property that the appropriate object class defines or inherits (and that is mapped by hibernate). [These do not correspond to the properties of the objects returned by the API calls.].description-imp Class definitions:

Datasets: javaDoc gitHub

Platforms: javaDoc gitHub E.g: curationDetails or curationDetails.lastTroubledEvent.date.

Any property of a supported type. Currently supported types are:

String - property of String type, required value can be any String.

Number - any Number implementation. Required value must be a string parseable to the specific Number type.

Boolean - required value will be parsed to true only if the string matches 'true', ignoring case. Accepted operator keywords are:

'=' - equality

'!=' - non-equality

'<' - smaller than

'>' - larger than

'<=' - smaller or equal

'=>' - larger or equal

'like' - similar string, effectively means 'contains', translates to the sql 'LIKE' operator (given value will be surrounded by % signs) Multiple filters can be chained using AND and OR keywords. Leave space between the keywords and the previous/next word! E.g: ?filter=property1 < value1 AND property2 like value2 If chained filters are mixed conjunctions and disjunctions, the query must be in conjunctive normal form (CNF). Parentheses are not necessary

every AND keyword separates blocks of disjunctions. Example: ?filter=p1 = v1 OR p1 != v2 AND p2 <= v2 AND p3 > v3 OR p3 < v4 Above query will translate to: (p1 = v1 OR p1 != v2) AND (p2 <= v2) AND (p3 > v3 OR p3 < v4;) Breaking the CNF results in an error. Filter curationDetails.troubled will be ignored if user is not an administrator.

raw
TRUE to receive results as-is from Gemma, or FALSE to enable parsing.

async
TRUE to run the API query on a separate worker, or FALSE to run synchronously. See the async package for details.

memoised
Whether or not to cache results so future requests for the same data will be faster. Use forgetGemmaMemoised to clear the cache.

file
The name of a file to save the results to, or NULL to not write results to a file. If raw == TRUE, the output will be a JSON file. Otherwise, it will be a RDS file.

overwrite
Whether or not to overwrite if a file exists at the specified filename.

Value

The data file for the given dataset. A 404 error if the given identifier does not map to any object.

Examples

getDatasetData("GSE2018")

#>                         Probe                         Sequence  GeneSymbol
#>     1:              1007_s_at              1007_s_at_collapsed        DDR1
#>     2:                1053_at                1053_at_collapsed        RFC2
#>     3:                 117_at                 117_at_collapsed HSPA6|HSPA7
#>     4:                 121_at                 121_at_collapsed        PAX8
#>     5:              1255_g_at              1255_g_at_collapsed      GUCA1A
#>    ---                                                                    
#> 22279: AFFX-r2-Hs28SrRNA-3_at AFFX-r2-Hs28SrRNA-3_at_collapsed            
#> 22280: AFFX-r2-Hs28SrRNA-5_at AFFX-r2-Hs28SrRNA-5_at_collapsed            
#> 22281: AFFX-r2-Hs28SrRNA-M_at AFFX-r2-Hs28SrRNA-M_at_collapsed            
#> 22282:    AFFX-r2-P1-cre-3_at    AFFX-r2-P1-cre-3_at_collapsed            
#> 22283:    AFFX-r2-P1-cre-5_at    AFFX-r2-P1-cre-5_at_collapsed            
#>                                                                                         GeneName
#>     1:                                               discoidin domain receptor tyrosine kinase 1
#>     2:                                                            replication factor C subunit 2
#>     3: heat shock protein family A (Hsp70) member 6|heat shock protein family A (Hsp70) member 7
#>     4:                                                                              paired box 8
#>     5:                                                            guanylate cyclase activator 1A
#>    ---                                                                                          
#> 22279:                                                                                          
#> 22280:                                                                                          
#> 22281:                                                                                          
#> 22282:                                                                                          
#> 22283:                                                                                          
#>            GemmaId    NCBIid
#>     1:       16908       780
#>     2:      139878      5982
#>     3: 73420|73442 3310|3311
#>     4:      173107      7849
#>     5:       58787      2978
#>    ---                      
#> 22279:                      
#> 22280:                      
#> 22281:                      
#> 22282:                      
#> 22283:                      
#>        GSE2018_bioMaterial_6.GSM36428___BioAssayId=29Name=BAL_13_A0B0
#>     1:                                                       7.664396
#>     2:                                                       6.818718
#>     3:                                                       7.702311
#>     4:                                                       8.872670
#>     5:                                                       4.023544
#>    ---                                                               
#> 22279:                                                       5.269825
#> 22280:                                                       4.032411
#> 22281:                                                       5.033840
#> 22282:                                                      12.634298
#> 22283:                                                      12.589158
#>        GSE2018_bioMaterial_28.GSM36450___BioAssayId=7Name=BAL_44b_A0B0
#>     1:                                                        7.555234
#>     2:                                                        6.817860
#>     3:                                                        7.736164
#>     4:                                                        8.728661
#>     5:                                                        4.022874
#>    ---                                                                
#> 22279:                                                        4.883055
#> 22280:                                                        4.443605
#> 22281:                                                        5.321434
#> 22282:                                                       11.756798
#> 22283:                                                       11.830940
#>        GSE2018_bioMaterial_24.GSM36446___BioAssayId=11Name=BAL_41_A1B0
#>     1:                                                        7.898749
#>     2:                                                        6.428031
#>     3:                                                        8.299871
#>     4:                                                        9.433338
#>     5:                                                        4.266776
#>    ---                                                                
#> 22279:                                                        6.307157
#> 22280:                                                        4.487075
#> 22281:                                                        5.462065
#> 22282:                                                       11.946831
#> 22283:                                                       11.781727
#>        GSE2018_bioMaterial_19.GSM36441___BioAssayId=16Name=BAL_31_A0B1
#>     1:                                                        7.482914
#>     2:                                                        6.873047
#>     3:                                                        7.714629
#>     4:                                                        9.106445
#>     5:                                                        4.127223
#>    ---                                                                
#> 22279:                                                        6.406686
#> 22280:                                                        4.551892
#> 22281:                                                        5.793712
#> 22282:                                                       12.617424
#> 22283:                                                       12.318279
#>        GSE2018_bioMaterial_5.GSM36427___BioAssayId=30Name=BAL_12_A1B0
#>     1:                                                       7.606950
#>     2:                                                       7.081941
#>     3:                                                       7.497553
#>     4:                                                       9.130426
#>     5:                                                       4.130228
#>    ---                                                               
#> 22279:                                                       6.242087
#> 22280:                                                       4.349244
#> 22281:                                                       5.663534
#> 22282:                                                      12.376739
#> 22283:                                                      12.476188
#>        GSE2018_bioMaterial_3.GSM36425___BioAssayId=32Name=BAL_15b_A0B0
#>     1:                                                        7.675714
#>     2:                                                        6.770301
#>     3:                                                        8.433980
#>     4:                                                        9.062922
#>     5:                                                        3.991846
#>    ---                                                                
#> 22279:                                                        5.699829
#> 22280:                                                        4.661939
#> 22281:                                                        5.803358
#> 22282:                                                       11.258021
#> 22283:                                                       11.643198
#>        GSE2018_bioMaterial_23.GSM36445___BioAssayId=12Name=BAL_40_A0B0
#>     1:                                                        7.568318
#>     2:                                                        6.754742
#>     3:                                                        7.307702
#>     4:                                                        9.126554
#>     5:                                                        4.057203
#>    ---                                                                
#> 22279:                                                        5.119069
#> 22280:                                                        4.238217
#> 22281:                                                        5.372956
#> 22282:                                                       12.462382
#> 22283:                                                       12.634298
#>        GSE2018_bioMaterial_33.GSM36455___BioAssayId=2Name=BAL_47b_A0B1
#>     1:                                                        7.605180
#>     2:                                                        6.785765
#>     3:                                                        7.800827
#>     4:                                                        9.024894
#>     5:                                                        4.179423
#>    ---                                                                
#> 22279:                                                        6.658047
#> 22280:                                                        4.462518
#> 22281:                                                        5.477958
#> 22282:                                                       12.871473
#> 22283:                                                       12.909754
#>        GSE2018_bioMaterial_34.GSM36456___BioAssayId=1Name=BAL_47c_A0B1
#>     1:                                                        7.629114
#>     2:                                                        6.910238
#>     3:                                                        7.474227
#>     4:                                                        9.173378
#>     5:                                                        3.955497
#>    ---                                                                
#> 22279:                                                        5.768701
#> 22280:                                                        4.355739
#> 22281:                                                        5.260586
#> 22282:                                                       12.033990
#> 22283:                                                       12.088908
#>        GSE2018_bioMaterial_29.GSM36451___BioAssayId=6Name=BAL_45a_A1B0
#>     1:                                                        7.446346
#>     2:                                                        6.803677
#>     3:                                                        7.623731
#>     4:                                                        9.293806
#>     5:                                                        4.034630
#>    ---                                                                
#> 22279:                                                        5.107271
#> 22280:                                                        4.260343
#> 22281:                                                        5.040337
#> 22282:                                                       11.272546
#> 22283:                                                       11.459026
#>        GSE2018_bioMaterial_32.GSM36454___BioAssayId=3Name=BAL_47a_A0B0
#>     1:                                                        7.787959
#>     2:                                                        6.440673
#>     3:                                                        7.844755
#>     4:                                                        9.040855
#>     5:                                                        3.909490
#>    ---                                                                
#> 22279:                                                        5.371702
#> 22280:                                                        4.276494
#> 22281:                                                        5.166088
#> 22282:                                                       12.462382
#> 22283:                                                       12.656571
#>        GSE2018_bioMaterial_27.GSM36449___BioAssayId=8Name=BAL_44a_A0B0
#>     1:                                                        7.734734
#>     2:                                                        6.940845
#>     3:                                                        7.603600
#>     4:                                                        9.248298
#>     5:                                                        4.004146
#>    ---                                                                
#> 22279:                                                        5.183981
#> 22280:                                                        4.627294
#> 22281:                                                        6.134356
#> 22282:                                                       12.257533
#> 22283:                                                       12.342556
#>        GSE2018_bioMaterial_21.GSM36443___BioAssayId=14Name=BAL_39_A0B1
#>     1:                                                        7.270490
#>     2:                                                        6.808251
#>     3:                                                        7.328971
#>     4:                                                        8.881946
#>     5:                                                        4.018348
#>    ---                                                                
#> 22279:                                                        5.413069
#> 22280:                                                        4.337369
#> 22281:                                                        5.647816
#> 22282:                                                       12.728011
#> 22283:                                                       12.855262
#>        GSE2018_bioMaterial_20.GSM36442___BioAssayId=15Name=BAL_33_A0B1
#>     1:                                                        7.428734
#>     2:                                                        7.139178
#>     3:                                                        7.568634
#>     4:                                                        8.946895
#>     5:                                                        4.025862
#>    ---                                                                
#> 22279:                                                        5.689970
#> 22280:                                                        4.311341
#> 22281:                                                        5.397744
#> 22282:                                                       11.887863
#> 22283:                                                       11.810861
#>        GSE2018_bioMaterial_17.GSM36439___BioAssayId=18Name=BAL_29_A1B0
#>     1:                                                        7.905805
#>     2:                                                        6.433760
#>     3:                                                        7.794066
#>     4:                                                        9.592714
#>     5:                                                        4.128827
#>    ---                                                                
#> 22279:                                                        5.304832
#> 22280:                                                        4.163611
#> 22281:                                                        5.566520
#> 22282:                                                       12.742010
#> 22283:                                                       12.880875
#>        GSE2018_bioMaterial_13.GSM36435___BioAssayId=22Name=BAL_25_A0B0
#>     1:                                                        7.565792
#>     2:                                                        6.887287
#>     3:                                                        7.444417
#>     4:                                                        8.923390
#>     5:                                                        3.979449
#>    ---                                                                
#> 22279:                                                        4.986458
#> 22280:                                                        4.339306
#> 22281:                                                        5.449715
#> 22282:                                                       12.323225
#> 22283:                                                       12.294077
#>        GSE2018_bioMaterial_15.GSM36437___BioAssayId=20Name=BAL_27a_A0B0
#>     1:                                                         7.912942
#>     2:                                                         6.975750
#>     3:                                                         7.775969
#>     4:                                                         9.249823
#>     5:                                                         4.027976
#>    ---                                                                 
#> 22279:                                                         5.660599
#> 22280:                                                         4.354635
#> 22281:                                                         5.502226
#> 22282:                                                        12.363355
#> 22283:                                                        12.294077
#>        GSE2018_bioMaterial_4.GSM36426___BioAssayId=31Name=BAL_1_A0B0
#>     1:                                                      7.520084
#>     2:                                                      7.158184
#>     3:                                                      7.843651
#>     4:                                                      9.057032
#>     5:                                                      4.168197
#>    ---                                                              
#> 22279:                                                      5.754021
#> 22280:                                                      4.305040
#> 22281:                                                      5.673235
#> 22282:                                                     12.419389
#> 22283:                                                     12.401875
#>        GSE2018_bioMaterial_2.GSM36424___BioAssayId=33Name=BAL_15a_A0B1
#>     1:                                                        7.633544
#>     2:                                                        6.906474
#>     3:                                                        7.689818
#>     4:                                                        8.764082
#>     5:                                                        4.025454
#>    ---                                                                
#> 22279:                                                        5.472515
#> 22280:                                                        4.360437
#> 22281:                                                        5.541409
#> 22282:                                                       12.462382
#> 22283:                                                       12.383058
#>        GSE2018_bioMaterial_26.GSM36448___BioAssayId=9Name=BAL_43_A0B0
#>     1:                                                       7.333711
#>     2:                                                       6.966837
#>     3:                                                       7.462743
#>     4:                                                       9.090870
#>     5:                                                       4.002775
#>    ---                                                               
#> 22279:                                                       5.713751
#> 22280:                                                       4.326695
#> 22281:                                                       5.575244
#> 22282:                                                      12.728011
#> 22283:                                                      12.708045
#>        GSE2018_bioMaterial_25.GSM36447___BioAssayId=10Name=BAL_42_A0B0
#>     1:                                                        7.893288
#>     2:                                                        6.649465
#>     3:                                                        8.579609
#>     4:                                                        8.930581
#>     5:                                                        4.113324
#>    ---                                                                
#> 22279:                                                        5.882000
#> 22280:                                                        4.201591
#> 22281:                                                        5.634897
#> 22282:                                                       12.758545
#> 22283:                                                       12.728011
#>        GSE2018_bioMaterial_22.GSM36444___BioAssayId=13Name=BAL_4_A0B0
#>     1:                                                       7.474848
#>     2:                                                       6.947905
#>     3:                                                       7.741327
#>     4:                                                       9.119985
#>     5:                                                       4.037982
#>    ---                                                               
#> 22279:                                                       5.745562
#> 22280:                                                       4.332832
#> 22281:                                                       5.390216
#> 22282:                                                      12.582693
#> 22283:                                                      12.742010
#>        GSE2018_bioMaterial_18.GSM36440___BioAssayId=17Name=BAL_3_A1B0
#>     1:                                                       7.543944
#>     2:                                                       6.895257
#>     3:                                                       7.542451
#>     4:                                                       8.873311
#>     5:                                                       3.861244
#>    ---                                                               
#> 22279:                                                       6.067514
#> 22280:                                                       4.337072
#> 22281:                                                       5.559285
#> 22282:                                                      12.595888
#> 22283:                                                      12.750864
#>        GSE2018_bioMaterial_8.GSM36430___BioAssayId=27Name=BAL_18_A1B0
#>     1:                                                       7.407762
#>     2:                                                       6.780074
#>     3:                                                       7.926313
#>     4:                                                       9.154426
#>     5:                                                       4.054710
#>    ---                                                               
#> 22279:                                                       7.299089
#> 22280:                                                       4.395083
#> 22281:                                                       5.974870
#> 22282:                                                      12.841431
#> 22283:                                                      12.862965
#>        GSE2018_bioMaterial_7.GSM36429___BioAssayId=28Name=BAL_17_A0B0
#>     1:                                                       7.229377
#>     2:                                                       6.963063
#>     3:                                                       7.972501
#>     4:                                                       8.806005
#>     5:                                                       4.024187
#>    ---                                                               
#> 22279:                                                       6.076631
#> 22280:                                                       4.215431
#> 22281:                                                       5.704353
#> 22282:                                                      12.516955
#> 22283:                                                      12.449719
#>        GSE2018_bioMaterial_12.GSM36434___BioAssayId=23Name=BAL_21_A0B0
#>     1:                                                        7.867851
#>     2:                                                        6.615844
#>     3:                                                        7.421375
#>     4:                                                        9.553362
#>     5:                                                        4.137057
#>    ---                                                                
#> 22279:                                                        7.957031
#> 22280:                                                        5.204943
#> 22281:                                                        7.226784
#> 22282:                                                       12.589158
#> 22283:                                                       12.909754
#>        GSE2018_bioMaterial_9.GSM36431___BioAssayId=26Name=BAL_19_A0B0
#>     1:                                                       7.821547
#>     2:                                                       6.709687
#>     3:                                                       7.825478
#>     4:                                                       9.574804
#>     5:                                                       3.989322
#>    ---                                                               
#> 22279:                                                       8.356067
#> 22280:                                                       5.687296
#> 22281:                                                       7.929119
#> 22282:                                                      12.671734
#> 22283:                                                      12.909754
#>        GSE2018_bioMaterial_30.GSM36452___BioAssayId=5Name=BAL_45b_A2B0
#>     1:                                                        7.782975
#>     2:                                                        6.593049
#>     3:                                                        7.932626
#>     4:                                                        8.999097
#>     5:                                                        3.944738
#>    ---                                                                
#> 22279:                                                        5.328197
#> 22280:                                                        4.119304
#> 22281:                                                        5.046239
#> 22282:                                                       12.123510
#> 22283:                                                       12.093898
#>        GSE2018_bioMaterial_14.GSM36436___BioAssayId=21Name=BAL_26_A2B0
#>     1:                                                        7.257932
#>     2:                                                        6.940845
#>     3:                                                        7.969855
#>     4:                                                        8.656306
#>     5:                                                        4.013794
#>    ---                                                                
#> 22279:                                                        4.948480
#> 22280:                                                        4.171280
#> 22281:                                                        5.167862
#> 22282:                                                       12.088908
#> 22283:                                                       12.211794
#>        GSE2018_bioMaterial_10.GSM36432___BioAssayId=25Name=BAL_20a_A1B1
#>     1:                                                              NaN
#>     2:                                                              NaN
#>     3:                                                              NaN
#>     4:                                                              NaN
#>     5:                                                              NaN
#>    ---                                                                 
#> 22279:                                                              NaN
#> 22280:                                                              NaN
#> 22281:                                                              NaN
#> 22282:                                                              NaN
#> 22283:                                                              NaN
#>        GSE2018_bioMaterial_11.GSM36433___BioAssayId=24Name=BAL_20b_A1B1
#>     1:                                                         7.680408
#>     2:                                                         6.650542
#>     3:                                                         7.738031
#>     4:                                                         9.239082
#>     5:                                                         4.030433
#>    ---                                                                 
#> 22279:                                                         6.952453
#> 22280:                                                         4.250091
#> 22281:                                                         5.876946
#> 22282:                                                        12.624101
#> 22283:                                                        12.449719
#>        GSE2018_bioMaterial_1.GSM36423___BioAssayId=34Name=BAL_37_A1B1
#>     1:                                                       7.466471
#>     2:                                                       6.858354
#>     3:                                                       7.689818
#>     4:                                                       9.055730
#>     5:                                                       4.062315
#>    ---                                                               
#> 22279:                                                       5.377125
#> 22280:                                                       4.253467
#> 22281:                                                       5.369570
#> 22282:                                                      11.975114
#> 22283:                                                      12.088908
#>        GSE2018_bioMaterial_31.GSM36453___BioAssayId=4Name=BAL_46_A2B0
#>     1:                                                       7.932626
#>     2:                                                       6.537636
#>     3:                                                       7.809553
#>     4:                                                      10.078527
#>     5:                                                       3.859072
#>    ---                                                               
#> 22279:                                                       8.848175
#> 22280:                                                       4.782150
#> 22281:                                                       7.281428
#> 22282:                                                      13.047083
#> 22283:                                                      13.239051
#>        GSE2018_bioMaterial_16.GSM36438___BioAssayId=19Name=BAL_27b_A2B1
#>     1:                                                         7.699664
#>     2:                                                         6.845171
#>     3:                                                         7.555798
#>     4:                                                         9.141695
#>     5:                                                         4.121438
#>    ---                                                                 
#> 22279:                                                         5.789951
#> 22280:                                                         4.440793
#> 22281:                                                         5.875893
#> 22282:                                                        12.443970
#> 22283:                                                        12.671734