Dataset design

Retrieves the design for the given dataset

getDatasetDesign(
  dataset = NA_character_,
  raw = getOption("gemma.raw", FALSE),
  async = getOption("gemma.async", FALSE),
  memoised = getOption("gemma.memoise", FALSE),
  file = getOption("gemma.file", NA_character_),
  overwrite = getOption("gemma.overwrite", FALSE)
)

Arguments

dataset
Required, part of the URL path. Can either be the dataset ID or its short name (e.g. GSE1234). Retrieval by ID is more efficient. Only datasets that user has access to will be available

raw
TRUE to receive results as-is from Gemma, or FALSE to enable parsing.

async
TRUE to run the API query on a separate worker, or FALSE to run synchronously. See the async package for details.

memoised
Whether or not to cache results so future requests for the same data will be faster. Use forgetGemmaMemoised to clear the cache.

file
The name of a file to save the results to, or NULL to not write results to a file. If raw == TRUE, the output will be a JSON file. Otherwise, it will be a RDS file.

overwrite
Whether or not to overwrite if a file exists at the specified filename.

Value

The design file for the given dataset. A 404 error if the given identifier does not map to any object.

Examples

getDatasetDesign("GSE2018")

#>                                                             Bioassay ExternalID
#>  1:  GSE2018_bioMaterial_34.GSM36456___BioAssayId=1Name=BAL_47c_A0B1   GSM36456
#>  2:  GSE2018_bioMaterial_33.GSM36455___BioAssayId=2Name=BAL_47b_A0B1   GSM36455
#>  3:  GSE2018_bioMaterial_32.GSM36454___BioAssayId=3Name=BAL_47a_A0B0   GSM36454
#>  4:   GSE2018_bioMaterial_31.GSM36453___BioAssayId=4Name=BAL_46_A2B0   GSM36453
#>  5:  GSE2018_bioMaterial_30.GSM36452___BioAssayId=5Name=BAL_45b_A2B0   GSM36452
#>  6:  GSE2018_bioMaterial_29.GSM36451___BioAssayId=6Name=BAL_45a_A1B0   GSM36451
#>  7:  GSE2018_bioMaterial_28.GSM36450___BioAssayId=7Name=BAL_44b_A0B0   GSM36450
#>  8:  GSE2018_bioMaterial_27.GSM36449___BioAssayId=8Name=BAL_44a_A0B0   GSM36449
#>  9:   GSE2018_bioMaterial_26.GSM36448___BioAssayId=9Name=BAL_43_A0B0   GSM36448
#> 10:  GSE2018_bioMaterial_25.GSM36447___BioAssayId=10Name=BAL_42_A0B0   GSM36447
#> 11:  GSE2018_bioMaterial_24.GSM36446___BioAssayId=11Name=BAL_41_A1B0   GSM36446
#> 12:  GSE2018_bioMaterial_23.GSM36445___BioAssayId=12Name=BAL_40_A0B0   GSM36445
#> 13:   GSE2018_bioMaterial_22.GSM36444___BioAssayId=13Name=BAL_4_A0B0   GSM36444
#> 14:  GSE2018_bioMaterial_21.GSM36443___BioAssayId=14Name=BAL_39_A0B1   GSM36443
#> 15:  GSE2018_bioMaterial_20.GSM36442___BioAssayId=15Name=BAL_33_A0B1   GSM36442
#> 16:  GSE2018_bioMaterial_19.GSM36441___BioAssayId=16Name=BAL_31_A0B1   GSM36441
#> 17:   GSE2018_bioMaterial_18.GSM36440___BioAssayId=17Name=BAL_3_A1B0   GSM36440
#> 18:  GSE2018_bioMaterial_17.GSM36439___BioAssayId=18Name=BAL_29_A1B0   GSM36439
#> 19: GSE2018_bioMaterial_16.GSM36438___BioAssayId=19Name=BAL_27b_A2B1   GSM36438
#> 20: GSE2018_bioMaterial_15.GSM36437___BioAssayId=20Name=BAL_27a_A0B0   GSM36437
#> 21:  GSE2018_bioMaterial_14.GSM36436___BioAssayId=21Name=BAL_26_A2B0   GSM36436
#> 22:  GSE2018_bioMaterial_13.GSM36435___BioAssayId=22Name=BAL_25_A0B0   GSM36435
#> 23:  GSE2018_bioMaterial_12.GSM36434___BioAssayId=23Name=BAL_21_A0B0   GSM36434
#> 24: GSE2018_bioMaterial_11.GSM36433___BioAssayId=24Name=BAL_20b_A1B1   GSM36433
#> 25: GSE2018_bioMaterial_10.GSM36432___BioAssayId=25Name=BAL_20a_A1B1   GSM36432
#> 26:   GSE2018_bioMaterial_9.GSM36431___BioAssayId=26Name=BAL_19_A0B0   GSM36431
#> 27:   GSE2018_bioMaterial_8.GSM36430___BioAssayId=27Name=BAL_18_A1B0   GSM36430
#> 28:   GSE2018_bioMaterial_7.GSM36429___BioAssayId=28Name=BAL_17_A0B0   GSM36429
#> 29:   GSE2018_bioMaterial_6.GSM36428___BioAssayId=29Name=BAL_13_A0B0   GSM36428
#> 30:   GSE2018_bioMaterial_5.GSM36427___BioAssayId=30Name=BAL_12_A1B0   GSM36427
#> 31:    GSE2018_bioMaterial_4.GSM36426___BioAssayId=31Name=BAL_1_A0B0   GSM36426
#> 32:  GSE2018_bioMaterial_3.GSM36425___BioAssayId=32Name=BAL_15b_A0B0   GSM36425
#> 33:  GSE2018_bioMaterial_2.GSM36424___BioAssayId=33Name=BAL_15a_A0B1   GSM36424
#> 34:   GSE2018_bioMaterial_1.GSM36423___BioAssayId=34Name=BAL_37_A1B1   GSM36423
#>                                                             Bioassay ExternalID
#>                                            diseaseState             batch
#>  1:                              reference_subject_role  Batch_06_7/24/02
#>  2:                              reference_subject_role  Batch_07_10/2/02
#>  3:                              reference_subject_role Batch_09_10/15/02
#>  4: Transplanted_Organ_Complication_|_Acute_rejection_|  Batch_01_1/31/02
#>  5: Transplanted_Organ_Complication_|_Acute_rejection_|  Batch_03_5/22/02
#>  6:                              reference_subject_role  Batch_06_7/24/02
#>  7:                              reference_subject_role   Batch_02_4/3/02
#>  8:                              reference_subject_role Batch_09_10/15/02
#>  9:                              reference_subject_role Batch_10_10/25/02
#> 10:                              reference_subject_role Batch_12_11/27/02
#> 11:                              reference_subject_role   Batch_02_4/3/02
#> 12:                              reference_subject_role  Batch_04_5/23/02
#> 13:                              reference_subject_role Batch_12_11/27/02
#> 14:                              reference_subject_role Batch_09_10/15/02
#> 15:                              reference_subject_role Batch_09_10/15/02
#> 16:                              reference_subject_role   Batch_02_4/3/02
#> 17:                              reference_subject_role Batch_12_11/27/02
#> 18:                              reference_subject_role  Batch_08_10/4/02
#> 19: Transplanted_Organ_Complication_|_Acute_rejection_|  Batch_01_1/31/02
#> 20:                              reference_subject_role  Batch_11_11/7/02
#> 21: Transplanted_Organ_Complication_|_Acute_rejection_|  Batch_04_5/23/02
#> 22:                              reference_subject_role  Batch_08_10/4/02
#> 23:                              reference_subject_role  Batch_01_1/31/02
#> 24: Transplanted_Organ_Complication_|_Acute_rejection_| Batch_09_10/15/02
#> 25: Transplanted_Organ_Complication_|_Acute_rejection_|  Batch_07_10/2/02
#> 26:                              reference_subject_role  Batch_01_1/31/02
#> 27:                              reference_subject_role Batch_12_11/27/02
#> 28:                              reference_subject_role Batch_12_11/27/02
#> 29:                              reference_subject_role  Batch_03_5/22/02
#> 30:                              reference_subject_role  Batch_05_5/31/02
#> 31:                              reference_subject_role  Batch_11_11/7/02
#> 32:                              reference_subject_role  Batch_05_5/31/02
#> 33:                              reference_subject_role  Batch_11_11/7/02
#> 34: Transplanted_Organ_Complication_|_Acute_rejection_| Batch_10_10/25/02
#>                                            diseaseState             batch