Datasets pca component expression levels

Retrieves the expression levels most correlated with the given principal component

getDatasetPCA(
  datasets = NA_character_,
  component = 1L,
  limit = 100L,
  keepNonSpecific = FALSE,
  consolidate = NA_character_,
  raw = getOption("gemma.raw", FALSE),
  async = getOption("gemma.async", FALSE),
  memoised = getOption("gemma.memoise", FALSE),
  file = getOption("gemma.file", NA_character_),
  overwrite = getOption("gemma.overwrite", FALSE)
)

Arguments

datasets
Optional, defaults to empty. Limits the result to entities with given identifiers. A list of identifiers, separated by commas (e.g: GSE2871,GSE2869,GSE2868). Identifiers can either be the Dataset ID or its short name. Retrieval by ID is more efficient. Only datasets that user has access to will be available. Do not combine different identifiers in one query.

component
Required, defaults to empty. The pca component to limit the results to.

limit
Optional, defaults to 20. Limits the result to specified amount of objects. Use 0 for no limit.

keepNonSpecific
Optional, defaults to false. If set to false, the response will only include elements that map exclusively to each queried gene If set to true, the response will include all elements that map to each queried gene, even if they also map to other genes.

consolidate
Optional, defaults to empty. What action to take when there is more than one element per gene in a dataset. The choices are:

[empty] - will list all vectors separately

pickmax - only return the vector that has the highest expression (mean over all its bioAssays)

pickvar - only return the vector with highest variance of expression across its bioAssays

average - create a new vector that will average the bioAssay values from all vectors

raw
TRUE to receive results as-is from Gemma, or FALSE to enable parsing.

async
TRUE to run the API query on a separate worker, or FALSE to run synchronously. See the async package for details.

memoised
Whether or not to cache results so future requests for the same data will be faster. Use forgetGemmaMemoised to clear the cache.

file
The name of a file to save the results to, or NULL to not write results to a file. If raw == TRUE, the output will be a JSON file. Otherwise, it will be a RDS file.

overwrite
Whether or not to overwrite if a file exists at the specified filename.

Value

The expression levels for each given experiment (experiment expression levels value object) of genes that are most correlated with the given principal component. A 404 error if the given identifier does not map to any object.

Examples

dat <- getDatasetPCA("GSE2018")
str(dat$expr)

#> List of 1
#>  $ :Classes ‘data.table’ and 'data.frame':	96 obs. of  37 variables:
#>   ..$ gene.ID     : int [1:96] 57404 25978 54700 23560 11179 23074 84060 25803 113251 8087 ...
#>   ..$ gene.Symbol : chr [1:96] "CYP20A1" "CHMP2B" "RRN3" "GTPBP4" ...
#>   ..$ probe       : chr [1:96] "219565_at" "202536_at" "222204_s_at" "218239_s_at" ...
#>   ..$ BAL_21_A0B0 : num [1:96] 5.78 6.88 5.45 6.34 5.53 ...
#>   ..$ BAL_27b_A2B1: num [1:96] 5.9 8.06 6.27 6.71 6.45 ...
#>   ..$ BAL_45b_A2B0: num [1:96] 6.09 7.56 5.65 6.62 6.31 ...
#>   ..$ BAL_13_A0B0 : num [1:96] 6.04 8.05 5.98 6.63 6.71 ...
#>   ..$ BAL_20b_A1B1: num [1:96] 5.68 7.84 5.35 6.5 6.48 ...
#>   ..$ BAL_19_A0B0 : num [1:96] 5.26 7.21 5.08 6.36 5.21 ...
#>   ..$ BAL_47b_A0B1: num [1:96] 6.01 7.9 5.61 6.78 6.65 ...
#>   ..$ BAL_43_A0B0 : num [1:96] 6.4 8.24 5.64 6.75 6.51 ...
#>   ..$ BAL_40_A0B0 : num [1:96] 6.01 8.44 5.94 6.75 6.81 ...
#>   ..$ BAL_33_A0B1 : num [1:96] 6.17 7.95 5.62 6.78 6.56 ...
#>   ..$ BAL_39_A0B1 : num [1:96] 6.36 8.59 6.58 6.97 7.33 ...
#>   ..$ BAL_1_A0B0  : num [1:96] 5.89 7.83 5.84 6.57 6.56 ...
#>   ..$ BAL_47a_A0B0: num [1:96] 5.75 7.54 5.33 6.39 5.92 ...
#>   ..$ BAL_27a_A0B0: num [1:96] 5.86 7.91 5.51 6.54 6.59 ...
#>   ..$ BAL_4_A0B0  : num [1:96] 6.14 7.94 5.78 6.72 6.71 ...
#>   ..$ BAL_15b_A0B0: num [1:96] 5.44 7.24 5.25 6.43 5.87 ...
#>   ..$ BAL_41_A1B0 : num [1:96] 5.36 7.36 5.21 6.37 6.06 ...
#>   ..$ BAL_25_A0B0 : num [1:96] 6.17 7.88 5.88 6.81 6.47 ...
#>   ..$ BAL_44a_A0B0: num [1:96] 5.61 7.33 4.79 6.11 5.85 ...
#>   ..$ BAL_15a_A0B1: num [1:96] 5.87 7.88 5.62 6.66 6.27 ...
#>   ..$ BAL_47c_A0B1: num [1:96] 5.87 7.38 4.99 6.37 6.18 ...
#>   ..$ BAL_3_A1B0  : num [1:96] 6.17 8.12 5.88 6.7 6.74 ...
#>   ..$ BAL_37_A1B1 : num [1:96] 5.98 7.59 5.7 6.78 6.81 ...
#>   ..$ BAL_17_A0B0 : num [1:96] 6.4 8.5 6.22 6.85 6.73 ...
#>   ..$ BAL_42_A0B0 : num [1:96] 5.72 7.74 5.42 6.33 6.18 ...
#>   ..$ BAL_44b_A0B0: num [1:96] 5.89 7.69 6.13 6.61 6.47 ...
#>   ..$ BAL_18_A1B0 : num [1:96] 6.17 8.31 5.84 6.82 6.98 ...
#>   ..$ BAL_12_A1B0 : num [1:96] 5.84 7.79 5.7 6.6 6.5 ...
#>   ..$ BAL_29_A1B0 : num [1:96] 5.64 7.39 5.34 6.28 6.1 ...
#>   ..$ BAL_46_A2B0 : num [1:96] 5.45 5.87 5.12 6.4 5.38 ...
#>   ..$ BAL_20a_A1B1: num [1:96] NaN NaN NaN NaN NaN NaN NaN NaN NaN NaN ...
#>   ..$ BAL_45a_A1B0: num [1:96] 6.18 7.79 6.02 6.83 6.71 ...
#>   ..$ BAL_26_A2B0 : num [1:96] 6.4 8.36 6.27 6.97 7.01 ...
#>   ..$ BAL_31_A0B1 : num [1:96] 5.89 7.25 5.26 6.17 6.1 ...
#>   ..- attr(*, ".internal.selfref")=<externalptr>