Dataset SVD information

Retrieves the SVD information for the given dataset

getDatasetSVD(
  dataset = NA_character_,
  raw = getOption("gemma.raw", FALSE),
  async = getOption("gemma.async", FALSE),
  memoised = getOption("gemma.memoise", FALSE),
  file = getOption("gemma.file", NA_character_),
  overwrite = getOption("gemma.overwrite", FALSE)
)

Arguments

dataset
Required, part of the URL path. Can either be the dataset ID or its short name (e.g. GSE1234). Retrieval by ID is more efficient. Only datasets that user has access to will be available

raw
TRUE to receive results as-is from Gemma, or FALSE to enable parsing.

async
TRUE to run the API query on a separate worker, or FALSE to run synchronously. See the async package for details.

memoised
Whether or not to cache results so future requests for the same data will be faster. Use forgetGemmaMemoised to clear the cache.

file
The name of a file to save the results to, or NULL to not write results to a file. If raw == TRUE, the output will be a JSON file. Otherwise, it will be a RDS file.

overwrite
Whether or not to overwrite if a file exists at the specified filename.

Value

A simple SVD value object for the given dataset, containing information about SVD of expression data A 404 error if the given identifier does not map to any object. Properties of the returned object are:

bioMaterialIds - Array of Bio Material IDs, in same order as the rows of the v matrix

variances - An array of values representing the fraction of the variance each component accounts for

vMatrix - the V Matrix (DoubleMatrix object)

Examples

getDatasetSVD("GSE2018")

#>         variance     VMatrix.0    VMatrix.1    VMatrix.2    VMatrix.3
#>  1: 1.474620e-01  0.2565417035 -0.167010041  0.049788449 -0.115597391
#>  2: 8.879356e-02 -0.0541179839 -0.267319955 -0.031958555  0.163699109
#>  3: 7.156613e-02 -0.0940208936  0.104490903 -0.305648063 -0.016504801
#>  4: 5.630170e-02  0.0002710698  0.324659349  0.203998806  0.177687730
#>  5: 5.411161e-02  0.2867611383 -0.012404342  0.131708008 -0.262558201
#>  6: 4.469272e-02  0.2301552376  0.056669732 -0.088796381  0.042894099
#>  7: 3.610912e-02 -0.0355673458  0.021461446 -0.149942673  0.298044826
#>  8: 3.440618e-02  0.2567207128 -0.222474178 -0.032381295  0.166867378
#>  9: 3.250666e-02  0.1971013531 -0.238300922  0.047572081 -0.002486732
#> 10: 3.166715e-02 -0.1066055279 -0.149704984  0.386137269  0.089778692
#> 11: 2.888146e-02 -0.0689621645 -0.006969266 -0.058568314 -0.200448114
#> 12: 2.765948e-02 -0.0140871809  0.007534199  0.059319737 -0.001581319
#> 13: 2.698289e-02  0.2038517023  0.302320218  0.188674117  0.056034788
#> 14: 2.435104e-02 -0.1755076341  0.063967300  0.149695797  0.237635497
#> 15: 2.363000e-02 -0.2915941535 -0.044384121 -0.040426955  0.007718541
#> 16: 2.154741e-02 -0.1773410872  0.008116184 -0.220339216 -0.042049046
#> 17: 2.012772e-02  0.1784410270 -0.061665447 -0.093924363  0.051036808
#> 18: 1.970868e-02  0.2012261112  0.342224179  0.072196028  0.062582263
#> 19: 1.889959e-02 -0.1410828652  0.233499739  0.009646479 -0.168372786
#> 20: 1.840312e-02  0.2054826495 -0.148696609 -0.007989942  0.044538873
#> 21: 1.665425e-02 -0.1693693758 -0.033075385  0.095327860 -0.057859537
#> 22: 1.591834e-02  0.0119224364 -0.290432713 -0.047415158 -0.087786718
#> 23: 1.566925e-02 -0.2779111049  0.039415714  0.138508663  0.207132537
#> 24: 1.487207e-02 -0.2087926801 -0.171721119 -0.122363126  0.120442639
#> 25: 1.412241e-02  0.0844670733  0.039109689 -0.366984730  0.063645297
#> 26: 1.379686e-02 -0.1338327571  0.041668876  0.135227045 -0.226457687
#> 27: 1.317481e-02 -0.1707941519  0.130216382  0.192904445  0.164869111
#> 28: 1.269458e-02  0.0028224669 -0.067144621  0.206614528  0.106948383
#> 29: 1.206858e-02 -0.1157888989  0.086720940 -0.130727667 -0.432985535
#> 30: 1.170639e-02 -0.0495173339 -0.366562657  0.098344158 -0.077070869
#> 31: 1.126706e-02  0.0869463552  0.089838255 -0.387769731  0.229190344
#> 32: 1.030402e-02  0.2962102460  0.180379601  0.126482196 -0.122705052
#> 33: 9.943119e-03 -0.0643830571  0.181317089 -0.243304287 -0.069086397
#> 34: 1.550993e-29 -0.1496450867 -0.005743437  0.036394788 -0.407196731
#>         variance     VMatrix.0    VMatrix.1    VMatrix.2    VMatrix.3
#>         VMatrix.4
#>  1:  0.2320494422
#>  2: -0.2216791580
#>  3: -0.2336448897
#>  4: -0.0940908804
#>  5:  0.0794813982
#>  6: -0.0748897943
#>  7:  0.4723832648
#>  8: -0.1496877092
#>  9: -0.0363681199
#> 10:  0.0004814266
#> 11:  0.2851787597
#> 12: -0.0165917812
#> 13:  0.0413235624
#> 14:  0.0968249138
#> 15: -0.0398001913
#> 16:  0.2228432999
#> 17: -0.1420940305
#> 18: -0.0058123943
#> 19: -0.2430865505
#> 20: -0.1537186371
#> 21:  0.2189013345
#> 22: -0.0687201079
#> 23:  0.1192975965
#> 24: -0.0828572188
#> 25:  0.1583922717
#> 26: -0.0702153039
#> 27: -0.2470380423
#> 28:  0.1152555512
#> 29: -0.0394305647
#> 30: -0.1153430898
#> 31:  0.0834110689
#> 32:  0.0610444126
#> 33: -0.3133474314
#> 34:  0.1615475923
#>         VMatrix.4