Dataset samples

Retrieves samples for the given dataset

getDatasetSamples(
  dataset = NA_character_,
  raw = getOption("gemma.raw", FALSE),
  async = getOption("gemma.async", FALSE),
  memoised = getOption("gemma.memoise", FALSE),
  file = getOption("gemma.file", NA_character_),
  overwrite = getOption("gemma.overwrite", FALSE)
)

Arguments

dataset
Required, part of the URL path. Can either be the dataset ID or its short name (e.g. GSE1234). Retrieval by ID is more efficient. Only datasets that user has access to will be available

raw
TRUE to receive results as-is from Gemma, or FALSE to enable parsing.

async
TRUE to run the API query on a separate worker, or FALSE to run synchronously. See the async package for details.

memoised
Whether or not to cache results so future requests for the same data will be faster. Use forgetGemmaMemoised to clear the cache.

file
The name of a file to save the results to, or NULL to not write results to a file. If raw == TRUE, the output will be a JSON file. Otherwise, it will be a RDS file.

overwrite
Whether or not to overwrite if a file exists at the specified filename.

Value

An array of samples (bio assay value objects) in the given dataset. A 404 error if the given identifier does not map to any object.

Examples

getDatasetSamples("GSE2018")

#>                    bioMaterial.Name  sample.Name sample.ID
#>  1:  GSE2018_bioMaterial_4|GSM36426   BAL_1_A0B0        31
#>  2: GSE2018_bioMaterial_15|GSM36437 BAL_27a_A0B0        20
#>  3: GSE2018_bioMaterial_26|GSM36448  BAL_43_A0B0         9
#>  4:  GSE2018_bioMaterial_6|GSM36428  BAL_13_A0B0        29
#>  5: GSE2018_bioMaterial_17|GSM36439  BAL_29_A1B0        18
#>  6: GSE2018_bioMaterial_28|GSM36450 BAL_44b_A0B0         7
#>  7:  GSE2018_bioMaterial_8|GSM36430  BAL_18_A1B0        27
#>  8: GSE2018_bioMaterial_19|GSM36441  BAL_31_A0B1        16
#>  9: GSE2018_bioMaterial_30|GSM36452 BAL_45b_A2B0         5
#> 10: GSE2018_bioMaterial_10|GSM36432 BAL_20a_A1B1        25
#> 11: GSE2018_bioMaterial_21|GSM36443  BAL_39_A0B1        14
#> 12: GSE2018_bioMaterial_32|GSM36454 BAL_47a_A0B0         3
#> 13:  GSE2018_bioMaterial_1|GSM36423  BAL_37_A1B1        34
#> 14: GSE2018_bioMaterial_12|GSM36434  BAL_21_A0B0        23
#> 15: GSE2018_bioMaterial_23|GSM36445  BAL_40_A0B0        12
#> 16: GSE2018_bioMaterial_34|GSM36456 BAL_47c_A0B1         1
#> 17:  GSE2018_bioMaterial_3|GSM36425 BAL_15b_A0B0        32
#> 18: GSE2018_bioMaterial_14|GSM36436  BAL_26_A2B0        21
#> 19: GSE2018_bioMaterial_25|GSM36447  BAL_42_A0B0        10
#> 20:  GSE2018_bioMaterial_5|GSM36427  BAL_12_A1B0        30
#> 21: GSE2018_bioMaterial_16|GSM36438 BAL_27b_A2B1        19
#> 22: GSE2018_bioMaterial_27|GSM36449 BAL_44a_A0B0         8
#> 23:  GSE2018_bioMaterial_7|GSM36429  BAL_17_A0B0        28
#> 24: GSE2018_bioMaterial_18|GSM36440   BAL_3_A1B0        17
#> 25: GSE2018_bioMaterial_29|GSM36451 BAL_45a_A1B0         6
#> 26:  GSE2018_bioMaterial_9|GSM36431  BAL_19_A0B0        26
#> 27: GSE2018_bioMaterial_20|GSM36442  BAL_33_A0B1        15
#> 28: GSE2018_bioMaterial_31|GSM36453  BAL_46_A2B0         4
#> 29: GSE2018_bioMaterial_11|GSM36433 BAL_20b_A1B1        24
#> 30: GSE2018_bioMaterial_22|GSM36444   BAL_4_A0B0        13
#> 31: GSE2018_bioMaterial_33|GSM36455 BAL_47b_A0B1         2
#> 32:  GSE2018_bioMaterial_2|GSM36424 BAL_15a_A0B1        33
#> 33: GSE2018_bioMaterial_13|GSM36435  BAL_25_A0B0        22
#> 34: GSE2018_bioMaterial_24|GSM36446  BAL_41_A1B0        11
#>                    bioMaterial.Name  sample.Name sample.ID
#>                                                  sample.Correspondence
#>  1:                                Corresponding to Samples:  GSM36426
#>  2:                                Corresponding to Samples:  GSM36437
#>  3: Corresponding to samples:  ial corresponding to samples:  GSM36448
#>  4:                                Corresponding to Samples:  GSM36428
#>  5:                                Corresponding to Samples:  GSM36439
#>  6: Corresponding to samples:  ial corresponding to samples:  GSM36450
#>  7:                                Corresponding to Samples:  GSM36430
#>  8:                                Corresponding to Samples:  GSM36441
#>  9: Corresponding to samples:  ial corresponding to samples:  GSM36452
#> 10:                                Corresponding to Samples:  GSM36432
#> 11:                                Corresponding to Samples:  GSM36443
#> 12: Corresponding to samples:  ial corresponding to samples:  GSM36454
#> 13:                                Corresponding to Samples:  GSM36423
#> 14:                                Corresponding to Samples:  GSM36434
#> 15:                                Corresponding to Samples:  GSM36445
#> 16: Corresponding to samples:  ial corresponding to samples:  GSM36456
#> 17:                                Corresponding to Samples:  GSM36425
#> 18:                                Corresponding to Samples:  GSM36436
#> 19: Corresponding to samples:  ial corresponding to samples:  GSM36447
#> 20:                                Corresponding to Samples:  GSM36427
#> 21:                                Corresponding to Samples:  GSM36438
#> 22: Corresponding to samples:  ial corresponding to samples:  GSM36449
#> 23:                                Corresponding to Samples:  GSM36429
#> 24:                                Corresponding to Samples:  GSM36440
#> 25: Corresponding to samples:  ial corresponding to samples:  GSM36451
#> 26:                                Corresponding to Samples:  GSM36431
#> 27:                                Corresponding to Samples:  GSM36442
#> 28: Corresponding to samples:  ial corresponding to samples:  GSM36453
#> 29:                                Corresponding to Samples:  GSM36433
#> 30:                                Corresponding to Samples:  GSM36444
#> 31: Corresponding to samples:  ial corresponding to samples:  GSM36455
#> 32:                                Corresponding to Samples:  GSM36424
#> 33:                                Corresponding to Samples:  GSM36435
#> 34:                                Corresponding to Samples:  GSM36446
#>                                                  sample.Correspondence
#>                                                                                                                                                                                                                                           sample.Description
#>  1:                                                                Bronchoalveolar lavage cells from a patient with an ISHLT biopsy score of A=0 and B=0 (Non-acute rejection).\nSource GEO sample is GSM36426\nLast updated (according to GEO): May 27 2005
#>  2:     Bronchoalveolar lavage cells from a patient with an ISHLT biopsy score of A=0 and B=0 (Non-acute rejection). One of two samples from the same patient, different visits\nSource GEO sample is GSM36437\nLast updated (according to GEO): May 27 2005
#>  3:                                                                Bronchoalveolar lavage cells from a patient with an ISHLT biopsy score of A=0 and B=0 (Non-acute rejection).\nSource GEO sample is GSM36448\nLast updated (according to GEO): May 27 2005
#>  4:                                                                Bronchoalveolar lavage cells from a patient with an ISHLT biopsy score of A=0 and B=0 (Non-acute rejection).\nSource GEO sample is GSM36428\nLast updated (according to GEO): May 27 2005
#>  5:                                                                Bronchoalveolar lavage cells from a patient with an ISHLT biopsy score of A=1 and B=0 (Non-acute rejection).\nSource GEO sample is GSM36439\nLast updated (according to GEO): May 27 2005
#>  6:     Bronchoalveolar lavage cells from a patient with an ISHLT biopsy score of A=0 and B=0 (Non-acute rejection). One of two samples from the same patient, different visits\nSource GEO sample is GSM36450\nLast updated (according to GEO): May 27 2005
#>  7:                                                                Bronchoalveolar lavage cells from a patient with an ISHLT biopsy score of A=1 and B=0 (Non-acute rejection).\nSource GEO sample is GSM36430\nLast updated (according to GEO): May 27 2005
#>  8:                                                                Bronchoalveolar lavage cells from a patient with an ISHLT biopsy score of A=0 and B=1 (Non-acute rejection).\nSource GEO sample is GSM36441\nLast updated (according to GEO): May 27 2005
#>  9:         Bronchoalveolar lavage cells from a patient with an ISHLT biopsy score of A=2 and B=0 (Acute rejection). One of two samples from the same patient, different visits\nSource GEO sample is GSM36452\nLast updated (according to GEO): May 27 2005
#> 10:        Bronchoalveolar lavage cells from a patient with an ISHLT biopsy score of A=1 and B=1 (Acute rejection). One of two samples from the same patient, different visits.\nSource GEO sample is GSM36432\nLast updated (according to GEO): May 27 2005
#> 11:                                                                Bronchoalveolar lavage cells from a patient with an ISHLT biopsy score of A=0 and B=1 (Non-acute rejection).\nSource GEO sample is GSM36443\nLast updated (according to GEO): May 27 2005
#> 12:  Bronchoalveolar lavage cells from a patient with an ISHLT biopsy score of A=0 and B=0 (Non-acute rejection). One of three samples from the same patient, different visits.\nSource GEO sample is GSM36454\nLast updated (according to GEO): May 27 2005
#> 13:                                                                     Bronchoalveolar lavage cells from a patient with an ISHLT biopsy score of A=1 and B=1 (Acute rejection)\nSource GEO sample is GSM36423\nLast updated (according to GEO): May 27 2005
#> 14:                                                                Bronchoalveolar lavage cells from a patient with an ISHLT biopsy score of A=0 and B=0 (Non-acute rejection).\nSource GEO sample is GSM36434\nLast updated (according to GEO): May 27 2005
#> 15:                                                                Bronchoalveolar lavage cells from a patient with an ISHLT biopsy score of A=0 and B=0 (Non-acute rejection).\nSource GEO sample is GSM36445\nLast updated (according to GEO): May 27 2005
#> 16:  Bronchoalveolar lavage cells from a patient with an ISHLT biopsy score of A=0 and B=1 (Non-acute rejection). One of three samples from the same patient, different visits.\nSource GEO sample is GSM36456\nLast updated (according to GEO): May 27 2005
#> 17:     Bronchoalveolar lavage cells from a patient with an ISHLT biopsy score of A=0 and B=0 (Non-acute rejection). One of two samples from the same patient, different visits\nSource GEO sample is GSM36425\nLast updated (according to GEO): May 27 2005
#> 18:                                                                    Bronchoalveolar lavage cells from a patient with an ISHLT biopsy score of A=2 and B=0 (Acute rejection).\nSource GEO sample is GSM36436\nLast updated (according to GEO): May 27 2005
#> 19:                                                                Bronchoalveolar lavage cells from a patient with an ISHLT biopsy score of A=0 and B=0 (Non-acute rejection).\nSource GEO sample is GSM36447\nLast updated (according to GEO): May 27 2005
#> 20:                                                                Bronchoalveolar lavage cells from a patient with an ISHLT biopsy score of A=1 and B=0 (Non-acute rejection).\nSource GEO sample is GSM36427\nLast updated (according to GEO): May 27 2005
#> 21:         Bronchoalveolar lavage cells from a patient with an ISHLT biopsy score of A=2 and B=1 (Acute rejection). One of two samples from the same patient, different visits\nSource GEO sample is GSM36438\nLast updated (according to GEO): May 27 2005
#> 22:     Bronchoalveolar lavage cells from a patient with an ISHLT biopsy score of A=0 and B=0 (Non-acute rejection). One of two samples from the same patient, different visits\nSource GEO sample is GSM36449\nLast updated (according to GEO): May 27 2005
#> 23:                                                                Bronchoalveolar lavage cells from a patient with an ISHLT biopsy score of A=0 and B=0 (Non-acute rejection).\nSource GEO sample is GSM36429\nLast updated (according to GEO): May 27 2005
#> 24:                                                                Bronchoalveolar lavage cells from a patient with an ISHLT biopsy score of A=1 and B=0 (Non-acute rejection).\nSource GEO sample is GSM36440\nLast updated (according to GEO): May 27 2005
#> 25:    Bronchoalveolar lavage cells from a patient with an ISHLT biopsy score of A=1 and B=0 (Non-acute rejection). One of two samples from the same patient, different visits.\nSource GEO sample is GSM36451\nLast updated (according to GEO): May 27 2005
#> 26:                                                                Bronchoalveolar lavage cells from a patient with an ISHLT biopsy score of A=0 and B=0 (Non-acute rejection).\nSource GEO sample is GSM36431\nLast updated (according to GEO): May 27 2005
#> 27:                                                                Bronchoalveolar lavage cells from a patient with an ISHLT biopsy score of A=0 and B=1 (Non-acute rejection).\nSource GEO sample is GSM36442\nLast updated (according to GEO): May 27 2005
#> 28:                                                                    Bronchoalveolar lavage cells from a patient with an ISHLT biopsy score of A=2 and B=0 (Acute rejection).\nSource GEO sample is GSM36453\nLast updated (according to GEO): May 27 2005
#> 29:        Bronchoalveolar lavage cells from a patient with an ISHLT biopsy score of A=1 and B=1 (Acute rejection). One of two samples from the same patient, different visits.\nSource GEO sample is GSM36433\nLast updated (according to GEO): May 27 2005
#> 30:                                                                Bronchoalveolar lavage cells from a patient with an ISHLT biopsy score of A=0 and B=0 (Non-acute rejection).\nSource GEO sample is GSM36444\nLast updated (according to GEO): May 27 2005
#> 31:  Bronchoalveolar lavage cells from a patient with an ISHLT biopsy score of A=0 and B=1 (Non-acute rejection). One of three samples from the same patient, different visits.\nSource GEO sample is GSM36455\nLast updated (according to GEO): May 27 2005
#> 32:    Bronchoalveolar lavage cells from a patient with an ISHLT biopsy score of A=0 and B=1 (Non-acute rejection). One of two samples from the same patient, different visits.\nSource GEO sample is GSM36424\nLast updated (according to GEO): May 27 2005
#> 33:                                                                Bronchoalveolar lavage cells from a patient with an ISHLT biopsy score of A=0 and B=0 (Non-acute rejection).\nSource GEO sample is GSM36435\nLast updated (according to GEO): May 27 2005
#> 34:                                                                Bronchoalveolar lavage cells from a patient with an ISHLT biopsy score of A=1 and B=0 (Non-acute rejection).\nSource GEO sample is GSM36446\nLast updated (according to GEO): May 27 2005
#>                                                                                                                                                                                                                                           sample.Description
#>     sample.Outlier sample.Accession sample.Processed sample.Characteristics
#>  1:          FALSE         GSM36426     1.036695e+12      <data.table[2x4]>
#>  2:          FALSE         GSM36437     1.036693e+12      <data.table[2x4]>
#>  3:          FALSE         GSM36448     1.035573e+12      <data.table[2x4]>
#>  4:          FALSE         GSM36428     1.022102e+12      <data.table[2x4]>
#>  5:          FALSE         GSM36439     1.033757e+12      <data.table[2x4]>
#>  6:          FALSE         GSM36450     1.018980e+12      <data.table[2x4]>
#>  7:          FALSE         GSM36430     1.038437e+12      <data.table[2x4]>
#>  8:          FALSE         GSM36441     1.017868e+12      <data.table[2x4]>
#>  9:          FALSE         GSM36452     1.022094e+12      <data.table[2x4]>
#> 10:           TRUE         GSM36432     1.033585e+12      <data.table[2x4]>
#> 11:          FALSE         GSM36443     1.034711e+12      <data.table[2x4]>
#> 12:          FALSE         GSM36454     1.035314e+12      <data.table[2x4]>
#> 13:          FALSE         GSM36423     1.035575e+12      <data.table[2x4]>
#> 14:          FALSE         GSM36434     1.014920e+12      <data.table[2x4]>
#> 15:          FALSE         GSM36445     1.022189e+12      <data.table[2x4]>
#> 16:          FALSE         GSM36456     1.031678e+12      <data.table[2x4]>
#> 17:          FALSE         GSM36425     1.023216e+12      <data.table[2x4]>
#> 18:          FALSE         GSM36436     1.022179e+12      <data.table[2x4]>
#> 19:          FALSE         GSM36447     1.038437e+12      <data.table[2x4]>
#> 20:          FALSE         GSM36427     1.022872e+12      <data.table[2x4]>
#> 21:          FALSE         GSM36438     1.014919e+12      <data.table[2x4]>
#> 22:          FALSE         GSM36449     1.035308e+12      <data.table[2x4]>
#> 23:          FALSE         GSM36429     1.038435e+12      <data.table[2x4]>
#> 24:          FALSE         GSM36440     1.038433e+12      <data.table[2x4]>
#> 25:          FALSE         GSM36451     1.027532e+12      <data.table[2x4]>
#> 26:          FALSE         GSM36431     1.016133e+12      <data.table[2x4]>
#> 27:          FALSE         GSM36442     1.035307e+12      <data.table[2x4]>
#> 28:          FALSE         GSM36453     1.012521e+12      <data.table[2x4]>
#> 29:          FALSE         GSM36433     1.035313e+12      <data.table[2x4]>
#> 30:          FALSE         GSM36444     1.038439e+12      <data.table[2x4]>
#> 31:          FALSE         GSM36455     1.033677e+12      <data.table[2x4]>
#> 32:          FALSE         GSM36424     1.036696e+12      <data.table[2x4]>
#> 33:          FALSE         GSM36435     1.033758e+12      <data.table[2x4]>
#> 34:          FALSE         GSM36446     1.017866e+12      <data.table[2x4]>
#>     sample.Outlier sample.Accession sample.Processed sample.Characteristics
#>     sample.FactorValues array.ShortName
#>  1:   <data.table[2x4]>           GPL96
#>  2:   <data.table[2x4]>           GPL96
#>  3:   <data.table[2x4]>           GPL96
#>  4:   <data.table[2x4]>           GPL96
#>  5:   <data.table[2x4]>           GPL96
#>  6:   <data.table[2x4]>           GPL96
#>  7:   <data.table[2x4]>           GPL96
#>  8:   <data.table[2x4]>           GPL96
#>  9:   <data.table[3x4]>           GPL96
#> 10:   <data.table[3x4]>           GPL96
#> 11:   <data.table[2x4]>           GPL96
#> 12:   <data.table[2x4]>           GPL96
#> 13:   <data.table[3x4]>           GPL96
#> 14:   <data.table[2x4]>           GPL96
#> 15:   <data.table[2x4]>           GPL96
#> 16:   <data.table[2x4]>           GPL96
#> 17:   <data.table[2x4]>           GPL96
#> 18:   <data.table[3x4]>           GPL96
#> 19:   <data.table[2x4]>           GPL96
#> 20:   <data.table[2x4]>           GPL96
#> 21:   <data.table[3x4]>           GPL96
#> 22:   <data.table[2x4]>           GPL96
#> 23:   <data.table[2x4]>           GPL96
#> 24:   <data.table[2x4]>           GPL96
#> 25:   <data.table[2x4]>           GPL96
#> 26:   <data.table[2x4]>           GPL96
#> 27:   <data.table[2x4]>           GPL96
#> 28:   <data.table[3x4]>           GPL96
#> 29:   <data.table[3x4]>           GPL96
#> 30:   <data.table[2x4]>           GPL96
#> 31:   <data.table[2x4]>           GPL96
#> 32:   <data.table[2x4]>           GPL96
#> 33:   <data.table[2x4]>           GPL96
#> 34:   <data.table[2x4]>           GPL96
#>     sample.FactorValues array.ShortName
#>                                                   array.Name array.Taxon
#>  1: Affymetrix GeneChip Human Genome U133 Array Set HG-U133A          NA
#>  2: Affymetrix GeneChip Human Genome U133 Array Set HG-U133A          NA
#>  3: Affymetrix GeneChip Human Genome U133 Array Set HG-U133A          NA
#>  4: Affymetrix GeneChip Human Genome U133 Array Set HG-U133A          NA
#>  5: Affymetrix GeneChip Human Genome U133 Array Set HG-U133A          NA
#>  6: Affymetrix GeneChip Human Genome U133 Array Set HG-U133A          NA
#>  7: Affymetrix GeneChip Human Genome U133 Array Set HG-U133A          NA
#>  8: Affymetrix GeneChip Human Genome U133 Array Set HG-U133A          NA
#>  9: Affymetrix GeneChip Human Genome U133 Array Set HG-U133A          NA
#> 10: Affymetrix GeneChip Human Genome U133 Array Set HG-U133A          NA
#> 11: Affymetrix GeneChip Human Genome U133 Array Set HG-U133A          NA
#> 12: Affymetrix GeneChip Human Genome U133 Array Set HG-U133A          NA
#> 13: Affymetrix GeneChip Human Genome U133 Array Set HG-U133A          NA
#> 14: Affymetrix GeneChip Human Genome U133 Array Set HG-U133A          NA
#> 15: Affymetrix GeneChip Human Genome U133 Array Set HG-U133A          NA
#> 16: Affymetrix GeneChip Human Genome U133 Array Set HG-U133A          NA
#> 17: Affymetrix GeneChip Human Genome U133 Array Set HG-U133A          NA
#> 18: Affymetrix GeneChip Human Genome U133 Array Set HG-U133A          NA
#> 19: Affymetrix GeneChip Human Genome U133 Array Set HG-U133A          NA
#> 20: Affymetrix GeneChip Human Genome U133 Array Set HG-U133A          NA
#> 21: Affymetrix GeneChip Human Genome U133 Array Set HG-U133A          NA
#> 22: Affymetrix GeneChip Human Genome U133 Array Set HG-U133A          NA
#> 23: Affymetrix GeneChip Human Genome U133 Array Set HG-U133A          NA
#> 24: Affymetrix GeneChip Human Genome U133 Array Set HG-U133A          NA
#> 25: Affymetrix GeneChip Human Genome U133 Array Set HG-U133A          NA
#> 26: Affymetrix GeneChip Human Genome U133 Array Set HG-U133A          NA
#> 27: Affymetrix GeneChip Human Genome U133 Array Set HG-U133A          NA
#> 28: Affymetrix GeneChip Human Genome U133 Array Set HG-U133A          NA
#> 29: Affymetrix GeneChip Human Genome U133 Array Set HG-U133A          NA
#> 30: Affymetrix GeneChip Human Genome U133 Array Set HG-U133A          NA
#> 31: Affymetrix GeneChip Human Genome U133 Array Set HG-U133A          NA
#> 32: Affymetrix GeneChip Human Genome U133 Array Set HG-U133A          NA
#> 33: Affymetrix GeneChip Human Genome U133 Array Set HG-U133A          NA
#> 34: Affymetrix GeneChip Human Genome U133 Array Set HG-U133A          NA
#>                                                   array.Name array.Taxon
#>     array.TaxonID array.Type array.Description array.Troubled
#>  1:            NA         NA                NA          FALSE
#>  2:            NA         NA                NA          FALSE
#>  3:            NA         NA                NA          FALSE
#>  4:            NA         NA                NA          FALSE
#>  5:            NA         NA                NA          FALSE
#>  6:            NA         NA                NA          FALSE
#>  7:            NA         NA                NA          FALSE
#>  8:            NA         NA                NA          FALSE
#>  9:            NA         NA                NA          FALSE
#> 10:            NA         NA                NA          FALSE
#> 11:            NA         NA                NA          FALSE
#> 12:            NA         NA                NA          FALSE
#> 13:            NA         NA                NA          FALSE
#> 14:            NA         NA                NA          FALSE
#> 15:            NA         NA                NA          FALSE
#> 16:            NA         NA                NA          FALSE
#> 17:            NA         NA                NA          FALSE
#> 18:            NA         NA                NA          FALSE
#> 19:            NA         NA                NA          FALSE
#> 20:            NA         NA                NA          FALSE
#> 21:            NA         NA                NA          FALSE
#> 22:            NA         NA                NA          FALSE
#> 23:            NA         NA                NA          FALSE
#> 24:            NA         NA                NA          FALSE
#> 25:            NA         NA                NA          FALSE
#> 26:            NA         NA                NA          FALSE
#> 27:            NA         NA                NA          FALSE
#> 28:            NA         NA                NA          FALSE
#> 29:            NA         NA                NA          FALSE
#> 30:            NA         NA                NA          FALSE
#> 31:            NA         NA                NA          FALSE
#> 32:            NA         NA                NA          FALSE
#> 33:            NA         NA                NA          FALSE
#> 34:            NA         NA                NA          FALSE
#>     array.TaxonID array.Type array.Description array.Troubled