Gene probes

Retrieves the probes (composite sequences) with this gene

getGeneProbes(
  gene = NA_character_,
  offset = 0L,
  limit = 20L,
  raw = getOption("gemma.raw", FALSE),
  async = getOption("gemma.async", FALSE),
  memoised = getOption("gemma.memoise", FALSE),
  file = getOption("gemma.file", NA_character_),
  overwrite = getOption("gemma.overwrite", FALSE)
)

Arguments

gene
Required, part of the URL path. Can either be the NCBI ID (1859), Ensembl ID (ENSG00000157540) or official symbol (DYRK1A) of the gene. NCBI ID is the most efficient (and guaranteed to be unique) identifier. Official symbol represents a gene homologue for a random taxon, unless used in a specific taxon (see Taxon Endpoints).

offset
Optional, defaults to 0. Skips the specified amount of objects when retrieving them from the database.

limit
Optional, defaults to 20. Limits the result to specified amount of objects. Use 0 for no limit.

raw
TRUE to receive results as-is from Gemma, or FALSE to enable parsing.

async
TRUE to run the API query on a separate worker, or FALSE to run synchronously. See the async package for details.

memoised
Whether or not to cache results so future requests for the same data will be faster. Use forgetGemmaMemoised to clear the cache.

file
The name of a file to save the results to, or NULL to not write results to a file. If raw == TRUE, the output will be a JSON file. Otherwise, it will be a RDS file.

overwrite
Whether or not to overwrite if a file exists at the specified filename.

Value

An array of probes (composite sequence value objects) that are mapped to this gene. Note, that it is possible for probes to map to multiple genes. A 404 error if the given identifier does not map to any object.

Examples

getGeneProbes("DYRK1A")

#>          mapping.Name mapping.Summary mapping.Description array.ShortName
#>  1:             26311              NA              Dyrk1a        GPL14745
#>  2:             20686              NA              Dyrk1a        GPL14745
#>  3:              1646              NA              Dyrk1a        GPL14745
#>  4:             44226              NA              Dyrk1a        GPL14745
#>  5:             34715              NA              Dyrk1a        GPL14745
#>  6:             23687              NA              Dyrk1a        GPL14745
#>  7:             33197              NA              Dyrk1a        GPL14745
#>  8:             37065              NA              Dyrk1a        GPL14745
#>  9:             38256              NA              Dyrk1a        GPL14745
#> 10:             42251              NA              Dyrk1a        GPL14745
#> 11:           7087018              NA                <NA>         GPL6543
#> 12:          10750437              NA                            GPL11534
#> 13: TC1100000329.rn.1              NA                            GPL22388
#> 14:             23876              NA              Dyrk1a        GPL15011
#> 15:             41351              NA                            GPL15011
#> 16:              9388              NA                            GPL15011
#> 17:             15610              NA                            GPL15011
#> 18:             24083              NA                            GPL11280
#> 19:             26644              NA                            GPL11280
#> 20:              2949              NA                            GPL11280
#>                                                                           array.Name
#>  1:                              Agilent-028282 Whole Rat Genome Microarray 4x44K v3
#>  2:                              Agilent-028282 Whole Rat Genome Microarray 4x44K v3
#>  3:                              Agilent-028282 Whole Rat Genome Microarray 4x44K v3
#>  4:                              Agilent-028282 Whole Rat Genome Microarray 4x44K v3
#>  5:                              Agilent-028282 Whole Rat Genome Microarray 4x44K v3
#>  6:                              Agilent-028282 Whole Rat Genome Microarray 4x44K v3
#>  7:                              Agilent-028282 Whole Rat Genome Microarray 4x44K v3
#>  8:                              Agilent-028282 Whole Rat Genome Microarray 4x44K v3
#>  9:                              Agilent-028282 Whole Rat Genome Microarray 4x44K v3
#> 10:                              Agilent-028282 Whole Rat Genome Microarray 4x44K v3
#> 11:       [RaEx-1_0-st] Affymetrix Rat Exon 1.0 ST Array [transcript (gene) version]
#> 12:     [RaGene-1_1-st] Affymetrix Rat Gene 1.1 ST Array [transcript (gene) version]
#> 13: [RTA-1_0] Affymetrix Rat Transcriptome Array 1.0 [transcript (gene) CSV version]
#> 14:                                              Agilent-016475 Patkany genome 4x44K
#> 15:                                              Agilent-016475 Patkany genome 4x44K
#> 16:                                              Agilent-016475 Patkany genome 4x44K
#> 17:                                              Agilent-016475 Patkany genome 4x44K
#> 18:        Agilent-024196 Whole Rat Genome Microarray 4x44K (Feature Number version)
#> 19:        Agilent-024196 Whole Rat Genome Microarray 4x44K (Feature Number version)
#> 20:        Agilent-024196 Whole Rat Genome Microarray 4x44K (Feature Number version)
#>     array.Taxon array.TaxonID array.Type
#>  1:         rat             3   ONECOLOR
#>  2:         rat             3   ONECOLOR
#>  3:         rat             3   ONECOLOR
#>  4:         rat             3   ONECOLOR
#>  5:         rat             3   ONECOLOR
#>  6:         rat             3   ONECOLOR
#>  7:         rat             3   ONECOLOR
#>  8:         rat             3   ONECOLOR
#>  9:         rat             3   ONECOLOR
#> 10:         rat             3   ONECOLOR
#> 11:         rat             3   ONECOLOR
#> 12:         rat             3   ONECOLOR
#> 13:         rat             3   ONECOLOR
#> 14:         rat             3   ONECOLOR
#> 15:         rat             3   ONECOLOR
#> 16:         rat             3   ONECOLOR
#> 17:         rat             3   ONECOLOR
#> 18:         rat             3   ONECOLOR
#> 19:         rat             3   ONECOLOR
#> 20:         rat             3   ONECOLOR
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                          array.Description
#>  1:   Whole Rat Genome Microarray 4x44K v3  Arrays of this design have barcodes that begin with 16028282 or 2528282.  Orientation: Features are numbered numbered Left-to-Right, Top-to-Bottom as scanned by an Agilent scanner (barcode on the left, DNA on the back surface, scanned through the glass), matching the FeatureNum output from Agilent's Feature Extraction software.  The ID column represents the Agilent Feature Extraction feature number.  Rows and columns are numbered as scanned by an Axon Scanner (barcode on the bottom, DNA on the front surface).  To match data scanned on an Axon scanner, use the RefNumber column contained in the Agilent-provided GAL file as the ID_REF column in sample submissions. 
#>  2:   Whole Rat Genome Microarray 4x44K v3  Arrays of this design have barcodes that begin with 16028282 or 2528282.  Orientation: Features are numbered numbered Left-to-Right, Top-to-Bottom as scanned by an Agilent scanner (barcode on the left, DNA on the back surface, scanned through the glass), matching the FeatureNum output from Agilent's Feature Extraction software.  The ID column represents the Agilent Feature Extraction feature number.  Rows and columns are numbered as scanned by an Axon Scanner (barcode on the bottom, DNA on the front surface).  To match data scanned on an Axon scanner, use the RefNumber column contained in the Agilent-provided GAL file as the ID_REF column in sample submissions. 
#>  3:   Whole Rat Genome Microarray 4x44K v3  Arrays of this design have barcodes that begin with 16028282 or 2528282.  Orientation: Features are numbered numbered Left-to-Right, Top-to-Bottom as scanned by an Agilent scanner (barcode on the left, DNA on the back surface, scanned through the glass), matching the FeatureNum output from Agilent's Feature Extraction software.  The ID column represents the Agilent Feature Extraction feature number.  Rows and columns are numbered as scanned by an Axon Scanner (barcode on the bottom, DNA on the front surface).  To match data scanned on an Axon scanner, use the RefNumber column contained in the Agilent-provided GAL file as the ID_REF column in sample submissions. 
#>  4:   Whole Rat Genome Microarray 4x44K v3  Arrays of this design have barcodes that begin with 16028282 or 2528282.  Orientation: Features are numbered numbered Left-to-Right, Top-to-Bottom as scanned by an Agilent scanner (barcode on the left, DNA on the back surface, scanned through the glass), matching the FeatureNum output from Agilent's Feature Extraction software.  The ID column represents the Agilent Feature Extraction feature number.  Rows and columns are numbered as scanned by an Axon Scanner (barcode on the bottom, DNA on the front surface).  To match data scanned on an Axon scanner, use the RefNumber column contained in the Agilent-provided GAL file as the ID_REF column in sample submissions. 
#>  5:   Whole Rat Genome Microarray 4x44K v3  Arrays of this design have barcodes that begin with 16028282 or 2528282.  Orientation: Features are numbered numbered Left-to-Right, Top-to-Bottom as scanned by an Agilent scanner (barcode on the left, DNA on the back surface, scanned through the glass), matching the FeatureNum output from Agilent's Feature Extraction software.  The ID column represents the Agilent Feature Extraction feature number.  Rows and columns are numbered as scanned by an Axon Scanner (barcode on the bottom, DNA on the front surface).  To match data scanned on an Axon scanner, use the RefNumber column contained in the Agilent-provided GAL file as the ID_REF column in sample submissions. 
#>  6:   Whole Rat Genome Microarray 4x44K v3  Arrays of this design have barcodes that begin with 16028282 or 2528282.  Orientation: Features are numbered numbered Left-to-Right, Top-to-Bottom as scanned by an Agilent scanner (barcode on the left, DNA on the back surface, scanned through the glass), matching the FeatureNum output from Agilent's Feature Extraction software.  The ID column represents the Agilent Feature Extraction feature number.  Rows and columns are numbered as scanned by an Axon Scanner (barcode on the bottom, DNA on the front surface).  To match data scanned on an Axon scanner, use the RefNumber column contained in the Agilent-provided GAL file as the ID_REF column in sample submissions. 
#>  7:   Whole Rat Genome Microarray 4x44K v3  Arrays of this design have barcodes that begin with 16028282 or 2528282.  Orientation: Features are numbered numbered Left-to-Right, Top-to-Bottom as scanned by an Agilent scanner (barcode on the left, DNA on the back surface, scanned through the glass), matching the FeatureNum output from Agilent's Feature Extraction software.  The ID column represents the Agilent Feature Extraction feature number.  Rows and columns are numbered as scanned by an Axon Scanner (barcode on the bottom, DNA on the front surface).  To match data scanned on an Axon scanner, use the RefNumber column contained in the Agilent-provided GAL file as the ID_REF column in sample submissions. 
#>  8:   Whole Rat Genome Microarray 4x44K v3  Arrays of this design have barcodes that begin with 16028282 or 2528282.  Orientation: Features are numbered numbered Left-to-Right, Top-to-Bottom as scanned by an Agilent scanner (barcode on the left, DNA on the back surface, scanned through the glass), matching the FeatureNum output from Agilent's Feature Extraction software.  The ID column represents the Agilent Feature Extraction feature number.  Rows and columns are numbered as scanned by an Axon Scanner (barcode on the bottom, DNA on the front surface).  To match data scanned on an Axon scanner, use the RefNumber column contained in the Agilent-provided GAL file as the ID_REF column in sample submissions. 
#>  9:   Whole Rat Genome Microarray 4x44K v3  Arrays of this design have barcodes that begin with 16028282 or 2528282.  Orientation: Features are numbered numbered Left-to-Right, Top-to-Bottom as scanned by an Agilent scanner (barcode on the left, DNA on the back surface, scanned through the glass), matching the FeatureNum output from Agilent's Feature Extraction software.  The ID column represents the Agilent Feature Extraction feature number.  Rows and columns are numbered as scanned by an Axon Scanner (barcode on the bottom, DNA on the front surface).  To match data scanned on an Axon scanner, use the RefNumber column contained in the Agilent-provided GAL file as the ID_REF column in sample submissions. 
#> 10:   Whole Rat Genome Microarray 4x44K v3  Arrays of this design have barcodes that begin with 16028282 or 2528282.  Orientation: Features are numbered numbered Left-to-Right, Top-to-Bottom as scanned by an Agilent scanner (barcode on the left, DNA on the back surface, scanned through the glass), matching the FeatureNum output from Agilent's Feature Extraction software.  The ID column represents the Agilent Feature Extraction feature number.  Rows and columns are numbered as scanned by an Axon Scanner (barcode on the bottom, DNA on the front surface).  To match data scanned on an Axon scanner, use the RefNumber column contained in the Agilent-provided GAL file as the ID_REF column in sample submissions. 
#> 11:                                                                                                                                                                                                                    Affymetrix submissions are typically submitted to GEO using the GEOarchive method described at http://www.ncbi.nlm.nih.gov/projects/geo/info/geo_affy.html  #%create_date=Mon Mar 16 20:01:52 2009 PDT #%chip_type=RaEx-1_0-st-v1 #%lib_set_name=RaEx-1_0-st #%lib_set_version=v1 #%genome-species=Rattus norvegicus #%genome-version=rn4 #%genome-version-ucsc=rn4 #%genome-version-ncbi=HGSC v3.4 #%genome-version-create_date=2004-11 #%netaffx-annotation-date=2009-03-16 #%netaffx-annotation-netaffx-build=28
#> 12:                                      Affymetrix submissions are typically submitted to GEO using the GEOarchive method described at http://www.ncbi.nlm.nih.gov/projects/geo/info/geo_affy.html  RaGene-1_1-st-v1.na31.rn4.transcript.csv #%create_date=Fri Sep 17 11:36:51 2010 PDT #%chip_type=RaGene-1_1-st-v1 #%lib_set_name=RaGene-1_1-st #%lib_set_version=v1 #%genome-species=Rattus norvegicus #%genome-version=rn4 #%genome-version-ucsc=rn4 #%genome-version-ncbi=Rnor3.4 #%genome-version-create_date=2004-11-00 #%genome-lifted_from-species=Rattus norvegicus #%genome-lifted_from-version-ucsc=rn4 #%genome-lifted_from-version-ncbi=Rnor3.4 #%netaffx-annotation-date=2009-11-23 #%netaffx-annotation-netaffx-build=31 
#> 13:                                                                       Annotation was derived from the Affymetrix annotation file RTA-1_0.na36.rn6.transcript.csv:  #%create_date=Wed Jun  8 16:48:57 2016 PDT #%chip_type=RTA-1_0 #%lib_set_name=RTA-1_0 #%lib_set_version=r1 #%genome-species=Rattus norvegicus #%genome-version=rn6 #%genome-version-ucsc=rn6 #%genome-version-ncbi=Rnor_6.0 #%genome-version-create_date=2014-07-30 #%genome-lifted-method=psl-map.pl #%genome-lifted_from-species=Rattus norvegicus #%genome-lifted_from-version-ucsc=rn6 #%genome-lifted_from-version-ncbi=Rnor_6.0 #%netaffx-annotation-date=2016-03-15 #%netaffx-annotation-netaffx-build=36 #%netaffx-annotation-data-type=transcript_cluster 
#> 14:                                         Arrays of this design have barcodes that begin with 16016475 or 2516475.  Orientation: Features are numbered numbered Left-to-Right, Top-to-Bottom as scanned by an Agilent scanner (barcode on the left, DNA on the back surface, scanned through the glass), matching the FeatureNum output from Agilent's Feature Extraction software.  The ID column represents the Agilent Feature Extraction feature number.  Rows and columns are numbered as scanned by an Axon Scanner (barcode on the bottom, DNA on the front surface).  To match data scanned on an Axon scanner, use the RefNumber column contained in the Agilent-provided GAL file as the ID_REF column in sample submissions. 
#> 15:                                         Arrays of this design have barcodes that begin with 16016475 or 2516475.  Orientation: Features are numbered numbered Left-to-Right, Top-to-Bottom as scanned by an Agilent scanner (barcode on the left, DNA on the back surface, scanned through the glass), matching the FeatureNum output from Agilent's Feature Extraction software.  The ID column represents the Agilent Feature Extraction feature number.  Rows and columns are numbered as scanned by an Axon Scanner (barcode on the bottom, DNA on the front surface).  To match data scanned on an Axon scanner, use the RefNumber column contained in the Agilent-provided GAL file as the ID_REF column in sample submissions. 
#> 16:                                         Arrays of this design have barcodes that begin with 16016475 or 2516475.  Orientation: Features are numbered numbered Left-to-Right, Top-to-Bottom as scanned by an Agilent scanner (barcode on the left, DNA on the back surface, scanned through the glass), matching the FeatureNum output from Agilent's Feature Extraction software.  The ID column represents the Agilent Feature Extraction feature number.  Rows and columns are numbered as scanned by an Axon Scanner (barcode on the bottom, DNA on the front surface).  To match data scanned on an Axon scanner, use the RefNumber column contained in the Agilent-provided GAL file as the ID_REF column in sample submissions. 
#> 17:                                         Arrays of this design have barcodes that begin with 16016475 or 2516475.  Orientation: Features are numbered numbered Left-to-Right, Top-to-Bottom as scanned by an Agilent scanner (barcode on the left, DNA on the back surface, scanned through the glass), matching the FeatureNum output from Agilent's Feature Extraction software.  The ID column represents the Agilent Feature Extraction feature number.  Rows and columns are numbered as scanned by an Axon Scanner (barcode on the bottom, DNA on the front surface).  To match data scanned on an Axon scanner, use the RefNumber column contained in the Agilent-provided GAL file as the ID_REF column in sample submissions. 
#> 18:                                                                                                                                                                                                                                                                                             With a focus on well known rat genes and homologues to human and mouse genes with useful annotation, Agilent's Whole Rat Genome Oligo – 4X44K microarray provides researchers with a new tool for modeling human biology in the rat model organism. For researchers, this means they now have access to a microarray made up of relevant content that has been empirically validated by Agilent.\nFrom GPL11280\nLast Updated: Dec 06 2010
#> 19:                                                                                                                                                                                                                                                                                             With a focus on well known rat genes and homologues to human and mouse genes with useful annotation, Agilent's Whole Rat Genome Oligo – 4X44K microarray provides researchers with a new tool for modeling human biology in the rat model organism. For researchers, this means they now have access to a microarray made up of relevant content that has been empirically validated by Agilent.\nFrom GPL11280\nLast Updated: Dec 06 2010
#> 20:                                                                                                                                                                                                                                                                                             With a focus on well known rat genes and homologues to human and mouse genes with useful annotation, Agilent's Whole Rat Genome Oligo – 4X44K microarray provides researchers with a new tool for modeling human biology in the rat model organism. For researchers, this means they now have access to a microarray made up of relevant content that has been empirically validated by Agilent.\nFrom GPL11280\nLast Updated: Dec 06 2010
#>     array.Troubled
#>  1:          FALSE
#>  2:          FALSE
#>  3:          FALSE
#>  4:          FALSE
#>  5:          FALSE
#>  6:          FALSE
#>  7:          FALSE
#>  8:          FALSE
#>  9:          FALSE
#> 10:          FALSE
#> 11:          FALSE
#> 12:          FALSE
#> 13:          FALSE
#> 14:          FALSE
#> 15:          FALSE
#> 16:          FALSE
#> 17:          FALSE
#> 18:          FALSE
#> 19:          FALSE
#> 20:          FALSE