Genes

Retrieves all genes matching the identifiers

getGenes(
  genes = NA_character_,
  raw = getOption("gemma.raw", FALSE),
  async = getOption("gemma.async", FALSE),
  memoised = getOption("gemma.memoise", FALSE),
  file = getOption("gemma.file", NA_character_),
  overwrite = getOption("gemma.overwrite", FALSE)
)

Arguments

genes
Required, part of the URL path. A list of identifiers, separated by commas (e.g: 1859, 5728). Can either be the NCBI ID (1859), Ensembl ID (ENSG00000157540) or official symbol (DYRK1A) of the gene. NCBI ID is the most efficient (and guaranteed to be unique) identifier. Official symbol represents a gene homologue for a random taxon, unless used in a specific taxon (see Taxon Endpoints). Do not combine different identifiers in one query.

raw
TRUE to receive results as-is from Gemma, or FALSE to enable parsing.

async
TRUE to run the API query on a separate worker, or FALSE to run synchronously. See the async package for details.

memoised
Whether or not to cache results so future requests for the same data will be faster. Use forgetGemmaMemoised to clear the cache.

file
The name of a file to save the results to, or NULL to not write results to a file. If raw == TRUE, the output will be a JSON file. Otherwise, it will be a RDS file.

overwrite
Whether or not to overwrite if a file exists at the specified filename.

Value

An array of value objects representing the objects that matched the query. Empty array if no objects matched. A 400 error if required parameters are missing.

Examples

getGenes("DYRK1A")

#>    gene.Symbol       gene.Ensembl gene.NCBI
#> 1:      DYRK1A    ENSG00000157540      1859
#> 2:      Dyrk1a ENSMUSG00000022897     13548
#> 3:      Dyrk1a ENSRNOG00000001662     25255
#>                                                            gene.Name
#> 1:     dual specificity tyrosine phosphorylation regulated kinase 1A
#> 2: dual-specificity tyrosine-(Y)-phosphorylation regulated kinase 1a
#> 3:     dual specificity tyrosine phosphorylation regulated kinase 1A
#>    gene.Aliases                                      gene.Description gene.GO
#> 1:           NA Imported from NCBI gene; Nomenclature status: INTERIM       0
#> 2:           NA Imported from NCBI gene; Nomenclature status: INTERIM       0
#> 3:           NA Imported from NCBI gene; Nomenclature status: INTERIM       0
#>    gene.Homologues gene.MFX.Rank taxon.Name taxon.ID  taxon.Scientific
#> 1:              NA             0      human        1      Homo sapiens
#> 2:              NA             0      mouse        2      Mus musculus
#> 3:              NA             0        rat        3 Rattus norvegicus
#>    phenotypes
#> 1:         NA
#> 2:         NA
#> 3:         NA

getGenes(c("DYRK1A", "PTEN"))

#>    gene.Symbol       gene.Ensembl gene.NCBI
#> 1:      DYRK1A    ENSG00000157540      1859
#> 2:        PTEN    ENSG00000171862      5728
#> 3:      Dyrk1a ENSMUSG00000022897     13548
#> 4:        Pten ENSMUSG00000013663     19211
#> 5:      Dyrk1a ENSRNOG00000001662     25255
#> 6:        Pten ENSRNOG00000020723     50557
#> 7:        Pten        FBgn0026379     43991
#>                                                            gene.Name
#> 1:     dual specificity tyrosine phosphorylation regulated kinase 1A
#> 2:                                    phosphatase and tensin homolog
#> 3: dual-specificity tyrosine-(Y)-phosphorylation regulated kinase 1a
#> 4:                                    phosphatase and tensin homolog
#> 5:     dual specificity tyrosine phosphorylation regulated kinase 1A
#> 6:                                    phosphatase and tensin homolog
#> 7:                                    Phosphatase and tensin homolog
#>    gene.Aliases                                      gene.Description gene.GO
#> 1:           NA Imported from NCBI gene; Nomenclature status: INTERIM       0
#> 2:           NA Imported from NCBI gene; Nomenclature status: INTERIM       0
#> 3:           NA Imported from NCBI gene; Nomenclature status: INTERIM       0
#> 4:           NA Imported from NCBI gene; Nomenclature status: INTERIM       0
#> 5:           NA Imported from NCBI gene; Nomenclature status: INTERIM       0
#> 6:           NA Imported from NCBI gene; Nomenclature status: INTERIM       0
#> 7:           NA Imported from NCBI gene; Nomenclature status: INTERIM       0
#>    gene.Homologues gene.MFX.Rank taxon.Name taxon.ID        taxon.Scientific
#> 1:              NA             0      human        1            Homo sapiens
#> 2:              NA             0      human        1            Homo sapiens
#> 3:              NA             0      mouse        2            Mus musculus
#> 4:              NA             0      mouse        2            Mus musculus
#> 5:              NA             0        rat        3       Rattus norvegicus
#> 6:              NA             0        rat        3       Rattus norvegicus
#> 7:              NA             0        fly       13 Drosophila melanogaster
#>    phenotypes
#> 1:         NA
#> 2:         NA
#> 3:         NA
#> 4:         NA
#> 5:         NA
#> 6:         NA
#> 7:         NA