Platforms

List platforms filtered and organized by given parameters

getPlatforms(
  platforms = NA_character_,
  filter = NA_character_,
  offset = 0L,
  limit = 20L,
  sort = "+id",
  raw = getOption("gemma.raw", FALSE),
  async = getOption("gemma.async", FALSE),
  memoised = getOption("gemma.memoise", FALSE),
  file = getOption("gemma.file", NA_character_),
  overwrite = getOption("gemma.overwrite", FALSE)
)

Arguments

platforms
Optional, defaults to empty. Limits the result to entities with given identifiers. A list of identifiers, separated by commas (e.g: GPL96,GPL1355,GPL1261). Identifiers can either be the Platform ID or its short name. Retrieval by ID is more efficient. Only platforms that user has access to will be available. Do not combine different identifiers in one query.

filter
Optional, defaults to empty. Filtering can be done on any* property or nested property that the appropriate object class defines or inherits (and that is mapped by hibernate). [These do not correspond to the properties of the objects returned by the API calls.].description-imp Class definitions:

Datasets: javaDoc gitHub

Platforms: javaDoc gitHub E.g: curationDetails or curationDetails.lastTroubledEvent.date.

Any property of a supported type. Currently supported types are:

String - property of String type, required value can be any String.

Number - any Number implementation. Required value must be a string parseable to the specific Number type.

Boolean - required value will be parsed to true only if the string matches 'true', ignoring case. Accepted operator keywords are:

'=' - equality

'!=' - non-equality

'<' - smaller than

'>' - larger than

'<=' - smaller or equal

'=>' - larger or equal

'like' - similar string, effectively means 'contains', translates to the sql 'LIKE' operator (given value will be surrounded by % signs) Multiple filters can be chained using AND and OR keywords. Leave space between the keywords and the previous/next word! E.g: ?filter=property1 < value1 AND property2 like value2 If chained filters are mixed conjunctions and disjunctions, the query must be in conjunctive normal form (CNF). Parentheses are not necessary

every AND keyword separates blocks of disjunctions. Example: ?filter=p1 = v1 OR p1 != v2 AND p2 <= v2 AND p3 > v3 OR p3 < v4 Above query will translate to: (p1 = v1 OR p1 != v2) AND (p2 <= v2) AND (p3 > v3 OR p3 < v4;) Breaking the CNF results in an error. Filter curationDetails.troubled will be ignored if user is not an administrator.

offset
Optional, defaults to 0. Skips the specified amount of objects when retrieving them from the database.

limit
Optional, defaults to 20. Limits the result to specified amount of objects. Use 0 for no limit.

sort
Optional, defaults to +id. Sets the ordering property and direction. Format is [+,-][property name]. E.g. -accession will translate to descending ordering by the 'Accession' property. Note that this does [not guarantee the order of the returned entities!].description-imp This is merely a signal to how the data should be pre-sorted before the limit and offset are applied. Nested properties are also supported (recursively). E.g: +curationDetails.lastTroubledEvent.date When using in scripts, remember to URL-encode the '+' plus character (see the compiled URL below).

raw
TRUE to receive results as-is from Gemma, or FALSE to enable parsing.

async
TRUE to run the API query on a separate worker, or FALSE to run synchronously. See the async package for details.

memoised
Whether or not to cache results so future requests for the same data will be faster. Use forgetGemmaMemoised to clear the cache.

file
The name of a file to save the results to, or NULL to not write results to a file. If raw == TRUE, the output will be a JSON file. Otherwise, it will be a RDS file.

overwrite
Whether or not to overwrite if a file exists at the specified filename.

Value

An array of value objects representing the objects that matched the query. Empty array if no objects matched.

Examples

getPlatforms("GPL1355")

#>    platform.ID platform.ShortName                                platform.Name
#> 1:           2            GPL1355 Affymetrix GeneChip Rat Genome 230 2.0 Array
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                        platform.Description
#> 1:  The GeneChip Rat Genome 230 2.0 Array is a powerful tool for toxicology, neurobiology, and other applications using rat as a model organism.  - Provides comprehensive coverage of the transcribed rat genome on a single array - Comprised of more than 31,000 probe sets, analyzing over 30,000 transcripts and variants from over 28,000 well-substantiated rat genes - The publicly available draft of the rat genome and leading public rat databases were used to refine sequences and provide a higher quality of data output  All probe sets represented on the GeneChip Rat Expression Set 230 are included on the GeneChip Rat Genome 230 2.0 Array. Sequences used in the design of the GeneChip Rat Genome 230 2.0 Array were selected from GenBank, dbEST, and RefSeq. The sequence clusters were created from the UniGene database (Build 99, June 2002) and then refined by analysis and comparison with the publicly available draft assembly of the rat genome from the Baylor College of Medicine Human Genome Sequencing Center (June 2002).  The GeneChip Rat Genome 230 2.0 Array includes representation of the RefSeq database sequences and probe sets related to sequences and refined EST clusters previously represented on the GeneChip Rat Genome U34 Set.  Oligonucleotide probes complementary to each corresponding sequence are synthesized in situ on the arrays. Eleven pairs of oligonucleotide probes are used to measure the level of transcription of each sequence represented on the GeneChip Rat Genome 230 2.0 Array.  Annotations derived from Affymetrix CSV file dated 6/23/2004\nFrom GPL1355\nLast Updated: May 31 2005
#>    platform.Troubled platform.ExperimentCount platform.GeneCount
#> 1:             FALSE                      288                 NA
#>    platform.ProbeSequenceCount platform.ProbeAlignmentCount
#> 1:                          NA                           NA
#>    platform.ProbeGeneCount platform.ElementCount taxon.Name taxon.ID
#> 1:                      NA                    NA        rat       NA
#>    technology.Type technology.Color
#> 1:        ONECOLOR         ONECOLOR

getPlatforms(c("GPL1355", "GPL96"))

#>    platform.ID platform.ShortName
#> 1:           1              GPL96
#> 2:           2            GPL1355
#>                                               platform.Name
#> 1: Affymetrix GeneChip Human Genome U133 Array Set HG-U133A
#> 2:             Affymetrix GeneChip Rat Genome 230 2.0 Array
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                        platform.Description
#> 1:                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                The U133 set includes 2 arrays with a total of 44928 entries and was indexed 29-Jan-2002. The set includes over 1,000,000 unique oligonucleotide features covering more than 39,000 transcript variants, which in turn represent greater than 33,000 of the best characterized human genes. Sequences were selected from GenBank, dbEST, and RefSeq. Sequence clusters were created from Build 133 of UniGene (April 20, 2001) and refined by analysis and comparison with a number of other publicly available databases including the Washington University EST trace repository and the University of California, Santa Cruz golden-path human genome database (April 2001 release). In addition, ESTs were analyzed for untrimmed low-quality sequence information, correct orientation, false priming, false clustering, alternative splicing and alternative polyadenylation.  Keywords = high density oligonucleotide array\nFrom GPL96\nLast Updated: Mar 09 2006
#> 2:  The GeneChip Rat Genome 230 2.0 Array is a powerful tool for toxicology, neurobiology, and other applications using rat as a model organism.  - Provides comprehensive coverage of the transcribed rat genome on a single array - Comprised of more than 31,000 probe sets, analyzing over 30,000 transcripts and variants from over 28,000 well-substantiated rat genes - The publicly available draft of the rat genome and leading public rat databases were used to refine sequences and provide a higher quality of data output  All probe sets represented on the GeneChip Rat Expression Set 230 are included on the GeneChip Rat Genome 230 2.0 Array. Sequences used in the design of the GeneChip Rat Genome 230 2.0 Array were selected from GenBank, dbEST, and RefSeq. The sequence clusters were created from the UniGene database (Build 99, June 2002) and then refined by analysis and comparison with the publicly available draft assembly of the rat genome from the Baylor College of Medicine Human Genome Sequencing Center (June 2002).  The GeneChip Rat Genome 230 2.0 Array includes representation of the RefSeq database sequences and probe sets related to sequences and refined EST clusters previously represented on the GeneChip Rat Genome U34 Set.  Oligonucleotide probes complementary to each corresponding sequence are synthesized in situ on the arrays. Eleven pairs of oligonucleotide probes are used to measure the level of transcription of each sequence represented on the GeneChip Rat Genome 230 2.0 Array.  Annotations derived from Affymetrix CSV file dated 6/23/2004\nFrom GPL1355\nLast Updated: May 31 2005
#>    platform.Troubled platform.ExperimentCount platform.GeneCount
#> 1:             FALSE                      412                 NA
#> 2:             FALSE                      288                 NA
#>    platform.ProbeSequenceCount platform.ProbeAlignmentCount
#> 1:                          NA                           NA
#> 2:                          NA                           NA
#>    platform.ProbeGeneCount platform.ElementCount taxon.Name taxon.ID
#> 1:                      NA                    NA      human       NA
#> 2:                      NA                    NA        rat       NA
#>    technology.Type technology.Color
#> 1:        ONECOLOR         ONECOLOR
#> 2:        ONECOLOR         ONECOLOR