Taxa

List taxa filtered by given parameters

getTaxa(
  taxa = NA_character_,
  raw = getOption("gemma.raw", FALSE),
  async = getOption("gemma.async", FALSE),
  memoised = getOption("gemma.memoise", FALSE),
  file = getOption("gemma.file", NA_character_),
  overwrite = getOption("gemma.overwrite", FALSE)
)

Arguments

taxa
Optional, defaults to empty. Limits the result to entities with given identifiers. A list of identifiers, separated by commas (e.g: human, mouse, fly). Identifiers can be the any of the following:

taxon ID

scientific name

common name Retrieval by ID is more efficient. Do not combine different identifiers in one query. For convenience, below is a list of officially supported taxa ID Comm.name Scient.name NcbiID

1 human Homo sapiens 9606 2 mouse Mus musculus 10090 3 rat Rattus norvegicus 10116 11 yeast Saccharomyces cerevisiae 4932 12 zebrafish Danio rerio 7955 13 fly Drosophila melanogaster 7227 14 worm Caenorhabditis elegans 6239

raw
TRUE to receive results as-is from Gemma, or FALSE to enable parsing.

async
TRUE to run the API query on a separate worker, or FALSE to run synchronously. See the async package for details.

memoised
Whether or not to cache results so future requests for the same data will be faster. Use forgetGemmaMemoised to clear the cache.

file
The name of a file to save the results to, or NULL to not write results to a file. If raw == TRUE, the output will be a JSON file. Otherwise, it will be a RDS file.

overwrite
Whether or not to overwrite if a file exists at the specified filename.

Value

An array of taxa (taxon value objects) matching the given identifiers. A 400 error if all identifiers are invalid. An array of all available taxa, if no identifiers were provided.

Examples

getTaxa("human")

#> $url
#> [1] "https://gemma.msl.ubc.ca/rest/v2/taxa/human"
#> 
#> $status_code
#> [1] 400
#> 
#> $type
#> [1] "application/json"
#> 
#> $headers
#>   [1] 48 54 54 50 2f 31 2e 31 20 34 30 30 20 0d 0a 44 61 74 65 3a 20 4d 6f 6e 2c
#>  [26] 20 31 36 20 41 75 67 20 32 30 32 31 20 32 33 3a 34 35 3a 35 36 20 47 4d 54
#>  [51] 0d 0a 53 65 72 76 65 72 3a 20 41 70 61 63 68 65 2f 32 2e 34 2e 36 20 28 43
#>  [76] 65 6e 74 4f 53 29 20 4f 70 65 6e 53 53 4c 2f 31 2e 30 2e 32 6b 2d 66 69 70
#> [101] 73 20 50 48 50 2f 35 2e 34 2e 31 36 0d 0a 53 74 72 69 63 74 2d 54 72 61 6e
#> [126] 73 70 6f 72 74 2d 53 65 63 75 72 69 74 79 3a 20 6d 61 78 2d 61 67 65 3d 36
#> [151] 33 30 37 32 30 30 30 0d 0a 43 6f 6e 74 65 6e 74 2d 54 79 70 65 3a 20 61 70
#> [176] 70 6c 69 63 61 74 69 6f 6e 2f 6a 73 6f 6e 0d 0a 43 6f 6e 74 65 6e 74 2d 4c
#> [201] 65 6e 67 74 68 3a 20 31 36 34 0d 0a 58 2d 43 6f 6e 74 65 6e 74 2d 54 79 70
#> [226] 65 2d 4f 70 74 69 6f 6e 73 3a 20 6e 6f 73 6e 69 66 66 0d 0a 58 2d 46 72 61
#> [251] 6d 65 2d 4f 70 74 69 6f 6e 73 3a 20 53 41 4d 45 4f 52 49 47 49 4e 0d 0a 58
#> [276] 2d 58 53 53 2d 50 72 6f 74 65 63 74 69 6f 6e 3a 20 31 3b 20 6d 6f 64 65 3d
#> [301] 62 6c 6f 63 6b 0d 0a 43 6f 6e 6e 65 63 74 69 6f 6e 3a 20 63 6c 6f 73 65 0d
#> [326] 0a 0d 0a
#> 
#> $modified
#> [1] NA
#> 
#> $times
#>      redirect    namelookup       connect   pretransfer starttransfer 
#>      0.000000      0.001131      0.070684      0.225680      0.335069 
#>         total 
#>      0.337605 
#> 
#> $content
#>   [1] 7b 22 61 70 69 56 65 72 73 69 6f 6e 22 3a 22 32 2e 33 2e 34 22 2c 22 65 72
#>  [26] 72 6f 72 22 3a 7b 22 63 6f 64 65 22 3a 34 30 30 2c 22 6d 65 73 73 61 67 65
#>  [51] 22 3a 22 45 6e 74 69 74 79 20 64 6f 65 73 20 6e 6f 74 20 63 6f 6e 74 61 69
#>  [76] 6e 20 74 68 65 20 67 69 76 65 6e 20 70 72 6f 70 65 72 74 79 2c 20 6f 72 20
#> [101] 74 68 65 20 70 72 6f 76 69 64 65 64 20 76 61 6c 75 65 20 63 61 6e 20 6e 6f
#> [126] 74 20 62 65 20 63 6f 6e 76 65 72 74 65 64 20 74 6f 20 74 68 65 20 70 72 6f
#> [151] 70 65 72 74 79 20 74 79 70 65 2e 22 7d 7d
#> 

getTaxa(c("human", "rat"))

#> $url
#> [1] "https://gemma.msl.ubc.ca/rest/v2/taxa/human%2Crat"
#> 
#> $status_code
#> [1] 400
#> 
#> $type
#> [1] "application/json"
#> 
#> $headers
#>   [1] 48 54 54 50 2f 31 2e 31 20 34 30 30 20 0d 0a 44 61 74 65 3a 20 4d 6f 6e 2c
#>  [26] 20 31 36 20 41 75 67 20 32 30 32 31 20 32 33 3a 34 35 3a 35 36 20 47 4d 54
#>  [51] 0d 0a 53 65 72 76 65 72 3a 20 41 70 61 63 68 65 2f 32 2e 34 2e 36 20 28 43
#>  [76] 65 6e 74 4f 53 29 20 4f 70 65 6e 53 53 4c 2f 31 2e 30 2e 32 6b 2d 66 69 70
#> [101] 73 20 50 48 50 2f 35 2e 34 2e 31 36 0d 0a 53 74 72 69 63 74 2d 54 72 61 6e
#> [126] 73 70 6f 72 74 2d 53 65 63 75 72 69 74 79 3a 20 6d 61 78 2d 61 67 65 3d 36
#> [151] 33 30 37 32 30 30 30 0d 0a 43 6f 6e 74 65 6e 74 2d 54 79 70 65 3a 20 61 70
#> [176] 70 6c 69 63 61 74 69 6f 6e 2f 6a 73 6f 6e 0d 0a 43 6f 6e 74 65 6e 74 2d 4c
#> [201] 65 6e 67 74 68 3a 20 31 36 34 0d 0a 58 2d 43 6f 6e 74 65 6e 74 2d 54 79 70
#> [226] 65 2d 4f 70 74 69 6f 6e 73 3a 20 6e 6f 73 6e 69 66 66 0d 0a 58 2d 46 72 61
#> [251] 6d 65 2d 4f 70 74 69 6f 6e 73 3a 20 53 41 4d 45 4f 52 49 47 49 4e 0d 0a 58
#> [276] 2d 58 53 53 2d 50 72 6f 74 65 63 74 69 6f 6e 3a 20 31 3b 20 6d 6f 64 65 3d
#> [301] 62 6c 6f 63 6b 0d 0a 43 6f 6e 6e 65 63 74 69 6f 6e 3a 20 63 6c 6f 73 65 0d
#> [326] 0a 0d 0a
#> 
#> $modified
#> [1] NA
#> 
#> $times
#>      redirect    namelookup       connect   pretransfer starttransfer 
#>      0.000000      0.001240      0.074313      0.235185      0.323711 
#>         total 
#>      0.325876 
#> 
#> $content
#>   [1] 7b 22 61 70 69 56 65 72 73 69 6f 6e 22 3a 22 32 2e 33 2e 34 22 2c 22 65 72
#>  [26] 72 6f 72 22 3a 7b 22 63 6f 64 65 22 3a 34 30 30 2c 22 6d 65 73 73 61 67 65
#>  [51] 22 3a 22 45 6e 74 69 74 79 20 64 6f 65 73 20 6e 6f 74 20 63 6f 6e 74 61 69
#>  [76] 6e 20 74 68 65 20 67 69 76 65 6e 20 70 72 6f 70 65 72 74 79 2c 20 6f 72 20
#> [101] 74 68 65 20 70 72 6f 76 69 64 65 64 20 76 61 6c 75 65 20 63 61 6e 20 6e 6f
#> [126] 74 20 62 65 20 63 6f 6e 76 65 72 74 65 64 20 74 6f 20 74 68 65 20 70 72 6f
#> [151] 70 65 72 74 79 20 74 79 70 65 2e 22 7d 7d
#>