All taxa

Lists all available taxa.

allTaxa(taxa = NULL, file = NULL, return = TRUE, overwrite = FALSE,
  memoised = FALSE)

Arguments

taxa
Limits the result to entities with given identifiers.

A list of identifiers, separated by commas (e.g: human, mouse, fly). Identifiers can be the any of the following:

taxon ID

scientific name

common name

abbreviation

file
Character. File path. If provided, response will be saved to file

return
Logical. If the response should be returned. Set to false when you only want to save a file

overwrite
Logical. If TRUE, existing files will be overwritten. If FALSE a warning will be thrown and no action is taken.

memoised
Logical. If TRUE a memoised version of the function will be used which is faster for repeated requests. Use forgetGemmaMemoised to clear memory.

Value

List of lists containing experiment object.

Examples

allTaxa()

#> $`Homo sapiens`
#> $`Homo sapiens`$externalDatabase
#> $`Homo sapiens`$externalDatabase$checked
#> [1] FALSE
#> 
#> $`Homo sapiens`$externalDatabase$name
#> [1] "hg38"
#> 
#> $`Homo sapiens`$externalDatabase$id
#> [1] 87
#> 
#> 
#> $`Homo sapiens`$commonName
#> [1] "human"
#> 
#> $`Homo sapiens`$scientificName
#> [1] "Homo sapiens"
#> 
#> $`Homo sapiens`$parentTaxon
#> NULL
#> 
#> $`Homo sapiens`$isGenesUsable
#> [1] TRUE
#> 
#> $`Homo sapiens`$isSpecies
#> [1] TRUE
#> 
#> $`Homo sapiens`$ncbiId
#> [1] 9606
#> 
#> $`Homo sapiens`$abbreviation
#> NULL
#> 
#> $`Homo sapiens`$`_totalInQuery`
#> [1] 0
#> 
#> $`Homo sapiens`$id
#> [1] 1
#> 
#> 
#> $`Mus musculus`
#> $`Mus musculus`$externalDatabase
#> $`Mus musculus`$externalDatabase$checked
#> [1] FALSE
#> 
#> $`Mus musculus`$externalDatabase$name
#> [1] "mm10"
#> 
#> $`Mus musculus`$externalDatabase$id
#> [1] 81
#> 
#> 
#> $`Mus musculus`$commonName
#> [1] "mouse"
#> 
#> $`Mus musculus`$scientificName
#> [1] "Mus musculus"
#> 
#> $`Mus musculus`$parentTaxon
#> NULL
#> 
#> $`Mus musculus`$isGenesUsable
#> [1] TRUE
#> 
#> $`Mus musculus`$isSpecies
#> [1] TRUE
#> 
#> $`Mus musculus`$ncbiId
#> [1] 10090
#> 
#> $`Mus musculus`$abbreviation
#> NULL
#> 
#> $`Mus musculus`$`_totalInQuery`
#> [1] 0
#> 
#> $`Mus musculus`$id
#> [1] 2
#> 
#> 
#> $`Rattus norvegicus`
#> $`Rattus norvegicus`$externalDatabase
#> $`Rattus norvegicus`$externalDatabase$checked
#> [1] FALSE
#> 
#> $`Rattus norvegicus`$externalDatabase$name
#> [1] "rn6"
#> 
#> $`Rattus norvegicus`$externalDatabase$id
#> [1] 86
#> 
#> 
#> $`Rattus norvegicus`$commonName
#> [1] "rat"
#> 
#> $`Rattus norvegicus`$scientificName
#> [1] "Rattus norvegicus"
#> 
#> $`Rattus norvegicus`$parentTaxon
#> NULL
#> 
#> $`Rattus norvegicus`$isGenesUsable
#> [1] TRUE
#> 
#> $`Rattus norvegicus`$isSpecies
#> [1] TRUE
#> 
#> $`Rattus norvegicus`$ncbiId
#> [1] 10116
#> 
#> $`Rattus norvegicus`$abbreviation
#> NULL
#> 
#> $`Rattus norvegicus`$`_totalInQuery`
#> [1] 0
#> 
#> $`Rattus norvegicus`$id
#> [1] 3
#> 
#> 
#> $`Saccharomyces cerevisiae`
#> $`Saccharomyces cerevisiae`$externalDatabase
#> NULL
#> 
#> $`Saccharomyces cerevisiae`$commonName
#> [1] "yeast"
#> 
#> $`Saccharomyces cerevisiae`$scientificName
#> [1] "Saccharomyces cerevisiae"
#> 
#> $`Saccharomyces cerevisiae`$parentTaxon
#> NULL
#> 
#> $`Saccharomyces cerevisiae`$isGenesUsable
#> [1] TRUE
#> 
#> $`Saccharomyces cerevisiae`$isSpecies
#> [1] TRUE
#> 
#> $`Saccharomyces cerevisiae`$ncbiId
#> [1] 4932
#> 
#> $`Saccharomyces cerevisiae`$abbreviation
#> NULL
#> 
#> $`Saccharomyces cerevisiae`$`_totalInQuery`
#> [1] 0
#> 
#> $`Saccharomyces cerevisiae`$id
#> [1] 11
#> 
#> 
#> $`Danio rerio`
#> $`Danio rerio`$externalDatabase
#> NULL
#> 
#> $`Danio rerio`$commonName
#> [1] "zebrafish"
#> 
#> $`Danio rerio`$scientificName
#> [1] "Danio rerio"
#> 
#> $`Danio rerio`$parentTaxon
#> NULL
#> 
#> $`Danio rerio`$isGenesUsable
#> [1] TRUE
#> 
#> $`Danio rerio`$isSpecies
#> [1] TRUE
#> 
#> $`Danio rerio`$ncbiId
#> [1] 7955
#> 
#> $`Danio rerio`$abbreviation
#> NULL
#> 
#> $`Danio rerio`$`_totalInQuery`
#> [1] 0
#> 
#> $`Danio rerio`$id
#> [1] 12
#> 
#> 
#> $`Drosophila melanogaster`
#> $`Drosophila melanogaster`$externalDatabase
#> NULL
#> 
#> $`Drosophila melanogaster`$commonName
#> [1] "fly"
#> 
#> $`Drosophila melanogaster`$scientificName
#> [1] "Drosophila melanogaster"
#> 
#> $`Drosophila melanogaster`$parentTaxon
#> NULL
#> 
#> $`Drosophila melanogaster`$isGenesUsable
#> [1] FALSE
#> 
#> $`Drosophila melanogaster`$isSpecies
#> [1] TRUE
#> 
#> $`Drosophila melanogaster`$ncbiId
#> [1] 7227
#> 
#> $`Drosophila melanogaster`$abbreviation
#> NULL
#> 
#> $`Drosophila melanogaster`$`_totalInQuery`
#> [1] 0
#> 
#> $`Drosophila melanogaster`$id
#> [1] 13
#> 
#> 
#> $`Caenorhabditis elegans`
#> $`Caenorhabditis elegans`$externalDatabase
#> NULL
#> 
#> $`Caenorhabditis elegans`$commonName
#> [1] "worm"
#> 
#> $`Caenorhabditis elegans`$scientificName
#> [1] "Caenorhabditis elegans"
#> 
#> $`Caenorhabditis elegans`$parentTaxon
#> NULL
#> 
#> $`Caenorhabditis elegans`$isGenesUsable
#> [1] TRUE
#> 
#> $`Caenorhabditis elegans`$isSpecies
#> [1] TRUE
#> 
#> $`Caenorhabditis elegans`$ncbiId
#> [1] 6239
#> 
#> $`Caenorhabditis elegans`$abbreviation
#> NULL
#> 
#> $`Caenorhabditis elegans`$`_totalInQuery`
#> [1] 0
#> 
#> $`Caenorhabditis elegans`$id
#> [1] 14
#> 
#> 
#> $`synthetic construct`
#> $`synthetic construct`$externalDatabase
#> NULL
#> 
#> $`synthetic construct`$commonName
#> [1] "synthetic construct"
#> 
#> $`synthetic construct`$scientificName
#> [1] "synthetic construct"
#> 
#> $`synthetic construct`$parentTaxon
#> NULL
#> 
#> $`synthetic construct`$isGenesUsable
#> [1] FALSE
#> 
#> $`synthetic construct`$isSpecies
#> [1] TRUE
#> 
#> $`synthetic construct`$ncbiId
#> NULL
#> 
#> $`synthetic construct`$abbreviation
#> NULL
#> 
#> $`synthetic construct`$`_totalInQuery`
#> [1] 0
#> 
#> $`synthetic construct`$id
#> [1] 19
#> 
#> 
#> $`Schizosaccharomyces pombe`
#> $`Schizosaccharomyces pombe`$externalDatabase
#> NULL
#> 
#> $`Schizosaccharomyces pombe`$commonName
#> [1] "fission yeast"
#> 
#> $`Schizosaccharomyces pombe`$scientificName
#> [1] "Schizosaccharomyces pombe"
#> 
#> $`Schizosaccharomyces pombe`$parentTaxon
#> NULL
#> 
#> $`Schizosaccharomyces pombe`$isGenesUsable
#> [1] FALSE
#> 
#> $`Schizosaccharomyces pombe`$isSpecies
#> [1] TRUE
#> 
#> $`Schizosaccharomyces pombe`$ncbiId
#> NULL
#> 
#> $`Schizosaccharomyces pombe`$abbreviation
#> NULL
#> 
#> $`Schizosaccharomyces pombe`$`_totalInQuery`
#> [1] 0
#> 
#> $`Schizosaccharomyces pombe`$id
#> [1] 33
#> 
#> 
#> $`Macaca mulatta`
#> $`Macaca mulatta`$externalDatabase
#> NULL
#> 
#> $`Macaca mulatta`$commonName
#> NULL
#> 
#> $`Macaca mulatta`$scientificName
#> [1] "Macaca mulatta"
#> 
#> $`Macaca mulatta`$parentTaxon
#> NULL
#> 
#> $`Macaca mulatta`$isGenesUsable
#> [1] TRUE
#> 
#> $`Macaca mulatta`$isSpecies
#> [1] TRUE
#> 
#> $`Macaca mulatta`$ncbiId
#> NULL
#> 
#> $`Macaca mulatta`$abbreviation
#> NULL
#> 
#> $`Macaca mulatta`$`_totalInQuery`
#> [1] 0
#> 
#> $`Macaca mulatta`$id
#> [1] 78
#> 
#> 
#> $`Rattus rattus`
#> $`Rattus rattus`$externalDatabase
#> NULL
#> 
#> $`Rattus rattus`$commonName
#> NULL
#> 
#> $`Rattus rattus`$scientificName
#> [1] "Rattus rattus"
#> 
#> $`Rattus rattus`$parentTaxon
#> NULL
#> 
#> $`Rattus rattus`$isGenesUsable
#> [1] FALSE
#> 
#> $`Rattus rattus`$isSpecies
#> [1] TRUE
#> 
#> $`Rattus rattus`$ncbiId
#> NULL
#> 
#> $`Rattus rattus`$abbreviation
#> NULL
#> 
#> $`Rattus rattus`$`_totalInQuery`
#> [1] 0
#> 
#> $`Rattus rattus`$id
#> [1] 79
#> 
#> 

allTaxa(c('human','mouse'))

#> $`Homo sapiens`
#> $`Homo sapiens`$externalDatabase
#> $`Homo sapiens`$externalDatabase$checked
#> [1] FALSE
#> 
#> $`Homo sapiens`$externalDatabase$name
#> [1] "hg38"
#> 
#> $`Homo sapiens`$externalDatabase$id
#> [1] 87
#> 
#> 
#> $`Homo sapiens`$commonName
#> [1] "human"
#> 
#> $`Homo sapiens`$scientificName
#> [1] "Homo sapiens"
#> 
#> $`Homo sapiens`$parentTaxon
#> NULL
#> 
#> $`Homo sapiens`$isGenesUsable
#> [1] TRUE
#> 
#> $`Homo sapiens`$isSpecies
#> [1] TRUE
#> 
#> $`Homo sapiens`$ncbiId
#> [1] 9606
#> 
#> $`Homo sapiens`$abbreviation
#> NULL
#> 
#> $`Homo sapiens`$`_totalInQuery`
#> [1] 0
#> 
#> $`Homo sapiens`$id
#> [1] 1
#> 
#> 
#> $`Mus musculus`
#> $`Mus musculus`$externalDatabase
#> $`Mus musculus`$externalDatabase$checked
#> [1] FALSE
#> 
#> $`Mus musculus`$externalDatabase$name
#> [1] "mm10"
#> 
#> $`Mus musculus`$externalDatabase$id
#> [1] 81
#> 
#> 
#> $`Mus musculus`$commonName
#> [1] "mouse"
#> 
#> $`Mus musculus`$scientificName
#> [1] "Mus musculus"
#> 
#> $`Mus musculus`$parentTaxon
#> NULL
#> 
#> $`Mus musculus`$isGenesUsable
#> [1] TRUE
#> 
#> $`Mus musculus`$isSpecies
#> [1] TRUE
#> 
#> $`Mus musculus`$ncbiId
#> [1] 10090
#> 
#> $`Mus musculus`$abbreviation
#> NULL
#> 
#> $`Mus musculus`$`_totalInQuery`
#> [1] 0
#> 
#> $`Mus musculus`$id
#> [1] 2
#> 
#>