All taxa

Lists all available taxa.

allTaxa(taxa = NULL, file = NULL, return = TRUE, overwrite = FALSE,
  memoised = FALSE)

Arguments

taxa
Limits the result to entities with given identifiers.

A list of identifiers, separated by commas (e.g: human, mouse, fly). Identifiers can be the any of the following:

taxon ID

scientific name

common name

abbreviation

file
Character. File path. If provided, response will be saved to file

return
Logical. If the response should be returned. Set to false when you only want to save a file

overwrite
Logical. If TRUE, existing files will be overwritten. If FALSE a warning will be thrown and no action is taken.

memoised
Logical. If TRUE a memoised version of the function will be used which is faster for repeated requests. Use forgetGemmaMemoised to clear memory.

Value

List of lists containing experiment object.

Examples

allTaxa()

#> $`Homo sapiens`
#> $`Homo sapiens`$commonName
#> [1] "human"
#> 
#> $`Homo sapiens`$externalDatabase
#> $`Homo sapiens`$externalDatabase$checked
#> [1] FALSE
#> 
#> $`Homo sapiens`$externalDatabase$name
#> [1] "hg19"
#> 
#> $`Homo sapiens`$externalDatabase$id
#> [1] 65
#> 
#> 
#> $`Homo sapiens`$isSpecies
#> [1] TRUE
#> 
#> $`Homo sapiens`$isGenesUsable
#> [1] TRUE
#> 
#> $`Homo sapiens`$scientificName
#> [1] "Homo sapiens"
#> 
#> $`Homo sapiens`$parentTaxon
#> NULL
#> 
#> $`Homo sapiens`$ncbiId
#> [1] 9606
#> 
#> $`Homo sapiens`$abbreviation
#> NULL
#> 
#> $`Homo sapiens`$id
#> [1] 1
#> 
#> 
#> $`Mus musculus`
#> $`Mus musculus`$commonName
#> [1] "mouse"
#> 
#> $`Mus musculus`$externalDatabase
#> $`Mus musculus`$externalDatabase$checked
#> [1] FALSE
#> 
#> $`Mus musculus`$externalDatabase$name
#> [1] "mm10"
#> 
#> $`Mus musculus`$externalDatabase$id
#> [1] 81
#> 
#> 
#> $`Mus musculus`$isSpecies
#> [1] TRUE
#> 
#> $`Mus musculus`$isGenesUsable
#> [1] TRUE
#> 
#> $`Mus musculus`$scientificName
#> [1] "Mus musculus"
#> 
#> $`Mus musculus`$parentTaxon
#> NULL
#> 
#> $`Mus musculus`$ncbiId
#> [1] 10090
#> 
#> $`Mus musculus`$abbreviation
#> NULL
#> 
#> $`Mus musculus`$id
#> [1] 2
#> 
#> 
#> $`Rattus norvegicus`
#> $`Rattus norvegicus`$commonName
#> [1] "rat"
#> 
#> $`Rattus norvegicus`$externalDatabase
#> $`Rattus norvegicus`$externalDatabase$checked
#> [1] FALSE
#> 
#> $`Rattus norvegicus`$externalDatabase$name
#> [1] "rn4"
#> 
#> $`Rattus norvegicus`$externalDatabase$id
#> [1] 13
#> 
#> 
#> $`Rattus norvegicus`$isSpecies
#> [1] TRUE
#> 
#> $`Rattus norvegicus`$isGenesUsable
#> [1] TRUE
#> 
#> $`Rattus norvegicus`$scientificName
#> [1] "Rattus norvegicus"
#> 
#> $`Rattus norvegicus`$parentTaxon
#> NULL
#> 
#> $`Rattus norvegicus`$ncbiId
#> [1] 10116
#> 
#> $`Rattus norvegicus`$abbreviation
#> NULL
#> 
#> $`Rattus norvegicus`$id
#> [1] 3
#> 
#> 
#> $Salmonidae
#> $Salmonidae$commonName
#> [1] "salmonid"
#> 
#> $Salmonidae$externalDatabase
#> NULL
#> 
#> $Salmonidae$isSpecies
#> [1] FALSE
#> 
#> $Salmonidae$isGenesUsable
#> [1] TRUE
#> 
#> $Salmonidae$scientificName
#> [1] "Salmonidae"
#> 
#> $Salmonidae$parentTaxon
#> NULL
#> 
#> $Salmonidae$ncbiId
#> [1] 8015
#> 
#> $Salmonidae$abbreviation
#> NULL
#> 
#> $Salmonidae$id
#> [1] 4
#> 
#> 
#> $`Salmo salar`
#> $`Salmo salar`$commonName
#> [1] "atlantic salmon"
#> 
#> $`Salmo salar`$externalDatabase
#> NULL
#> 
#> $`Salmo salar`$isSpecies
#> [1] TRUE
#> 
#> $`Salmo salar`$isGenesUsable
#> [1] FALSE
#> 
#> $`Salmo salar`$scientificName
#> [1] "Salmo salar"
#> 
#> $`Salmo salar`$parentTaxon
#> $`Salmo salar`$parentTaxon$commonName
#> [1] "salmonid"
#> 
#> $`Salmo salar`$parentTaxon$externalDatabase
#> NULL
#> 
#> $`Salmo salar`$parentTaxon$isSpecies
#> [1] FALSE
#> 
#> $`Salmo salar`$parentTaxon$isGenesUsable
#> [1] TRUE
#> 
#> $`Salmo salar`$parentTaxon$scientificName
#> [1] "Salmonidae"
#> 
#> $`Salmo salar`$parentTaxon$parentTaxon
#> NULL
#> 
#> $`Salmo salar`$parentTaxon$ncbiId
#> [1] 8015
#> 
#> $`Salmo salar`$parentTaxon$abbreviation
#> NULL
#> 
#> $`Salmo salar`$parentTaxon$id
#> [1] 4
#> 
#> 
#> $`Salmo salar`$ncbiId
#> [1] 8030
#> 
#> $`Salmo salar`$abbreviation
#> [1] "ssal"
#> 
#> $`Salmo salar`$id
#> [1] 5
#> 
#> 
#> $`Oncorhynchus mykiss`
#> $`Oncorhynchus mykiss`$commonName
#> [1] "rainbow trout"
#> 
#> $`Oncorhynchus mykiss`$externalDatabase
#> NULL
#> 
#> $`Oncorhynchus mykiss`$isSpecies
#> [1] TRUE
#> 
#> $`Oncorhynchus mykiss`$isGenesUsable
#> [1] FALSE
#> 
#> $`Oncorhynchus mykiss`$scientificName
#> [1] "Oncorhynchus mykiss"
#> 
#> $`Oncorhynchus mykiss`$parentTaxon
#> $`Oncorhynchus mykiss`$parentTaxon$commonName
#> [1] "salmonid"
#> 
#> $`Oncorhynchus mykiss`$parentTaxon$externalDatabase
#> NULL
#> 
#> $`Oncorhynchus mykiss`$parentTaxon$isSpecies
#> [1] FALSE
#> 
#> $`Oncorhynchus mykiss`$parentTaxon$isGenesUsable
#> [1] TRUE
#> 
#> $`Oncorhynchus mykiss`$parentTaxon$scientificName
#> [1] "Salmonidae"
#> 
#> $`Oncorhynchus mykiss`$parentTaxon$parentTaxon
#> NULL
#> 
#> $`Oncorhynchus mykiss`$parentTaxon$ncbiId
#> [1] 8015
#> 
#> $`Oncorhynchus mykiss`$parentTaxon$abbreviation
#> NULL
#> 
#> $`Oncorhynchus mykiss`$parentTaxon$id
#> [1] 4
#> 
#> 
#> $`Oncorhynchus mykiss`$ncbiId
#> [1] 8022
#> 
#> $`Oncorhynchus mykiss`$abbreviation
#> [1] "omyk"
#> 
#> $`Oncorhynchus mykiss`$id
#> [1] 6
#> 
#> 
#> $`Coregonus clupeaformis`
#> $`Coregonus clupeaformis`$commonName
#> [1] "whitefish"
#> 
#> $`Coregonus clupeaformis`$externalDatabase
#> NULL
#> 
#> $`Coregonus clupeaformis`$isSpecies
#> [1] TRUE
#> 
#> $`Coregonus clupeaformis`$isGenesUsable
#> [1] FALSE
#> 
#> $`Coregonus clupeaformis`$scientificName
#> [1] "Coregonus clupeaformis"
#> 
#> $`Coregonus clupeaformis`$parentTaxon
#> $`Coregonus clupeaformis`$parentTaxon$commonName
#> [1] "salmonid"
#> 
#> $`Coregonus clupeaformis`$parentTaxon$externalDatabase
#> NULL
#> 
#> $`Coregonus clupeaformis`$parentTaxon$isSpecies
#> [1] FALSE
#> 
#> $`Coregonus clupeaformis`$parentTaxon$isGenesUsable
#> [1] TRUE
#> 
#> $`Coregonus clupeaformis`$parentTaxon$scientificName
#> [1] "Salmonidae"
#> 
#> $`Coregonus clupeaformis`$parentTaxon$parentTaxon
#> NULL
#> 
#> $`Coregonus clupeaformis`$parentTaxon$ncbiId
#> [1] 8015
#> 
#> $`Coregonus clupeaformis`$parentTaxon$abbreviation
#> NULL
#> 
#> $`Coregonus clupeaformis`$parentTaxon$id
#> [1] 4
#> 
#> 
#> $`Coregonus clupeaformis`$ncbiId
#> [1] 59861
#> 
#> $`Coregonus clupeaformis`$abbreviation
#> [1] "cclu"
#> 
#> $`Coregonus clupeaformis`$id
#> [1] 7
#> 
#> 
#> $`Oncorhynchus tshawytscha`
#> $`Oncorhynchus tshawytscha`$commonName
#> [1] "chinook salmon"
#> 
#> $`Oncorhynchus tshawytscha`$externalDatabase
#> NULL
#> 
#> $`Oncorhynchus tshawytscha`$isSpecies
#> [1] TRUE
#> 
#> $`Oncorhynchus tshawytscha`$isGenesUsable
#> [1] FALSE
#> 
#> $`Oncorhynchus tshawytscha`$scientificName
#> [1] "Oncorhynchus tshawytscha"
#> 
#> $`Oncorhynchus tshawytscha`$parentTaxon
#> $`Oncorhynchus tshawytscha`$parentTaxon$commonName
#> [1] "salmonid"
#> 
#> $`Oncorhynchus tshawytscha`$parentTaxon$externalDatabase
#> NULL
#> 
#> $`Oncorhynchus tshawytscha`$parentTaxon$isSpecies
#> [1] FALSE
#> 
#> $`Oncorhynchus tshawytscha`$parentTaxon$isGenesUsable
#> [1] TRUE
#> 
#> $`Oncorhynchus tshawytscha`$parentTaxon$scientificName
#> [1] "Salmonidae"
#> 
#> $`Oncorhynchus tshawytscha`$parentTaxon$parentTaxon
#> NULL
#> 
#> $`Oncorhynchus tshawytscha`$parentTaxon$ncbiId
#> [1] 8015
#> 
#> $`Oncorhynchus tshawytscha`$parentTaxon$abbreviation
#> NULL
#> 
#> $`Oncorhynchus tshawytscha`$parentTaxon$id
#> [1] 4
#> 
#> 
#> $`Oncorhynchus tshawytscha`$ncbiId
#> [1] 74940
#> 
#> $`Oncorhynchus tshawytscha`$abbreviation
#> [1] "otsh"
#> 
#> $`Oncorhynchus tshawytscha`$id
#> [1] 8
#> 
#> 
#> $`Oncorhynchus keta`
#> $`Oncorhynchus keta`$commonName
#> [1] "chum salmon"
#> 
#> $`Oncorhynchus keta`$externalDatabase
#> NULL
#> 
#> $`Oncorhynchus keta`$isSpecies
#> [1] TRUE
#> 
#> $`Oncorhynchus keta`$isGenesUsable
#> [1] FALSE
#> 
#> $`Oncorhynchus keta`$scientificName
#> [1] "Oncorhynchus keta"
#> 
#> $`Oncorhynchus keta`$parentTaxon
#> $`Oncorhynchus keta`$parentTaxon$commonName
#> [1] "salmonid"
#> 
#> $`Oncorhynchus keta`$parentTaxon$externalDatabase
#> NULL
#> 
#> $`Oncorhynchus keta`$parentTaxon$isSpecies
#> [1] FALSE
#> 
#> $`Oncorhynchus keta`$parentTaxon$isGenesUsable
#> [1] TRUE
#> 
#> $`Oncorhynchus keta`$parentTaxon$scientificName
#> [1] "Salmonidae"
#> 
#> $`Oncorhynchus keta`$parentTaxon$parentTaxon
#> NULL
#> 
#> $`Oncorhynchus keta`$parentTaxon$ncbiId
#> [1] 8015
#> 
#> $`Oncorhynchus keta`$parentTaxon$abbreviation
#> NULL
#> 
#> $`Oncorhynchus keta`$parentTaxon$id
#> [1] 4
#> 
#> 
#> $`Oncorhynchus keta`$ncbiId
#> [1] 8018
#> 
#> $`Oncorhynchus keta`$abbreviation
#> NULL
#> 
#> $`Oncorhynchus keta`$id
#> [1] 9
#> 
#> 
#> $`Osmerus mordax`
#> $`Osmerus mordax`$commonName
#> [1] "rainbow smelt"
#> 
#> $`Osmerus mordax`$externalDatabase
#> NULL
#> 
#> $`Osmerus mordax`$isSpecies
#> [1] TRUE
#> 
#> $`Osmerus mordax`$isGenesUsable
#> [1] FALSE
#> 
#> $`Osmerus mordax`$scientificName
#> [1] "Osmerus mordax"
#> 
#> $`Osmerus mordax`$parentTaxon
#> $`Osmerus mordax`$parentTaxon$commonName
#> [1] "salmonid"
#> 
#> $`Osmerus mordax`$parentTaxon$externalDatabase
#> NULL
#> 
#> $`Osmerus mordax`$parentTaxon$isSpecies
#> [1] FALSE
#> 
#> $`Osmerus mordax`$parentTaxon$isGenesUsable
#> [1] TRUE
#> 
#> $`Osmerus mordax`$parentTaxon$scientificName
#> [1] "Salmonidae"
#> 
#> $`Osmerus mordax`$parentTaxon$parentTaxon
#> NULL
#> 
#> $`Osmerus mordax`$parentTaxon$ncbiId
#> [1] 8015
#> 
#> $`Osmerus mordax`$parentTaxon$abbreviation
#> NULL
#> 
#> $`Osmerus mordax`$parentTaxon$id
#> [1] 4
#> 
#> 
#> $`Osmerus mordax`$ncbiId
#> [1] 8014
#> 
#> $`Osmerus mordax`$abbreviation
#> [1] "omor"
#> 
#> $`Osmerus mordax`$id
#> [1] 10
#> 
#> 
#> $`Saccharomyces cerevisiae`
#> $`Saccharomyces cerevisiae`$commonName
#> [1] "yeast"
#> 
#> $`Saccharomyces cerevisiae`$externalDatabase
#> NULL
#> 
#> $`Saccharomyces cerevisiae`$isSpecies
#> [1] TRUE
#> 
#> $`Saccharomyces cerevisiae`$isGenesUsable
#> [1] TRUE
#> 
#> $`Saccharomyces cerevisiae`$scientificName
#> [1] "Saccharomyces cerevisiae"
#> 
#> $`Saccharomyces cerevisiae`$parentTaxon
#> NULL
#> 
#> $`Saccharomyces cerevisiae`$ncbiId
#> [1] 4932
#> 
#> $`Saccharomyces cerevisiae`$abbreviation
#> NULL
#> 
#> $`Saccharomyces cerevisiae`$id
#> [1] 11
#> 
#> 
#> $`Danio rerio`
#> $`Danio rerio`$commonName
#> [1] "zebrafish"
#> 
#> $`Danio rerio`$externalDatabase
#> NULL
#> 
#> $`Danio rerio`$isSpecies
#> [1] TRUE
#> 
#> $`Danio rerio`$isGenesUsable
#> [1] TRUE
#> 
#> $`Danio rerio`$scientificName
#> [1] "Danio rerio"
#> 
#> $`Danio rerio`$parentTaxon
#> NULL
#> 
#> $`Danio rerio`$ncbiId
#> [1] 7955
#> 
#> $`Danio rerio`$abbreviation
#> NULL
#> 
#> $`Danio rerio`$id
#> [1] 12
#> 
#> 
#> $`Drosophila melanogaster`
#> $`Drosophila melanogaster`$commonName
#> [1] "fly"
#> 
#> $`Drosophila melanogaster`$externalDatabase
#> NULL
#> 
#> $`Drosophila melanogaster`$isSpecies
#> [1] TRUE
#> 
#> $`Drosophila melanogaster`$isGenesUsable
#> [1] FALSE
#> 
#> $`Drosophila melanogaster`$scientificName
#> [1] "Drosophila melanogaster"
#> 
#> $`Drosophila melanogaster`$parentTaxon
#> $`Drosophila melanogaster`$parentTaxon$commonName
#> [1] "drosophila"
#> 
#> $`Drosophila melanogaster`$parentTaxon$externalDatabase
#> NULL
#> 
#> $`Drosophila melanogaster`$parentTaxon$isSpecies
#> [1] FALSE
#> 
#> $`Drosophila melanogaster`$parentTaxon$isGenesUsable
#> [1] TRUE
#> 
#> $`Drosophila melanogaster`$parentTaxon$scientificName
#> [1] "Drosophila"
#> 
#> $`Drosophila melanogaster`$parentTaxon$parentTaxon
#> NULL
#> 
#> $`Drosophila melanogaster`$parentTaxon$ncbiId
#> NULL
#> 
#> $`Drosophila melanogaster`$parentTaxon$abbreviation
#> NULL
#> 
#> $`Drosophila melanogaster`$parentTaxon$id
#> [1] 26
#> 
#> 
#> $`Drosophila melanogaster`$ncbiId
#> [1] 7227
#> 
#> $`Drosophila melanogaster`$abbreviation
#> NULL
#> 
#> $`Drosophila melanogaster`$id
#> [1] 13
#> 
#> 
#> $`Caenorhabditis elegans`
#> $`Caenorhabditis elegans`$commonName
#> [1] "worm"
#> 
#> $`Caenorhabditis elegans`$externalDatabase
#> NULL
#> 
#> $`Caenorhabditis elegans`$isSpecies
#> [1] TRUE
#> 
#> $`Caenorhabditis elegans`$isGenesUsable
#> [1] TRUE
#> 
#> $`Caenorhabditis elegans`$scientificName
#> [1] "Caenorhabditis elegans"
#> 
#> $`Caenorhabditis elegans`$parentTaxon
#> NULL
#> 
#> $`Caenorhabditis elegans`$ncbiId
#> [1] 6239
#> 
#> $`Caenorhabditis elegans`$abbreviation
#> NULL
#> 
#> $`Caenorhabditis elegans`$id
#> [1] 14
#> 
#> 
#> $`Oncorhynchus kisutch`
#> $`Oncorhynchus kisutch`$commonName
#> [1] "coho salmon"
#> 
#> $`Oncorhynchus kisutch`$externalDatabase
#> NULL
#> 
#> $`Oncorhynchus kisutch`$isSpecies
#> [1] TRUE
#> 
#> $`Oncorhynchus kisutch`$isGenesUsable
#> [1] FALSE
#> 
#> $`Oncorhynchus kisutch`$scientificName
#> [1] "Oncorhynchus kisutch"
#> 
#> $`Oncorhynchus kisutch`$parentTaxon
#> $`Oncorhynchus kisutch`$parentTaxon$commonName
#> [1] "salmonid"
#> 
#> $`Oncorhynchus kisutch`$parentTaxon$externalDatabase
#> NULL
#> 
#> $`Oncorhynchus kisutch`$parentTaxon$isSpecies
#> [1] FALSE
#> 
#> $`Oncorhynchus kisutch`$parentTaxon$isGenesUsable
#> [1] TRUE
#> 
#> $`Oncorhynchus kisutch`$parentTaxon$scientificName
#> [1] "Salmonidae"
#> 
#> $`Oncorhynchus kisutch`$parentTaxon$parentTaxon
#> NULL
#> 
#> $`Oncorhynchus kisutch`$parentTaxon$ncbiId
#> [1] 8015
#> 
#> $`Oncorhynchus kisutch`$parentTaxon$abbreviation
#> NULL
#> 
#> $`Oncorhynchus kisutch`$parentTaxon$id
#> [1] 4
#> 
#> 
#> $`Oncorhynchus kisutch`$ncbiId
#> [1] 8019
#> 
#> $`Oncorhynchus kisutch`$abbreviation
#> NULL
#> 
#> $`Oncorhynchus kisutch`$id
#> [1] 15
#> 
#> 
#> $`Oncorhynchus gorbuscha`
#> $`Oncorhynchus gorbuscha`$commonName
#> [1] "pink salmon"
#> 
#> $`Oncorhynchus gorbuscha`$externalDatabase
#> NULL
#> 
#> $`Oncorhynchus gorbuscha`$isSpecies
#> [1] TRUE
#> 
#> $`Oncorhynchus gorbuscha`$isGenesUsable
#> [1] FALSE
#> 
#> $`Oncorhynchus gorbuscha`$scientificName
#> [1] "Oncorhynchus gorbuscha"
#> 
#> $`Oncorhynchus gorbuscha`$parentTaxon
#> $`Oncorhynchus gorbuscha`$parentTaxon$commonName
#> [1] "salmonid"
#> 
#> $`Oncorhynchus gorbuscha`$parentTaxon$externalDatabase
#> NULL
#> 
#> $`Oncorhynchus gorbuscha`$parentTaxon$isSpecies
#> [1] FALSE
#> 
#> $`Oncorhynchus gorbuscha`$parentTaxon$isGenesUsable
#> [1] TRUE
#> 
#> $`Oncorhynchus gorbuscha`$parentTaxon$scientificName
#> [1] "Salmonidae"
#> 
#> $`Oncorhynchus gorbuscha`$parentTaxon$parentTaxon
#> NULL
#> 
#> $`Oncorhynchus gorbuscha`$parentTaxon$ncbiId
#> [1] 8015
#> 
#> $`Oncorhynchus gorbuscha`$parentTaxon$abbreviation
#> NULL
#> 
#> $`Oncorhynchus gorbuscha`$parentTaxon$id
#> [1] 4
#> 
#> 
#> $`Oncorhynchus gorbuscha`$ncbiId
#> [1] 8017
#> 
#> $`Oncorhynchus gorbuscha`$abbreviation
#> NULL
#> 
#> $`Oncorhynchus gorbuscha`$id
#> [1] 16
#> 
#> 
#> $`Oncorhynchus nerka`
#> $`Oncorhynchus nerka`$commonName
#> [1] "sockeye salmon"
#> 
#> $`Oncorhynchus nerka`$externalDatabase
#> NULL
#> 
#> $`Oncorhynchus nerka`$isSpecies
#> [1] TRUE
#> 
#> $`Oncorhynchus nerka`$isGenesUsable
#> [1] FALSE
#> 
#> $`Oncorhynchus nerka`$scientificName
#> [1] "Oncorhynchus nerka"
#> 
#> $`Oncorhynchus nerka`$parentTaxon
#> $`Oncorhynchus nerka`$parentTaxon$commonName
#> [1] "salmonid"
#> 
#> $`Oncorhynchus nerka`$parentTaxon$externalDatabase
#> NULL
#> 
#> $`Oncorhynchus nerka`$parentTaxon$isSpecies
#> [1] FALSE
#> 
#> $`Oncorhynchus nerka`$parentTaxon$isGenesUsable
#> [1] TRUE
#> 
#> $`Oncorhynchus nerka`$parentTaxon$scientificName
#> [1] "Salmonidae"
#> 
#> $`Oncorhynchus nerka`$parentTaxon$parentTaxon
#> NULL
#> 
#> $`Oncorhynchus nerka`$parentTaxon$ncbiId
#> [1] 8015
#> 
#> $`Oncorhynchus nerka`$parentTaxon$abbreviation
#> NULL
#> 
#> $`Oncorhynchus nerka`$parentTaxon$id
#> [1] 4
#> 
#> 
#> $`Oncorhynchus nerka`$ncbiId
#> [1] 8023
#> 
#> $`Oncorhynchus nerka`$abbreviation
#> NULL
#> 
#> $`Oncorhynchus nerka`$id
#> [1] 17
#> 
#> 
#> $`Onthophagus taurus`
#> $`Onthophagus taurus`$commonName
#> NULL
#> 
#> $`Onthophagus taurus`$externalDatabase
#> NULL
#> 
#> $`Onthophagus taurus`$isSpecies
#> [1] TRUE
#> 
#> $`Onthophagus taurus`$isGenesUsable
#> [1] FALSE
#> 
#> $`Onthophagus taurus`$scientificName
#> [1] "Onthophagus taurus"
#> 
#> $`Onthophagus taurus`$parentTaxon
#> NULL
#> 
#> $`Onthophagus taurus`$ncbiId
#> NULL
#> 
#> $`Onthophagus taurus`$abbreviation
#> NULL
#> 
#> $`Onthophagus taurus`$id
#> [1] 18
#> 
#> 
#> $`synthetic construct`
#> $`synthetic construct`$commonName
#> NULL
#> 
#> $`synthetic construct`$externalDatabase
#> NULL
#> 
#> $`synthetic construct`$isSpecies
#> [1] TRUE
#> 
#> $`synthetic construct`$isGenesUsable
#> [1] FALSE
#> 
#> $`synthetic construct`$scientificName
#> [1] "synthetic construct"
#> 
#> $`synthetic construct`$parentTaxon
#> NULL
#> 
#> $`synthetic construct`$ncbiId
#> NULL
#> 
#> $`synthetic construct`$abbreviation
#> NULL
#> 
#> $`synthetic construct`$id
#> [1] 19
#> 
#> 
#> $`Drosophila simulans`
#> $`Drosophila simulans`$commonName
#> [1] "d.simulans"
#> 
#> $`Drosophila simulans`$externalDatabase
#> NULL
#> 
#> $`Drosophila simulans`$isSpecies
#> [1] TRUE
#> 
#> $`Drosophila simulans`$isGenesUsable
#> [1] FALSE
#> 
#> $`Drosophila simulans`$scientificName
#> [1] "Drosophila simulans"
#> 
#> $`Drosophila simulans`$parentTaxon
#> $`Drosophila simulans`$parentTaxon$commonName
#> [1] "drosophila"
#> 
#> $`Drosophila simulans`$parentTaxon$externalDatabase
#> NULL
#> 
#> $`Drosophila simulans`$parentTaxon$isSpecies
#> [1] FALSE
#> 
#> $`Drosophila simulans`$parentTaxon$isGenesUsable
#> [1] TRUE
#> 
#> $`Drosophila simulans`$parentTaxon$scientificName
#> [1] "Drosophila"
#> 
#> $`Drosophila simulans`$parentTaxon$parentTaxon
#> NULL
#> 
#> $`Drosophila simulans`$parentTaxon$ncbiId
#> NULL
#> 
#> $`Drosophila simulans`$parentTaxon$abbreviation
#> NULL
#> 
#> $`Drosophila simulans`$parentTaxon$id
#> [1] 26
#> 
#> 
#> $`Drosophila simulans`$ncbiId
#> NULL
#> 
#> $`Drosophila simulans`$abbreviation
#> NULL
#> 
#> $`Drosophila simulans`$id
#> [1] 20
#> 
#> 
#> $`Drosophila santomea`
#> $`Drosophila santomea`$commonName
#> [1] "d.santomea"
#> 
#> $`Drosophila santomea`$externalDatabase
#> NULL
#> 
#> $`Drosophila santomea`$isSpecies
#> [1] TRUE
#> 
#> $`Drosophila santomea`$isGenesUsable
#> [1] FALSE
#> 
#> $`Drosophila santomea`$scientificName
#> [1] "Drosophila santomea"
#> 
#> $`Drosophila santomea`$parentTaxon
#> $`Drosophila santomea`$parentTaxon$commonName
#> [1] "drosophila"
#> 
#> $`Drosophila santomea`$parentTaxon$externalDatabase
#> NULL
#> 
#> $`Drosophila santomea`$parentTaxon$isSpecies
#> [1] FALSE
#> 
#> $`Drosophila santomea`$parentTaxon$isGenesUsable
#> [1] TRUE
#> 
#> $`Drosophila santomea`$parentTaxon$scientificName
#> [1] "Drosophila"
#> 
#> $`Drosophila santomea`$parentTaxon$parentTaxon
#> NULL
#> 
#> $`Drosophila santomea`$parentTaxon$ncbiId
#> NULL
#> 
#> $`Drosophila santomea`$parentTaxon$abbreviation
#> NULL
#> 
#> $`Drosophila santomea`$parentTaxon$id
#> [1] 26
#> 
#> 
#> $`Drosophila santomea`$ncbiId
#> NULL
#> 
#> $`Drosophila santomea`$abbreviation
#> NULL
#> 
#> $`Drosophila santomea`$id
#> [1] 21
#> 
#> 
#> $`Drosophila yakuba`
#> $`Drosophila yakuba`$commonName
#> [1] "d.yakuba"
#> 
#> $`Drosophila yakuba`$externalDatabase
#> NULL
#> 
#> $`Drosophila yakuba`$isSpecies
#> [1] TRUE
#> 
#> $`Drosophila yakuba`$isGenesUsable
#> [1] FALSE
#> 
#> $`Drosophila yakuba`$scientificName
#> [1] "Drosophila yakuba"
#> 
#> $`Drosophila yakuba`$parentTaxon
#> $`Drosophila yakuba`$parentTaxon$commonName
#> [1] "drosophila"
#> 
#> $`Drosophila yakuba`$parentTaxon$externalDatabase
#> NULL
#> 
#> $`Drosophila yakuba`$parentTaxon$isSpecies
#> [1] FALSE
#> 
#> $`Drosophila yakuba`$parentTaxon$isGenesUsable
#> [1] TRUE
#> 
#> $`Drosophila yakuba`$parentTaxon$scientificName
#> [1] "Drosophila"
#> 
#> $`Drosophila yakuba`$parentTaxon$parentTaxon
#> NULL
#> 
#> $`Drosophila yakuba`$parentTaxon$ncbiId
#> NULL
#> 
#> $`Drosophila yakuba`$parentTaxon$abbreviation
#> NULL
#> 
#> $`Drosophila yakuba`$parentTaxon$id
#> [1] 26
#> 
#> 
#> $`Drosophila yakuba`$ncbiId
#> NULL
#> 
#> $`Drosophila yakuba`$abbreviation
#> NULL
#> 
#> $`Drosophila yakuba`$id
#> [1] 22
#> 
#> 
#> $`Drosophila teissieri`
#> $`Drosophila teissieri`$commonName
#> [1] "d.teissieri"
#> 
#> $`Drosophila teissieri`$externalDatabase
#> NULL
#> 
#> $`Drosophila teissieri`$isSpecies
#> [1] TRUE
#> 
#> $`Drosophila teissieri`$isGenesUsable
#> [1] FALSE
#> 
#> $`Drosophila teissieri`$scientificName
#> [1] "Drosophila teissieri"
#> 
#> $`Drosophila teissieri`$parentTaxon
#> $`Drosophila teissieri`$parentTaxon$commonName
#> [1] "drosophila"
#> 
#> $`Drosophila teissieri`$parentTaxon$externalDatabase
#> NULL
#> 
#> $`Drosophila teissieri`$parentTaxon$isSpecies
#> [1] FALSE
#> 
#> $`Drosophila teissieri`$parentTaxon$isGenesUsable
#> [1] TRUE
#> 
#> $`Drosophila teissieri`$parentTaxon$scientificName
#> [1] "Drosophila"
#> 
#> $`Drosophila teissieri`$parentTaxon$parentTaxon
#> NULL
#> 
#> $`Drosophila teissieri`$parentTaxon$ncbiId
#> NULL
#> 
#> $`Drosophila teissieri`$parentTaxon$abbreviation
#> NULL
#> 
#> $`Drosophila teissieri`$parentTaxon$id
#> [1] 26
#> 
#> 
#> $`Drosophila teissieri`$ncbiId
#> NULL
#> 
#> $`Drosophila teissieri`$abbreviation
#> NULL
#> 
#> $`Drosophila teissieri`$id
#> [1] 23
#> 
#> 
#> $`Drosophila mauritiana`
#> $`Drosophila mauritiana`$commonName
#> [1] "d.mauritiana"
#> 
#> $`Drosophila mauritiana`$externalDatabase
#> NULL
#> 
#> $`Drosophila mauritiana`$isSpecies
#> [1] TRUE
#> 
#> $`Drosophila mauritiana`$isGenesUsable
#> [1] FALSE
#> 
#> $`Drosophila mauritiana`$scientificName
#> [1] "Drosophila mauritiana"
#> 
#> $`Drosophila mauritiana`$parentTaxon
#> $`Drosophila mauritiana`$parentTaxon$commonName
#> [1] "drosophila"
#> 
#> $`Drosophila mauritiana`$parentTaxon$externalDatabase
#> NULL
#> 
#> $`Drosophila mauritiana`$parentTaxon$isSpecies
#> [1] FALSE
#> 
#> $`Drosophila mauritiana`$parentTaxon$isGenesUsable
#> [1] TRUE
#> 
#> $`Drosophila mauritiana`$parentTaxon$scientificName
#> [1] "Drosophila"
#> 
#> $`Drosophila mauritiana`$parentTaxon$parentTaxon
#> NULL
#> 
#> $`Drosophila mauritiana`$parentTaxon$ncbiId
#> NULL
#> 
#> $`Drosophila mauritiana`$parentTaxon$abbreviation
#> NULL
#> 
#> $`Drosophila mauritiana`$parentTaxon$id
#> [1] 26
#> 
#> 
#> $`Drosophila mauritiana`$ncbiId
#> NULL
#> 
#> $`Drosophila mauritiana`$abbreviation
#> NULL
#> 
#> $`Drosophila mauritiana`$id
#> [1] 24
#> 
#> 
#> $`Drosophila sechellia`
#> $`Drosophila sechellia`$commonName
#> [1] "d.sechellia"
#> 
#> $`Drosophila sechellia`$externalDatabase
#> NULL
#> 
#> $`Drosophila sechellia`$isSpecies
#> [1] TRUE
#> 
#> $`Drosophila sechellia`$isGenesUsable
#> [1] FALSE
#> 
#> $`Drosophila sechellia`$scientificName
#> [1] "Drosophila sechellia"
#> 
#> $`Drosophila sechellia`$parentTaxon
#> $`Drosophila sechellia`$parentTaxon$commonName
#> [1] "drosophila"
#> 
#> $`Drosophila sechellia`$parentTaxon$externalDatabase
#> NULL
#> 
#> $`Drosophila sechellia`$parentTaxon$isSpecies
#> [1] FALSE
#> 
#> $`Drosophila sechellia`$parentTaxon$isGenesUsable
#> [1] TRUE
#> 
#> $`Drosophila sechellia`$parentTaxon$scientificName
#> [1] "Drosophila"
#> 
#> $`Drosophila sechellia`$parentTaxon$parentTaxon
#> NULL
#> 
#> $`Drosophila sechellia`$parentTaxon$ncbiId
#> NULL
#> 
#> $`Drosophila sechellia`$parentTaxon$abbreviation
#> NULL
#> 
#> $`Drosophila sechellia`$parentTaxon$id
#> [1] 26
#> 
#> 
#> $`Drosophila sechellia`$ncbiId
#> NULL
#> 
#> $`Drosophila sechellia`$abbreviation
#> NULL
#> 
#> $`Drosophila sechellia`$id
#> [1] 25
#> 
#> 
#> $Drosophila
#> $Drosophila$commonName
#> [1] "drosophila"
#> 
#> $Drosophila$externalDatabase
#> NULL
#> 
#> $Drosophila$isSpecies
#> [1] FALSE
#> 
#> $Drosophila$isGenesUsable
#> [1] TRUE
#> 
#> $Drosophila$scientificName
#> [1] "Drosophila"
#> 
#> $Drosophila$parentTaxon
#> NULL
#> 
#> $Drosophila$ncbiId
#> NULL
#> 
#> $Drosophila$abbreviation
#> NULL
#> 
#> $Drosophila$id
#> [1] 26
#> 
#> 
#> $`Salvelinus fontinalis`
#> $`Salvelinus fontinalis`$commonName
#> [1] "brook trout"
#> 
#> $`Salvelinus fontinalis`$externalDatabase
#> NULL
#> 
#> $`Salvelinus fontinalis`$isSpecies
#> [1] TRUE
#> 
#> $`Salvelinus fontinalis`$isGenesUsable
#> [1] FALSE
#> 
#> $`Salvelinus fontinalis`$scientificName
#> [1] "Salvelinus fontinalis"
#> 
#> $`Salvelinus fontinalis`$parentTaxon
#> $`Salvelinus fontinalis`$parentTaxon$commonName
#> [1] "salmonid"
#> 
#> $`Salvelinus fontinalis`$parentTaxon$externalDatabase
#> NULL
#> 
#> $`Salvelinus fontinalis`$parentTaxon$isSpecies
#> [1] FALSE
#> 
#> $`Salvelinus fontinalis`$parentTaxon$isGenesUsable
#> [1] TRUE
#> 
#> $`Salvelinus fontinalis`$parentTaxon$scientificName
#> [1] "Salmonidae"
#> 
#> $`Salvelinus fontinalis`$parentTaxon$parentTaxon
#> NULL
#> 
#> $`Salvelinus fontinalis`$parentTaxon$ncbiId
#> [1] 8015
#> 
#> $`Salvelinus fontinalis`$parentTaxon$abbreviation
#> NULL
#> 
#> $`Salvelinus fontinalis`$parentTaxon$id
#> [1] 4
#> 
#> 
#> $`Salvelinus fontinalis`$ncbiId
#> [1] 8038
#> 
#> $`Salvelinus fontinalis`$abbreviation
#> NULL
#> 
#> $`Salvelinus fontinalis`$id
#> [1] 27
#> 
#> 
#> $`Fundulus heteroclitus`
#> $`Fundulus heteroclitus`$commonName
#> [1] "killifish"
#> 
#> $`Fundulus heteroclitus`$externalDatabase
#> NULL
#> 
#> $`Fundulus heteroclitus`$isSpecies
#> [1] TRUE
#> 
#> $`Fundulus heteroclitus`$isGenesUsable
#> [1] FALSE
#> 
#> $`Fundulus heteroclitus`$scientificName
#> [1] "Fundulus heteroclitus"
#> 
#> $`Fundulus heteroclitus`$parentTaxon
#> $`Fundulus heteroclitus`$parentTaxon$commonName
#> [1] "salmonid"
#> 
#> $`Fundulus heteroclitus`$parentTaxon$externalDatabase
#> NULL
#> 
#> $`Fundulus heteroclitus`$parentTaxon$isSpecies
#> [1] FALSE
#> 
#> $`Fundulus heteroclitus`$parentTaxon$isGenesUsable
#> [1] TRUE
#> 
#> $`Fundulus heteroclitus`$parentTaxon$scientificName
#> [1] "Salmonidae"
#> 
#> $`Fundulus heteroclitus`$parentTaxon$parentTaxon
#> NULL
#> 
#> $`Fundulus heteroclitus`$parentTaxon$ncbiId
#> [1] 8015
#> 
#> $`Fundulus heteroclitus`$parentTaxon$abbreviation
#> NULL
#> 
#> $`Fundulus heteroclitus`$parentTaxon$id
#> [1] 4
#> 
#> 
#> $`Fundulus heteroclitus`$ncbiId
#> [1] 8078
#> 
#> $`Fundulus heteroclitus`$abbreviation
#> NULL
#> 
#> $`Fundulus heteroclitus`$id
#> [1] 28
#> 
#> 
#> $`Sparus aurata`
#> $`Sparus aurata`$commonName
#> NULL
#> 
#> $`Sparus aurata`$externalDatabase
#> NULL
#> 
#> $`Sparus aurata`$isSpecies
#> [1] TRUE
#> 
#> $`Sparus aurata`$isGenesUsable
#> [1] FALSE
#> 
#> $`Sparus aurata`$scientificName
#> [1] "Sparus aurata"
#> 
#> $`Sparus aurata`$parentTaxon
#> NULL
#> 
#> $`Sparus aurata`$ncbiId
#> NULL
#> 
#> $`Sparus aurata`$abbreviation
#> NULL
#> 
#> $`Sparus aurata`$id
#> [1] 30
#> 
#> 
#> $`Macaca mulatta`
#> $`Macaca mulatta`$commonName
#> NULL
#> 
#> $`Macaca mulatta`$externalDatabase
#> NULL
#> 
#> $`Macaca mulatta`$isSpecies
#> [1] TRUE
#> 
#> $`Macaca mulatta`$isGenesUsable
#> [1] FALSE
#> 
#> $`Macaca mulatta`$scientificName
#> [1] "Macaca mulatta"
#> 
#> $`Macaca mulatta`$parentTaxon
#> NULL
#> 
#> $`Macaca mulatta`$ncbiId
#> NULL
#> 
#> $`Macaca mulatta`$abbreviation
#> NULL
#> 
#> $`Macaca mulatta`$id
#> [1] 31
#> 
#> 
#> $`Pan troglodytes`
#> $`Pan troglodytes`$commonName
#> NULL
#> 
#> $`Pan troglodytes`$externalDatabase
#> NULL
#> 
#> $`Pan troglodytes`$isSpecies
#> [1] TRUE
#> 
#> $`Pan troglodytes`$isGenesUsable
#> [1] FALSE
#> 
#> $`Pan troglodytes`$scientificName
#> [1] "Pan troglodytes"
#> 
#> $`Pan troglodytes`$parentTaxon
#> NULL
#> 
#> $`Pan troglodytes`$ncbiId
#> NULL
#> 
#> $`Pan troglodytes`$abbreviation
#> NULL
#> 
#> $`Pan troglodytes`$id
#> [1] 32
#> 
#> 
#> $`Schizosaccharomyces pombe`
#> $`Schizosaccharomyces pombe`$commonName
#> NULL
#> 
#> $`Schizosaccharomyces pombe`$externalDatabase
#> NULL
#> 
#> $`Schizosaccharomyces pombe`$isSpecies
#> [1] TRUE
#> 
#> $`Schizosaccharomyces pombe`$isGenesUsable
#> [1] FALSE
#> 
#> $`Schizosaccharomyces pombe`$scientificName
#> [1] "Schizosaccharomyces pombe"
#> 
#> $`Schizosaccharomyces pombe`$parentTaxon
#> NULL
#> 
#> $`Schizosaccharomyces pombe`$ncbiId
#> NULL
#> 
#> $`Schizosaccharomyces pombe`$abbreviation
#> NULL
#> 
#> $`Schizosaccharomyces pombe`$id
#> [1] 33
#> 
#> 
#> $`Canis lupus familiaris`
#> $`Canis lupus familiaris`$commonName
#> NULL
#> 
#> $`Canis lupus familiaris`$externalDatabase
#> NULL
#> 
#> $`Canis lupus familiaris`$isSpecies
#> [1] TRUE
#> 
#> $`Canis lupus familiaris`$isGenesUsable
#> [1] FALSE
#> 
#> $`Canis lupus familiaris`$scientificName
#> [1] "Canis lupus familiaris"
#> 
#> $`Canis lupus familiaris`$parentTaxon
#> NULL
#> 
#> $`Canis lupus familiaris`$ncbiId
#> NULL
#> 
#> $`Canis lupus familiaris`$abbreviation
#> NULL
#> 
#> $`Canis lupus familiaris`$id
#> [1] 70
#> 
#> 
#> $`Monodelphis domestica`
#> $`Monodelphis domestica`$commonName
#> NULL
#> 
#> $`Monodelphis domestica`$externalDatabase
#> NULL
#> 
#> $`Monodelphis domestica`$isSpecies
#> [1] TRUE
#> 
#> $`Monodelphis domestica`$isGenesUsable
#> [1] FALSE
#> 
#> $`Monodelphis domestica`$scientificName
#> [1] "Monodelphis domestica"
#> 
#> $`Monodelphis domestica`$parentTaxon
#> NULL
#> 
#> $`Monodelphis domestica`$ncbiId
#> NULL
#> 
#> $`Monodelphis domestica`$abbreviation
#> NULL
#> 
#> $`Monodelphis domestica`$id
#> [1] 71
#> 
#> 
#> $`Ornithorhynchus anatinus`
#> $`Ornithorhynchus anatinus`$commonName
#> NULL
#> 
#> $`Ornithorhynchus anatinus`$externalDatabase
#> NULL
#> 
#> $`Ornithorhynchus anatinus`$isSpecies
#> [1] TRUE
#> 
#> $`Ornithorhynchus anatinus`$isGenesUsable
#> [1] FALSE
#> 
#> $`Ornithorhynchus anatinus`$scientificName
#> [1] "Ornithorhynchus anatinus"
#> 
#> $`Ornithorhynchus anatinus`$parentTaxon
#> NULL
#> 
#> $`Ornithorhynchus anatinus`$ncbiId
#> NULL
#> 
#> $`Ornithorhynchus anatinus`$abbreviation
#> NULL
#> 
#> $`Ornithorhynchus anatinus`$id
#> [1] 72
#> 
#> 
#> $`Gallus gallus`
#> $`Gallus gallus`$commonName
#> NULL
#> 
#> $`Gallus gallus`$externalDatabase
#> NULL
#> 
#> $`Gallus gallus`$isSpecies
#> [1] TRUE
#> 
#> $`Gallus gallus`$isGenesUsable
#> [1] FALSE
#> 
#> $`Gallus gallus`$scientificName
#> [1] "Gallus gallus"
#> 
#> $`Gallus gallus`$parentTaxon
#> NULL
#> 
#> $`Gallus gallus`$ncbiId
#> NULL
#> 
#> $`Gallus gallus`$abbreviation
#> NULL
#> 
#> $`Gallus gallus`$id
#> [1] 73
#> 
#> 
#> $`Gorilla gorilla`
#> $`Gorilla gorilla`$commonName
#> NULL
#> 
#> $`Gorilla gorilla`$externalDatabase
#> NULL
#> 
#> $`Gorilla gorilla`$isSpecies
#> [1] TRUE
#> 
#> $`Gorilla gorilla`$isGenesUsable
#> [1] FALSE
#> 
#> $`Gorilla gorilla`$scientificName
#> [1] "Gorilla gorilla"
#> 
#> $`Gorilla gorilla`$parentTaxon
#> NULL
#> 
#> $`Gorilla gorilla`$ncbiId
#> NULL
#> 
#> $`Gorilla gorilla`$abbreviation
#> NULL
#> 
#> $`Gorilla gorilla`$id
#> [1] 74
#> 
#> 
#> $`Xenopus tropicalis`
#> $`Xenopus tropicalis`$commonName
#> NULL
#> 
#> $`Xenopus tropicalis`$externalDatabase
#> NULL
#> 
#> $`Xenopus tropicalis`$isSpecies
#> [1] TRUE
#> 
#> $`Xenopus tropicalis`$isGenesUsable
#> [1] FALSE
#> 
#> $`Xenopus tropicalis`$scientificName
#> [1] "Xenopus tropicalis"
#> 
#> $`Xenopus tropicalis`$parentTaxon
#> NULL
#> 
#> $`Xenopus tropicalis`$ncbiId
#> NULL
#> 
#> $`Xenopus tropicalis`$abbreviation
#> NULL
#> 
#> $`Xenopus tropicalis`$id
#> [1] 75
#> 
#> 
#> $`Pan paniscus`
#> $`Pan paniscus`$commonName
#> NULL
#> 
#> $`Pan paniscus`$externalDatabase
#> NULL
#> 
#> $`Pan paniscus`$isSpecies
#> [1] TRUE
#> 
#> $`Pan paniscus`$isGenesUsable
#> [1] FALSE
#> 
#> $`Pan paniscus`$scientificName
#> [1] "Pan paniscus"
#> 
#> $`Pan paniscus`$parentTaxon
#> NULL
#> 
#> $`Pan paniscus`$ncbiId
#> NULL
#> 
#> $`Pan paniscus`$abbreviation
#> NULL
#> 
#> $`Pan paniscus`$id
#> [1] 76
#> 
#> 
#> $`Bos taurus`
#> $`Bos taurus`$commonName
#> NULL
#> 
#> $`Bos taurus`$externalDatabase
#> NULL
#> 
#> $`Bos taurus`$isSpecies
#> [1] TRUE
#> 
#> $`Bos taurus`$isGenesUsable
#> [1] FALSE
#> 
#> $`Bos taurus`$scientificName
#> [1] "Bos taurus"
#> 
#> $`Bos taurus`$parentTaxon
#> NULL
#> 
#> $`Bos taurus`$ncbiId
#> NULL
#> 
#> $`Bos taurus`$abbreviation
#> NULL
#> 
#> $`Bos taurus`$id
#> [1] 77
#> 
#> 

allTaxa(c('human','mouse'))

#> $`Homo sapiens`
#> $`Homo sapiens`$commonName
#> [1] "human"
#> 
#> $`Homo sapiens`$externalDatabase
#> $`Homo sapiens`$externalDatabase$checked
#> [1] FALSE
#> 
#> $`Homo sapiens`$externalDatabase$name
#> [1] "hg19"
#> 
#> $`Homo sapiens`$externalDatabase$id
#> [1] 65
#> 
#> 
#> $`Homo sapiens`$isSpecies
#> [1] TRUE
#> 
#> $`Homo sapiens`$isGenesUsable
#> [1] TRUE
#> 
#> $`Homo sapiens`$scientificName
#> [1] "Homo sapiens"
#> 
#> $`Homo sapiens`$parentTaxon
#> NULL
#> 
#> $`Homo sapiens`$ncbiId
#> [1] 9606
#> 
#> $`Homo sapiens`$abbreviation
#> NULL
#> 
#> $`Homo sapiens`$id
#> [1] 1
#> 
#> 
#> $`Mus musculus`
#> $`Mus musculus`$commonName
#> [1] "mouse"
#> 
#> $`Mus musculus`$externalDatabase
#> $`Mus musculus`$externalDatabase$checked
#> [1] FALSE
#> 
#> $`Mus musculus`$externalDatabase$name
#> [1] "mm10"
#> 
#> $`Mus musculus`$externalDatabase$id
#> [1] 81
#> 
#> 
#> $`Mus musculus`$isSpecies
#> [1] TRUE
#> 
#> $`Mus musculus`$isGenesUsable
#> [1] TRUE
#> 
#> $`Mus musculus`$scientificName
#> [1] "Mus musculus"
#> 
#> $`Mus musculus`$parentTaxon
#> NULL
#> 
#> $`Mus musculus`$ncbiId
#> [1] 10090
#> 
#> $`Mus musculus`$abbreviation
#> NULL
#> 
#> $`Mus musculus`$id
#> [1] 2
#> 
#>