Datasets differential expression levels Retrieves differential expression levels for given datasets

Datasets differential expression levels Retrieves differential expression levels for given datasets

getDatasetDE(
  datasets = NA_character_,
  keepNonSpecific = FALSE,
  diffExSet = NA_integer_,
  threshold = 100,
  limit = 100L,
  consolidate = NA_character_,
  raw = getOption("gemma.raw", F),
  async = getOption("gemma.async", F),
  memoised = getOption("gemma.memoise", F),
  file = getOption("gemma.file", NA_character_),
  overwrite = getOption("gemma.overwrite", F)
)

Arguments

datasets
Optional, defaults to empty. Limits the result to entities with given identifiers. A list of identifiers, separated by commas (e.g: GSE2871,GSE2869,GSE2868). Identifiers can either be the Dataset ID or its short name. Retrieval by ID is more efficient. Only datasets that user has access to will be available. Do not combine different identifiers in one query.

keepNonSpecific
Optional, defaults to false. If set to false, the response will only include elements that map exclusively to each queried gene If set to true, the response will include all elements that map to each queried gene, even if they also map to other genes.

diffExSet
Required, defaults to empty. The ID of the differential expression set to retrieve the data from. This value can be obtained through the \'Dataset differential analysis\' endpoint in the \'Dataset endpoints\' category. See the resultSetId in one of the response objects in said endpoint.

threshold
Optional, defaults to 100.0. The threshold that the differential expression has to meet to be included in the response.

limit
Optional, defaults to 20. Limits the result to specified amount of objects. Use 0 for no limit.

consolidate
Optional, defaults to empty. What action to take when there is more than one element per gene in a dataset. The choices are:

[empty] - will list all vectors separately

pickmax - only return the vector that has the highest expression (mean over all its bioAssays)

pickvar - only return the vector with highest variance of expression across its bioAssays

average - create a new vector that will average the bioAssay values from all vectors

raw
FALSE to receive results as-is from Gemma, or TRUE to enable parsing.

async
TRUE to run the API query on a separate worker, or FALSE to run synchronously. See the async package for details.

memoised
Whether or not to cache results so future requests for the same data will be faster. Use forgetGemmaMemoised to clear the cache.

file
The name of a file to save the results to, or NULL to not write results to a file. If raw == TRUE, the output will be a JSON file. Otherwise, it will be a RDS file.

overwrite
Whether or not to overwrite if a file exists at the specified filename.

Value

The differential expression levels for each given experiment (experiment expression levels value object) in the given differential expression set. If the experiment is not in the given diff. exp. set, an empty array is returned. A 404 error if the given identifier does not map to any object.